% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{add1KG2SampleGDS}
\alias{add1KG2SampleGDS}
\title{TODO}
\usage{
add1KG2SampleGDS(gds, gdsSampleFile)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{gdsSampleFile}{the path of an object of class \code{gds} related to
the sample}
}
\value{
TODO a \code{vector} of \code{string}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
