% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{addBlockFromPlink2GDS}
\alias{addBlockFromPlink2GDS}
\title{Save the information about a specific block in a
Population Reference GDS Annotation file}
\usage{
addBlockFromPlink2GDS(
  gds,
  gdsOut,
  PATHBLOCK,
  superPop,
  blockName,
  blockDesc,
  verbose = FALSE
)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), TODO}

\item{gdsOut}{an object of class \code{gds} opened in writing mode.}

\item{PATHBLOCK}{TODO}

\item{superPop}{TODO}

\item{blockName}{a \code{character} string representing the unique
block name.}

\item{blockDesc}{a \code{character} string representing the description of
the current block.}

\item{verbose}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}.}
}
\value{
\code{OL} when the function is successful.
}
\description{
This function extracts the information for all
SNVs related to a specific block from the Population Reference GDS file. It
uses this information and the information from the block files associated
to a specific super-population to generate the final block information.
The block information is then saved in a Population ReferenceG DS Annotation
file.
}
\details{
More information about GDS file format can be found at the Bioconductor
gdsfmt website:
https://bioconductor.org/packages/gdsfmt/
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
