% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDSRef}
\alias{addGDSRef}
\title{The function add an array sample.ref to the gds file.It define base
on a list of unrelated samples.}
\usage{
addGDSRef(gds, filePart)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened GDS file.}

\item{filePart}{a \code{list} from the function pcairPartition in GENESIS}
}
\value{
The integer \code{0} when successful.
}
\description{
This function create the field sample.ref which is 1 when
the samples are a reference and 0 otherwise. The sample.ref is fill based
on the file filePart$unrels
from  in GENESIS TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
