% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDSStudyPruning}
\alias{addGDSStudyPruning}
\title{Fill the pruned.study entry in the GDS file}
\usage{
addGDSStudyPruning(gds, pruned, sample.id)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the sample}

\item{pruned}{TODO}

\item{sample.id}{a \code{character} string corresponding to
the sample identifier.}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
