% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper2_internal.R
\encoding{UTF-8}
\name{addGenoSynVcf}
\alias{addGenoSynVcf}
\title{Add the genotype from synthetic vcf information for the list
of pruned SNVs
into the Profile GDS file}
\usage{
addGenoSynVcf(synProfileRef, synVCFfile, pRAIDS = pRAIDS)
}
\arguments{
\item{synProfileRef}{a TODO.}

\item{synVCFfile}{a TODO.}

\item{pRAIDS}{a TODO.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
The function extracts the information about the pruned SNVs
from the 1KG GDS file and adds entries related to the pruned SNVs in
the Profile GDS file. The nodes are added to the Profile GDS file:
'sample.id', 'snp.id', 'snp.chromosome', 'snp.position', 'snp.index',
'genotype' and 'lap'.
}
\examples{

## TODO


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
