% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{addRef2GDS1KG}
\alias{addRef2GDS1KG}
\title{Add the information about the unrelated patients to the GDS 1KG file}
\usage{
addRef2GDS1KG(fileNameGDS, filePart)
}
\arguments{
\item{fileNameGDS}{a \code{character} string representing the path and file
name of the GDS file that contains the 1KG information. The GDS file must
contain the SNP information, the genotyping information and
the pedigree information from 1000 Genomes.
The extension of the file must be '.gds'.}

\item{filePart}{a \code{character} string representing the path and file
name of the RDS file that contains the
information about the 1KG patients that are unrelated.
The extension of the file must be '.rds'. The file must exists.}
}
\value{
\code{None}
}
\description{
This function adds the information about the unrelated patients
to the GDS 1KG file. More specificaly, it creates the field
\code{sample.ref} which as the value \code{1} when the sample
is unrelated and the value \code{0} otherwise.
The \code{sample.ref} is filled based on the information present in the
input RDS file.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
