% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper2_internal.R
\encoding{UTF-8}
\name{addStudyGDSProfileSyn}
\alias{addStudyGDSProfileSyn}
\title{Add information related to a specific study and specific samples
into a GDS Sample file}
\usage{
addStudyGDSProfileSyn(profileSyn, pRAIDS = pRAIDS)
}
\arguments{
\item{profileSyn}{an object of TODO.}

\item{pRAIDS}{a TODO.}
}
\value{
a \code{vector} of \code{character} strings representing the sample
identifiers that have been saved in the GDS Sample file.
}
\description{
This function add entries related to 1) a specific study and
2) specific samples into a GDS Sample file.
The study information is appended to the GDS Sample file "study.list" node
when the node is already present in the file. Otherwise, the node is
created and then, the information is added.
The sample information for all selected samples is appended to the GDS
Sample file "study.annot" node
when the node is already present in the file. Otherwise, the node is
created and then, the information is added.
}
\examples{

## TODO


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
