% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addStudyGDSSample}
\alias{addStudyGDSSample}
\title{Add information related to a specific study and a specific sample
into a GDS file}
\usage{
addStudyGDSSample(
  gds,
  pedDF,
  batch = 1,
  listSamples = NULL,
  studyDF,
  verbose = TRUE
)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened GDS file.}

\item{pedDF}{a \code{data.frame} with the sample information. The
\code{data.frame} must have the columns:
"Name.ID", "Case.ID", "Sample.Type", "Diagnosis" and "Source".
The unique sample identifier of the \code{data.frame} is the "Name.ID"
column and the row names of the \code{data.frame} must be the "Name.ID"
values.}

\item{batch}{a \code{integer} corresponding the batch associated to the
study. Default: \code{1}.}

\item{listSamples}{a \code{vector} of \code{character} string representing
the samples (samples identifiers) that are saved into the GDS.
If \code{NULL}, all samples present in the \code{pedDF} are used.
Default: \code{NULL}.}

\item{studyDF}{a \code{data.frame} with at least the 3 columns: "study.id",
"study.desc" and "study.platform".}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function. Default: \code{TRUE}.}
}
\value{
a \code{vector} of \code{character} strings representing the sample
identifiers present in the \code{pedDF} object.
}
\description{
This function add entries related to 1) a specific study and
2) a specific sample into a GDS file. More specifically, a
"study.annot" node is added to the GDS file. In addition, those nodes are
also created inside those 2 nodes: "study.list" and "study.annot".
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
