% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{addUpdateSegment}
\alias{addUpdateSegment}
\title{TODO}
\usage{
addUpdateSegment(gds, snp.seg)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the sample}

\item{snp.seg}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
