% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicsPaper.R
\docType{package}
\encoding{UTF-8}
\name{aicsPaper-package}
\alias{aicsPaper-package}
\alias{aicsPaper}
\title{aicsPaper: Ancestry Inference from Cancer Sequencing}
\description{
The aicsPaper package implements specialized algorithms that enable
ancestry inference from various cancer sequencing sources (RNA,
Exome and Whole-Genome sequencing).
}
\details{
The aicsPaper package also implements simulation algorithm that generates
synthetic cancer-derived data.

This code and analysis pipeline was designed and developed for the
following publication:
}
\references{
Belleau, P, Deschênes A, Tuveson DA, Krasnitz A. Accurate and robust
inference of genetic ancestry from cancer-derived molecular data across
genomic platforms. bioRxiv 2022.02.01.478737
[link to bioRxiv](https://doi.org/10.1101/2022.02.01.478737)
}
\seealso{
\itemize{
    \item \code{\link{TODO}} {for TODO}
}
}
\author{
Pascal Belleau, Astrid Deschênes and
Alexander Krasnitz

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
