% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{appendGDSgenotypeMat}
\alias{appendGDSgenotypeMat}
\title{This function append the field genotype in the gds file}
\usage{
appendGDSgenotypeMat(gds, matG)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{matG}{a \code{matrix} with the genotype}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
