% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{appendGDSSample}
\alias{appendGDSSample}
\title{This function append the fields related to the samples. If the
samples are part of a study you must uses the addStudyGDSSample}
\usage{
appendGDSSample(gds, pedDF, batch = 1, listSamples = NULL, verbose = TRUE)
}
\arguments{
\item{gds}{a \code{gds}.}

\item{pedDF}{a \code{data.frame} with the sample info. Must have the column
sample.id, Name.ID, sex, pop.group, superPop and batch. The unique id
of pedDF is Name.ID and the row.name is Name.ID too.}

\item{listSamples}{a \code{array} with the sample from pedDF$Name.ID to keep}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function. Default: \code{TRUE}.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function append the fields related to the samples.
The fields append are sample.id and the \code{data.frame} sample.annot.
If the samples are in the section study the field related to the
study must be fill.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
