% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{appendGDSSampleOnly}
\alias{appendGDSSampleOnly}
\title{This function append the fields sample.id.}
\usage{
appendGDSSampleOnly(gds, listSamples)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened GDS file.}

\item{listSample}{a \code{array} of sample.id to add.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function append the fields samples.id.
The fields append are sample.id with the listSample
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
