% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{basePCASample}
\alias{basePCASample}
\title{TODO}
\usage{
basePCASample(gds, listSample.Ref = NULL, listSNP = NULL, np = 1)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{listSample.Ref}{A \code{vector} of \code{string} corresponding to
the sample.ids}

\item{listSNP}{the list of snp.id keep}

\item{np}{a \code{integer} reprenting the number of thread}
}
\value{
TODO a \code{list}  with with two objects pca.unrel -> \code{snpgdsPCAClass}
and a snp.load -> \code{snpgdsPCASNPLoading}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
