% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{batch,RAIDSparam-method}
\alias{batch,RAIDSparam-method}
\title{A getter for the batch slot in a RAIDSparam class}
\usage{
\S4method{batch}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{integer} representing the current
identifier for the batch.
}
\description{
A function for getting the batch slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the batch slot for the object
batch(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
