% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{batch<-,RAIDSparam-method}
\alias{batch<-,RAIDSparam-method}
\title{Setter function for replacement of batch slot in
a RAIDSparam class}
\usage{
\S4method{batch}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single positive \code{integer} representing the current
identifier for the batch.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of batch slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the batch slot in the object
batch(paramDemo) <- 34L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
