% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{batch}
\alias{batch}
\title{Generic function for getting the batch slot in a class}
\usage{
batch(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the batch slot in the S4 object.
}
\description{
A generic function for getting the batch slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(batch="character"))

# Create an instance
obj <- new("MyClass", batch="123")

# Call the generic (assuming a method is implemented)
# batch(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
