% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{blockTypeId,RAIDSparam-method}
\alias{blockTypeId,RAIDSparam-method}
\title{A getter for the blockTypeId slot in a RAIDSparam class}
\usage{
\S4method{blockTypeId}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string corresponding to the block type
used to extract the block identifiers.
}
\description{
A function for getting the blockTypeId slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the blockTypeId slot for the object
blockTypeId(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
