% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{blockTypeId<-,RAIDSparam-method}
\alias{blockTypeId<-,RAIDSparam-method}
\title{Setter function for replacement of blockTypeId slot in a RAIDSparam class}
\usage{
\S4method{blockTypeId}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string corresponding to
the block type used to extract the block identifiers. The block type must
be present in the GDS Reference Annotation file.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of blockTypeId slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new blockTypeId to the blockTypeId slot in the object
blockTypeId(paramDemo) <- "TEST"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
