% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{blockTypeId}
\alias{blockTypeId}
\title{Generic function for getting the blockTypeId slot in a class}
\usage{
blockTypeId(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the blockTypeId slot in the S4 object.
}
\description{
A generic function for getting the blockTypeId slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(blockTypeId="character"))

# Create an instance
obj <- new("MyClass", blockTypeId="123")

# Call the generic (assuming a method is implemented)
# blockTypeId(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
