% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{calcAFMLRNA}
\alias{calcAFMLRNA}
\title{TOREVIEW Compute the log likelihood ratio base on the coverage (read depth)
of each allele in block (gene in the case of RNA-seq)}
\usage{
calcAFMLRNA(snpPosHetero)
}
\arguments{
\item{snpPosHetero}{For a specific gene (block) a \code{data.frame} with
lap for the SNV heterozygote dataset with
coverage > \code{minCov}. The \code{data.frame} must contain those columns:
'phase', 'cnt.ref', 'cnt.alt'. TODO}
}
\value{
TODO a \code{list} of \code{numeric} for the gene lR the score
for aFraction different than 0.5
aFraction allele estimation, nPhase number of SNV phase,
sumAlleleLow number of read overlapping the allele low
sumAlleleHigh number of read overlapping the allele high TODO
}
\description{
TOREVIEW For each hetero sum the log of read depth of the lowest depth
divide by the total depth of the position minus of likelhood of the allelic
fraction of 0.5. If the phase is known, the variant varaint in the same
haplotype are group.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
