% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\docType{class}
\encoding{UTF-8}
\name{CharacterOrNULL-class}
\alias{CharacterOrNULL-class}
\title{Class Union for character or NULL}
\description{
A virtual class that groups character and \code{NULL}
together.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
