% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{chrInfo,RAIDSparam-method}
\alias{chrInfo,RAIDSparam-method}
\title{A getter for the chrInfo slot in a RAIDSparam class}
\usage{
\S4method{chrInfo}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{vector} of positive \code{integer} values
representing the length of the chromosomes.
}
\description{
A function for getting the chrInfo slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the chrInfo slot for the object
chrInfo(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
