% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{chrInfo<-,RAIDSparam-method}
\alias{chrInfo<-,RAIDSparam-method}
\title{Setter function for replacement of chrInfo slot in a RAIDSparam class}
\usage{
\S4method{chrInfo}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{vector} of positive \code{integer} values
representing the length of the chromosomes.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of chrInfo slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new chrInfo to the chrInfo slot in the object
chrInfo(paramDemo) <- c(122222L, 333333L)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
