% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{chrInfo}
\alias{chrInfo}
\title{Generic function for getting the chrInfo slot in a class}
\usage{
chrInfo(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the chrInfo slot in the S4 object.
}
\description{
A generic function for getting the chrInfo slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(chrInfo="character"))

# Create an instance
obj <- new("MyClass", chrInfo="123")

# Call the generic (assuming a method is implemented)
# chrInfo(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
