% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeAlleleFraction}
\alias{computeAlleleFraction}
\title{TODO}
\usage{
computeAlleleFraction(snp.pos, chr, w = 10, cutOff = -3)
}
\arguments{
\item{snp.pos}{TODO}

\item{chr}{A integer for the chromosome TODO}

\item{w}{size of the window to compute the allelic Fraction TODO}

\item{cutOff}{TODO default -3}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
