% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeAllelicFraction}
\alias{computeAllelicFraction}
\title{TODO}
\usage{
computeAllelicFraction(gds, gdsSample, allHetero, normalSample)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{allHetero}{TODO}

\item{normalSample}{TODO}

\item{gdsPhase}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
