% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{computeAllelicFractionRNA}
\alias{computeAllelicFractionRNA}
\title{Estimate the allelic fraction of the pruned SNVs for a specific
RNA-seq sample}
\usage{
computeAllelicFractionRNA(
  gdsReference,
  gdsSample,
  gdsRefAnnot,
  currentProfile,
  studyID,
  blockID,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffAR = 3,
  verbose
)
}
\arguments{
\item{gdsReference}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened Reference GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened Profile GDS file.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened Reference SNV Annotation GDS file.}

\item{currentProfile}{a \code{character} string corresponding to
the sample identifier as used in \code{\link{pruningSample}} function.}

\item{studyID}{a \code{character} string corresponding to the name of
the study as
used in \code{\link{pruningSample}} function.}

\item{blockID}{a \code{character} string corresponding to the field gene
block in the GDS \code{gdsRefAnnot} to use split by gene.}

\item{chrInfo}{a \code{vector} of \code{integer} values representing
the length of the chromosomes.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{minProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
Default: \code{0.999}.}

\item{eProb}{a single \code{numeric} between 0 and 1 representing the
probability of sequencing error. Default: \code{0.001}.}

\item{cutOffLOH}{a single log of the score to be LOH TODO.
Default: \code{-5}.}

\item{cutOffAR}{a single \code{numeric} representing the cutoff, in
log score, to tag SNVs located in a gene has having an allelic fraction
different 0.5
Default: \code{3}.}

\item{verbose}{a \code{logicial} indicating if the function should print
message when running.}
}
\value{
a \code{data.frame} with lap for the pruned SNV dataset with
coverage > \code{minCov}. TODO
}
\description{
The function creates a \code{data.frame} containing the
allelic fraction for the pruned SNV dataset specific to a RNA-seq sample.
}
\examples{

## Required library for GDS
library(SNPRelate)

#' ## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG_GDS.gds")
fileAnnotGDS <- file.path(dataDir, "ex1_good_small_1KG_Annot_GDS.gds")

## Temporary Profile GDS file for one profile
fileProfile <- file.path(getwd(), "ex1.gds")

## Example can only be run if the current directory is in writing mode
if (file.access(getwd()) == 0 && !file.exists(fileProfile)) {

    ## Copy the Profile GDS file demo that has been pruned and annotated
    file.copy(file.path(dataDir, "ex1_demo_with_pruning_and_1KG_annot.gds"),
                 fileProfile)

    ## Open the reference GDS file (demo version)
    gds1KG <- snpgdsOpen(fileGDS)
   gdsRefAnnot <- openfn.gds(fileAnnotGDS)
    ## Open Profile GDS file for one profile
    profileGDS <- openfn.gds(fileProfile)

    ## Chromosome length information
    ## chr23 is chrX, chr24 is chrY and chrM is 25
    chrInfo <- c(248956422L, 242193529L, 198295559L, 190214555L,
        181538259L, 170805979L, 159345973L, 145138636L, 138394717L,
        133797422L, 135086622L, 133275309L, 114364328L, 107043718L,
        101991189L, 90338345L, 83257441L,  80373285L,  58617616L,
        64444167L,  46709983L, 50818468L, 156040895L, 57227415L,  16569L)

    ## The function returns a data frame containing the allelic fraction info
    result <- RAIDS:::computeAllelicFractionRNA(gdsReference=gds1KG,
        gdsSample=profileGDS, gdsRefAnnot=gdsRefAnnot,
        currentProfile="ex1", studyID="MYDATA",
        blockID="GeneS.Ensembl.Hsapiens.v86",
        chrInfo=chrInfo, minCov=10L,
        minProb=0.999, eProb=0.001, cutOffLOH=-5,
        cutOffAR=3, verbose=FALSE)
    head(result)

    ## Close both GDS files (important)
    closefn.gds(profileGDS)
    closefn.gds(gds1KG)
    closefn.gds(gdsRefAnnot)

    ## Remove Profile GDS file (created for demo purpose)
    unlink(fileProfile, force=TRUE)

}

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
