% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeAllelicImbDNAChr}
\alias{computeAllelicImbDNAChr}
\title{TODO}
\usage{
computeAllelicImbDNAChr(snp.pos, chr, wAR = 10, cutOffEmptyBox = -3)
}
\arguments{
\item{snp.pos}{TODO}

\item{chr}{A integer for the chromosome TODO}

\item{wAR}{size-1 of the window to compute an empty box}

\item{cutOffEmptyBox}{TODO default -3}
}
\value{
A vector of the imbAR TODO.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
