% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNRefSynthetic}
\alias{computeKNNRefSynthetic}
\title{Run a k-nearest neighbor analysis on a subset of the
synthetic dataset}
\usage{
computeKNNRefSynthetic(
  gdsSample,
  listEigenvector,
  listCatPop,
  study.id.syn,
  spRef,
  fieldPopInfAnc = "SuperPop",
  kList = seq_len(15),
  pcaList = seq(2, 15, 1)
)
}
\arguments{
\item{gdsSample}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, the
GDS Sample file.}

\item{listEigenvector}{TODO see return of computePCAsynthetic}

\item{listCatPop}{a \code{vector} of \code{character} string
representing the list of possible ancestry assignations. Default:
\code{("EAS", "EUR", "AFR", "AMR", "SAS")}.}

\item{study.id.syn}{a the study.id of the synthetic data}

\item{spRef}{TODO}

\item{fieldPopInfAnc}{TODO . Default: \code{"SuperPop"}.}

\item{kList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _K_ parameter. The _K_ parameter represents the
number of neighbors used in the K-nearest neighbor analysis.
Default: \code{seq_len(15)}.}

\item{pcaList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _D_ parameter. The _D_ parameter represents the
number of dimensions used in the PCA analysis.
Default: \code{seq(2,15,1)}.}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
