% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNSuperPoprSynthetic}
\alias{computeKNNSuperPoprSynthetic}
\title{Deprecated}
\usage{
computeKNNSuperPoprSynthetic(
  listEigenvector,
  sample.ref,
  study.annot,
  spRef,
  kList = seq_len(15),
  pcaList = 2:15
)
}
\arguments{
\item{listEigenvector}{TODO see return of computePCAsynthetic}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}

\item{spRef}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
Deprecated
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
