% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNSuperPopSample}
\alias{computeKNNSuperPopSample}
\title{TODO}
\usage{
computeKNNSuperPopSample(
  gdsSample,
  listEigenvector,
  name.id,
  spRef,
  study.id.ref = "Ref.1KG",
  kList = seq_len(15),
  pcaList = 2:15
)
}
\arguments{
\item{listEigenvector}{TODO see return of computePCARefSample}

\item{spRef}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO Deprecated
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
