% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeLOHBlocksDNAChr}
\alias{computeLOHBlocksDNAChr}
\title{TODO}
\usage{
computeLOHBlocksDNAChr(gds, chrInfo, snp.pos, chr, genoN = 1e-04)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{chrInfo}{a vector chrInfo[i] = length(Hsapiens[[paste0("chr", i)]])
Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)}

\item{snp.pos}{A \code{data.frame} TODO}

\item{chr}{A integer for the chromosome TODO}

\item{genoN}{\code{numeric} between 0 and 1 TODO}
}
\value{
a data.frame TODO.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
