% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCAMultiSynthetic}
\alias{computePCAMultiSynthetic}
\title{TODO}
\usage{
computePCAMultiSynthetic(
  gdsSample,
  listPCA,
  study.annot,
  algorithm = "exact",
  eigen.cnt = 32L,
  missing.rate = 0.025
)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{listPCA}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry for each synthetic
data to process. The format is the same as study.annot in the gdsSample}

\item{algorithm}{algorithm of the PCA "exact", "randomized"
(para snpgdsPCA)}

\item{eigen.cnt}{number of eigenvectors in PCA
(para snpgdsPCA)}

\item{missing.rate}{number of eigenvectors in PCA
(para snpgdsPCA)}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO
gds <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
