% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis_internal.R
\encoding{UTF-8}
\name{computePCAProfile}
\alias{computePCAProfile}
\title{Calculate Principal Component Analysis (PCA) on SNV genotype data set}
\usage{
computePCAProfile(listPCA, profileId, pRAIDS)
}
\arguments{
\item{listPCA}{a \code{list} containing the PCA \code{object} generated
with the 1KG reference profiles.}

\item{profileId}{a \code{character} string representing the profile
identifier.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{list} containing 2 entries:
\describe{
\item{pruned}{ a \code{vector} of SNV identifiers specifying selected SNVs
for the PCA analysis.}
\item{pca.unrel}{ a \code{snpgdsPCAClass} object containing the eigenvalues
as generated by \link[SNPRelate]{snpgdsPCA} function.}
}
}
\description{
The functions calculates the principal component analysis (PCA)
for a list of pruned SNVs present in a Profile GDS file. The
\link[SNPRelate]{snpgdsPCA} function is used to do the calculation.
}
\examples{

## Load the known ancestry for the demo 1KG reference profiles
data(demoKnownSuperPop1KG)

## The Reference GDS file
path1KG <- system.file("extdata/tests", package="RAIDS")

## Path to the demo Profile GDS file is located in this package
dataDir <- system.file("extdata/demoAncestryCall", package="RAIDS")

# The name of the synthetic study
studyID <- "MYDATA"

studyDF <- data.frame(study.id=studyID,
                             study.desc=studyID,
                             study.platform="NotDef",
                             stringsAsFactors=FALSE)
pathProfileGDS <- file.path(dataDir) # , "ex1.gds"
fileReferenceGDS <- system.file("extdata/tests/ex1_good_small_1KG.gds", package="RAIDS")
pedStudy <- data.frame(Name.ID=c("ex1"),
                             Case.ID=c("ex1"),
                             Sample.Type=c("type"),
                             Diagnosis="NotDef",
                             Source=c("NotDef"),
                             stringsAsFactors=FALSE)
     row.names(pedStudy) <- pedStudy$Name.ID

pRAIDS <- paramRAIDS(studyDF=studyDF,
                      pedStudy=pedStudy,
                      pathProfileGDS=pathProfileGDS,
                      fileReferenceGDS=fileReferenceGDS,
                      fieldPopInfAnc="SuperPop")

listPCARef <- RAIDS:::computePCARefRMMulti2(listRM = NULL, pRAIDS = pRAIDS)
listPCAProfile <- RAIDS:::computePCAProfile(listPCA=listPCARef, 
                      profileId=pRAIDS$pedStudy$Name.ID[1], 
                      pRAIDS=pRAIDS)

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
