% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCARefRMMulti}
\alias{computePCARefRMMulti}
\title{Calculate Principal Component Analysis (PCA) on SNV genotype dataset}
\usage{
computePCARefRMMulti(
  gdsSample,
  sample.ref,
  listRM,
  np = 1L,
  algorithm = "exact",
  eigen.cnt = 32L,
  missing.rate = 0.025
)
}
\arguments{
\item{gdsSample}{an object of class \link[SNPRelate]{SNPGDSFileClass},
the GDS Sample file.}

\item{sample.ref}{TODO}

\item{listRM}{a \code{vector} of \code{character} strings containing the
identifiers for the reference samples that need to be removed for the
PCA analysis.}

\item{np}{a single positive \code{integer} representing the number of CPU
that will be used. Default: \code{1L}.}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}

\item{missing.rate}{a \code{numeric} value representing the threshold
missing rate at with the SNVs are discarded; the SNVs are retained in the
\link[SNPRelate]{snpgdsPCA} only
with "<= missing.rate" only; if \code{NaN}, no missing threshold.
Default: \code{0.025}.}
}
\value{
a \code{list} containing 2 entries:
\itemize{
\item{pruned}{ a \code{vector} of SNV identifiers specifying selected SNVs
for the PCA analysis.}
\item{pca.unrel}{ a \code{snpgdsPCAClass} object containing the eigenvalues
as generated by \link[SNPRelate]{snpgdsPCA} function.}
}
}
\description{
The functions calculates the principal component analysis (PCA)
for a list of pruned SNVs present in a GDS Sample file. The
\link[SNPRelate]{snpgdsPCA} function is used to do the calculation.
}
\examples{

# TODO
gds <- "TOTO"

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
