% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCAsynthetic}
\alias{computePCAsynthetic}
\title{TODO}
\usage{
computePCAsynthetic(
  gdsSample,
  pruned,
  sample.id,
  sample.ref,
  study.annot,
  algorithm = "exact",
  eigen.cnt = 32L
)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{pruned}{TODO}

\item{sample.id}{TODO}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}

\item{algorithm}{algorithm of the PCA "exact", "randomized"}

\item{eigen.cnt}{number of eigenvectors in PCA}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO Deprecated
}
\examples{

# TODO
gds <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
