% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{computeSyntheticROC}
\alias{computeSyntheticROC}
\title{TODO}
\usage{
computeSyntheticROC(matKNN, pedCall, refCall, predCall, listCall)
}
\arguments{
\item{matKNN}{TODO}

\item{pedCall}{TODO see return of prepPedSynthetic1KG}

\item{refCall}{TODO column name in pedCall with the call}

\item{predCall}{TODO column name in matKNN with the call}

\item{listCall}{TODO array of the possible call}
}
\value{
\code{list} containing 3 entries:
\itemize{
\item{matAUROC.All}{ TODO }
\item{matAUROC.Call}{ TODO }
\item{listROC.Call}{a \code{list} TODO}
}
}
\description{
TODO
}
\examples{

## TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
