% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{createStudy2GDS1KG}
\alias{createStudy2GDS1KG}
\title{Create the GDS Sample file(s) for one or multiple specific samples
using the information from a RDS Sample description file and the 1KG
GDS file}
\usage{
createStudy2GDS1KG(
  PATHGENO = file.path("data", "sampleGeno"),
  fileNamePED = NULL,
  pedStudy = NULL,
  fileNameGDS,
  batch = 1,
  studyDF,
  listSamples = NULL,
  PATHSAMPLEGDS = NULL,
  verbose = TRUE
)
}
\arguments{
\item{PATHGENO}{a \code{character} string representing the path to the
directory containing the VCF output of SNP-pileup for each sample. The
SNP-pileup files must be compressed (gz files) and have the name identifiers
of the samples. A sample with "Name.ID" identifier would have an
associated SNP-pileup file called "Name.ID.txt.gz".}

\item{fileNamePED}{a \code{character} string representing the path to the
RDS file that contains the information about the sample to analyse.
The RDS file must
include a \code{data.frame} with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in
\code{character} strings. The \code{data.frame}
must contain the information for all the samples passed in the
\code{listSamples} parameter. Only \code{fileNamePED} or \code{pedStudy}
can be defined.}

\item{pedStudy}{a \code{data.frame} with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in
\code{character} strings (no factor). The \code{data.frame}
must contain the information for all the samples passed in the
\code{listSamples} parameter. Only \code{fileNamePED} or \code{pedStudy}
can be defined.}

\item{fileNameGDS}{a \code{character} string representing the file name of
the 1KG GDS file. The file must exist.}

\item{batch}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.
Default: \code{1}.}

\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor).}

\item{listSamples}{a \code{vector} of \code{character} string corresponding
to the sample identifiers that will have a GDS Sample file created. The
sample identifiers must be present in the "Name.ID" column of the RDS file
passed to the \code{fileNamePED} parameter.
If \code{NULL}, all samples in the \code{fileNamePED} are selected.
Default: \code{NULL}.}

\item{PATHSAMPLEGDS}{a \code{character} string representing the path to
the directory where the GDS Sample files will be created.
Default: \code{NULL}.}

\item{verbose}{a \code{logical} indicating if message information should be
printed. Default: \code{TRUE}.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
The function uses the information for the 1KG GDS file and the
RDS Sample Description file to create the GDS Sample file. One GDS Sample
file is created per sample. One GDS Sample file will be created for each
entry present in the \code{listSamples} parameter.
}
\examples{

## Path to the demo 1KG GDS file is located in this package
data.dir <- system.file("extdata/tests", package="RAIDS")
gdsFile <- file.path(data.dir, "ex1_good_small_1KG_GDS.gds")

## The data.frame containing the information about the study
## The 3 mandatory columns: "study.id", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyDF <- data.frame(study.id = "MYDATA",
                        study.desc = "Description",
                        study.platform = "PLATFORM",
                        stringsAsFactors = FALSE)

## The data.frame containing the information about the samples
## The entries should be strings, not factors (stringsAsFactors=FALSE)
samplePED <- data.frame(Name.ID = c("ex1", "ex2"),
                    Case.ID = c("Patient_h11", "Patient_h12"),
                    Diagnosis = rep("Cancer", 2),
                    Sample.Type = rep("Primary Tumor", 2),
                    Source = rep("Databank B", 2), stringsAsFactors = FALSE)
rownames(samplePED) <- samplePED$Name.ID

## Create the Sample GDS file for sample in listSamples vector
## (in this case, samples "ex1")
## The Sample GDS file is created in the PATHSAMPLEGDS directory
result <- createStudy2GDS1KG(PATHGENO=data.dir,
            pedStudy=samplePED, fileNameGDS=gdsFile,
            studyDF=studyDF, listSamples=c("ex1"),
            PATHSAMPLEGDS=data.dir, verbose=FALSE)

## The function returns OL when successful
result

## The Sample GDS file 'ex1.gds' has been created in the
## specified directory
list.files(data.fir)

## Unlink Sample GDS file (created for demo purpose)
unlink(file.path(data.dir, "ex1.gds"))

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
