% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\docType{class}
\encoding{UTF-8}
\name{DataFrameOrNULL-class}
\alias{DataFrameOrNULL-class}
\title{Class Union for data.frame or NULL}
\description{
A virtual class that groups data.frame and \code{NULL}
together.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
