% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{eigenCount<-,RAIDSparam-method}
\alias{eigenCount<-,RAIDSparam-method}
\title{Setter function for replacement of eigenCount slot in
a RAIDSparam class}
\usage{
\S4method{eigenCount}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigenCount' <= 0, then all eigenvectors are returned.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of eigenCount slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the eigenCount slot in the object
eigenCount(paramDemo) <-  33L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
