% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{eigenCount}
\alias{eigenCount}
\title{Generic function for getting the eigenCount slot in a class}
\usage{
eigenCount(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the eigenCount slot in the S4 object.
}
\description{
A generic function for getting the eigenCount slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(eigenCount="character"))

# Create an instance
obj <- new("MyClass", eigenCount="123")

# Call the generic (assuming a method is implemented)
# eigenCount(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
