% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{eigenCountSyn,RAIDSparam-method}
\alias{eigenCountSyn,RAIDSparam-method}
\title{A getter for the eigenCountSyn slot in a RAIDSparam class}
\usage{
\S4method{eigenCountSyn}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{integer} indicating the number of
eigenvectors TODO
}
\description{
A function for getting the eigenCountSyn slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the eigenCountSyn slot for the object
eigenCountSyn(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
