% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{eigenCountSyn<-,RAIDSparam-method}
\alias{eigenCountSyn<-,RAIDSparam-method}
\title{Setter function for replacement of eigenCountSyn slot in
a RAIDSparam class}
\usage{
\S4method{eigenCountSyn}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{integer} indicating the number of
eigenvectors TODO}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of eigenCountSyn slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the eigenCountSyn slot in the object
eigenCountSyn(paramDemo) <-  33L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
