% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{eigenCountSyn}
\alias{eigenCountSyn}
\title{Generic function for getting the eigenCountSyn slot in a class}
\usage{
eigenCountSyn(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the eigenCountSyn slot in the S4 object.
}
\description{
A generic function for getting the eigenCountSyn slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(eigenCountSyn="character"))

# Create an instance
obj <- new("MyClass", eigenCountSyn="123")

# Call the generic (assuming a method is implemented)
# eigenCountSyn(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
