% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{estimateAllelicFraction}
\alias{estimateAllelicFraction}
\title{Estimate the allelic fraction of the pruned SNVs for a specific
sample}
\usage{
estimateAllelicFraction(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  chrInfo,
  studyType = "DNA",
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffHomoScore = -3,
  wAR = 9,
  gdsRefAnnot = NULL
)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the 1KG GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the GDS Sample file.}

\item{sampleCurrent}{a \code{character} string corresponding to
the sample identifier as used in \code{\link{pruningSample}} function.}

\item{study.id}{a \code{character} string corresponding to the study
identifier.}

\item{chrInfo}{a \code{vector}
chrInfo[i] = length(Hsapiens[[paste0("chr", i)]])
       Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)}

\item{studyType}{a \code{character} string representing the type of study.
It can be values such as "DNA", "whole-exome", "RNA", etc..
Default: \code{"DNA"}.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{minProb}{a single \code{numeric} between 0 and 1 representing TODO.
Default: \code{0.999}.}

\item{eProb}{a single \code{numeric} between 0 and 1 representing the
probability of sequencing error. Default: \code{0.001}.}

\item{cutOffLOH}{a single \code{numeric} representing the cutoff, in log,
for the homozygote score to assign a region as LOH.
Default: \code{-5}.}

\item{cutOffHomoScore}{a single \code{numeric} representing the cutoff, in
log, that the SNVs in a block are called homozygote by error.
Default: \code{-3}.}

\item{wAR}{a single positive \code{integer} representing the size-1 of
the window used to compute an empty box. Default: \code{9}.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the1 1KG SNV Annotation GDS file. Default: \code{NULL}.}
}
\value{
The integer \code{0} when successful.
}
\description{
The function estimates the allelic fraction of the pruned
SNVs for a specific sample and add the information to the associated
GDS Sample file.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", "RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
