% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{extractNucleotide}
\alias{extractNucleotide}
\title{Read a VCF file with the genotypes use for the ancestry call}
\usage{
extractNucleotide(nucleotide, count, curNucleo)
}
\arguments{
\item{nucleotide}{a \code{vector} of  a \code{character} string representing A, C, G or T}

\item{count}{a \code{vector} of \code{numeric} with count for the nucleotide}

\item{curNucleoa}{\code{character} string representing A, C, G or T}
}
\value{
a \code{numeric}
}
\description{
The function reads VCF file and
returns a data frame
containing the information about the read counts for the SNVs present in
the file.
}
\examples{


## Directory where demo SNP-pileup file
dataDir <- system.file("extdata/example/snpPileup", package="RAIDS")


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
