% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInfAnc,RAIDSparam-method}
\alias{fieldPopInfAnc,RAIDSparam-method}
\title{A getter for the fieldPopInfAnc slot in a RAIDSparam class}
\usage{
\S4method{fieldPopInfAnc}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{character} string representing the name of
the column in the data frame that contains the inferred
super-population ancestry for the samples. The column should be
present in the data frame.
}
\description{
A function for getting the fieldPopInfAnc slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the fieldPopInfAnc slot for the object
fieldPopInfAnc(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
