% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInfAnc<-,RAIDSparam-method}
\alias{fieldPopInfAnc<-,RAIDSparam-method}
\title{Setter function for replacement of fieldPopInfAnc slot in
a RAIDSparam class}
\usage{
\S4method{fieldPopInfAnc}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{character} string representing the name of
the column in the data frame that contains the inferred
super-population ancestry for the samples. The column should be
present in the data frame.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of fieldPopInfAnc slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the fieldPopInfAnc slot in the object
fieldPopInfAnc(paramDemo) <-  "superPOPULATION"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
