% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInfAnc}
\alias{fieldPopInfAnc}
\title{Generic function for getting the fieldPopInfAnc slot in a class}
\usage{
fieldPopInfAnc(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the fieldPopInfAnc slot in the S4 object.
}
\description{
A generic function for getting the fieldPopInfAnc slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(fieldPopInfAnc="character"))

# Create an instance
obj <- new("MyClass", fieldPopInfAnc="123")

# Call the generic (assuming a method is implemented)
# fieldPopInfAnc(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
