% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInRef,RAIDSparam-method}
\alias{fieldPopInRef,RAIDSparam-method}
\title{A getter for the fieldPopInRef slot in a RAIDSparam class}
\usage{
\S4method{fieldPopInRef}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{character} string representing the name of the
column that contains the known ancestry for the reference profiles in
the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter).
}
\description{
A function for getting the fieldPopInRef slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the fieldPopInRef slot for the object
fieldPopInRef(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
