% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInRef<-,RAIDSparam-method}
\alias{fieldPopInRef<-,RAIDSparam-method}
\title{Setter function for replacement of fieldPopInRef slot in
a RAIDSparam class}
\usage{
\S4method{fieldPopInRef}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{character} string representing the name of the
column that contains the known ancestry for the reference profiles in
the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter).}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of fieldPopInRef slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the fieldPopInRef slot in the object
fieldPopInRef(paramDemo) <-  "newPop"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
