% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldPopInRef}
\alias{fieldPopInRef}
\title{Generic function for getting the fieldPopInRef slot in a class}
\usage{
fieldPopInRef(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the fieldPopInRef slot in the S4 object.
}
\description{
A generic function for getting the fieldPopInRef slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(fieldPopInRef="character"))

# Create an instance
obj <- new("MyClass", fieldPopInRef="123")

# Call the generic (assuming a method is implemented)
# fieldPopInRef(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
