% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldSubPop,RAIDSparam-method}
\alias{fieldSubPop,RAIDSparam-method}
\title{A getter for the fieldSubPop slot in a RAIDSparam class}
\usage{
\S4method{fieldSubPop}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{character} string representing the name of the column
in the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter) that contains the sub-population
information for the samples. The column should be present in the Population
Reference GDS file.
}
\description{
A function for getting the fieldSubPop slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the fieldSubPop slot for the object
fieldSubPop(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
