% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldSubPop<-,RAIDSparam-method}
\alias{fieldSubPop<-,RAIDSparam-method}
\title{Setter function for replacement of fieldSubPop slot in
a RAIDSparam class}
\usage{
\S4method{fieldSubPop}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{character} string representing the name of the
column in the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter) that contains the sub-population
information for the samples. The column should be present in the Population
Reference GDS file.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of fieldSubPop slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the fieldSubPop slot in the object
fieldSubPop(paramDemo) <-  "super.POPULATION.group"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
