% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fieldSubPop}
\alias{fieldSubPop}
\title{Generic function for getting the fieldSubPop slot in a class}
\usage{
fieldSubPop(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the fieldSubPop slot in the S4 object.
}
\description{
A generic function for getting the fieldSubPop slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(fieldSubPop="character"))

# Create an instance
obj <- new("MyClass", fieldSubPop="123")

# Call the generic (assuming a method is implemented)
# fieldSubPop(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
