% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fileReferenceAnnotGDS,RAIDSparam-method}
\alias{fileReferenceAnnotGDS,RAIDSparam-method}
\title{A getter for the fileReferenceAnnotGDS slot in a RAIDSparam class}
\usage{
\S4method{fileReferenceAnnotGDS}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or a \code{character} string representing the file
name of the Population Reference GDS Annotation file.
}
\description{
A function for getting the fileReferenceAnnotGDS slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the fileReferenceAnnotGDS slot for the object
fileReferenceAnnotGDS(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
