% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fileReferenceGDS,RAIDSparam-method}
\alias{fileReferenceGDS,RAIDSparam-method}
\title{A getter for the fileReferenceGDS slot in a RAIDSparam class}
\usage{
\S4method{fileReferenceGDS}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or a single \code{character} string representing
the file name of the Reference GDS file.
}
\description{
A function for getting the fileReferenceGDS slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the fileReferenceGDS slot for the object
fileReferenceGDS(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
