% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fileReferenceGDS<-,RAIDSparam-method}
\alias{fileReferenceGDS<-,RAIDSparam-method}
\title{Setter function for replacement of fileReferenceGDS slot in a RAIDSparam class}
\usage{
\S4method{fileReferenceGDS}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a single \code{character} string representing
the file name of the Reference GDS file. If specified, the file must exist.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of fileReferenceGDS slot in a
\code{RAIDSparam} class.
}
\examples{

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the existing file to the fileReferenceGDS slot in the object
fileReferenceGDS(paramDemo) <- file.path(dataDir, 
    "tests/ex1_good_small_1KG.gds")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
