% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fileReferenceGDS}
\alias{fileReferenceGDS}
\title{Generic function for getting the fileReferenceGDS slot in a class}
\usage{
fileReferenceGDS(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the fileReferenceGDS slot in the S4 object.
}
\description{
A generic function for getting the fileReferenceGDS slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(fileReferenceGDS="character"))

# Create an instance
obj <- new("MyClass", fileReferenceGDS="123")

# Call the generic (assuming a method is implemented)
# fileReferenceGDS(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
