% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{gds2tfam}
\alias{gds2tfam}
\title{create a file tfam file for plink from the gdsReference file}
\usage{
gds2tfam(gdsReference, listSample, pedOUT)
}
\arguments{
\item{gdsReference}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{listSample}{a \code{array} with the sample to keep TODO}

\item{pedOUT}{TODO a PATH and file name to the output file}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{export}
