% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDS1KGgenotypeFromSNPPileup}
\alias{generateGDS1KGgenotypeFromSNPPileup}
\title{TODO This function append the genotype and the file related to the
pileup}
\usage{
generateGDS1KGgenotypeFromSNPPileup(
  PATHGENO,
  listSamples,
  listPos,
  offset,
  minCov = 10,
  minProb = 0.999,
  seqError = 0.001,
  pedStudy,
  batch,
  studyDF,
  PATHGDSSAMPLE = NULL
)
}
\arguments{
\item{PATHGENO}{TODO a PATH to a directory with the a file for each
samples with the genotype.}

\item{listSamples}{a \code{array} with the sample to keep}

\item{listPos}{a \code{array}}

\item{offset}{a \code{integer} to adjust if the genome start at 0 or 1}

\item{minCov}{a \code{array} with the sample to keep}

\item{minProb}{a \code{array} with the sample to keep}

\item{seqError}{a \code{array} with the sample to keep}

\item{batch}{a \code{integer} corresponding}

\item{studyDF}{a \code{data.frame} with at least the column study.id,
study.desc and study.platform}

\item{PATHGDSSAMPLE}{TODO a PATH to a directory where a gds specific
to the samples with coverage info is keep}

\item{pedDF}{a \code{data.frame} with the sample info. Must have the column
sample.id, Name.ID, sex, pop.group, superPop and batch. The unique id of
pedDF is Name.ID and the row.name is Name.ID too.}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
