% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{generateMapSnvSel}
\alias{generateMapSnvSel}
\title{Generate the mapSNVSel file}
\usage{
generateMapSnvSel(cutOff = 0.01, fileSNV, fileLSNP, fileFREQ)
}
\arguments{
\item{cutOff}{a single \code{numeric} value, the cut off
for the frequency in at least one super population.}

\item{fileSNV}{}

\item{fileLSNP}{}

\item{fileFREQ}{}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
