% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{generateVCF}
\alias{generateVCF}
\title{Generate a VCF with the information from the SNPs that pass
a cut-off threshold}
\usage{
generateVCF(freqCutoff = NULL, chr = NULL, fileOut, pRAIDS)
}
\arguments{
\item{freqCutoff}{a single positive \code{numeric} specifying the cut-off to
keep a SNP. If \code{NULL}, all SNPs are retained. Default: \code{NULL}.}

\item{chr}{a \code{integer} between 1 to 22
chr1:22 specify SNP retained are on chr.
If \code{NULL}, all SNPs are retained. Default: \code{NULL}.}

\item{fileOut}{a \code{character} string representing the path and
the file name of the new VCF.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## Required library
library(gdsfmt)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## Demo 1KG Reference GDS file
fileGDS <- openfn.gds(file.path(dataDir,
                    "PopulationReferenceDemo.gds"))

## Output VCF file that will be created (temporary)
vcfFile <- file.path(tempdir(), "Demo_TMP_01.vcf")

## Create a VCF file with the SNV dataset present in the GDS file
## No cutoff on frequency, so all SNVs are saved
snvListVCF(gdsReference=fileGDS, fileOut=vcfFile, offset=0L,
                    freqCutoff=NULL)

## Close GDS file (IMPORTANT)
closefn.gds(fileGDS)

## Remove temporary VCF file
unlink(vcfFile, force=TRUE)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
