% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genome,RAIDSparam-method}
\alias{genome,RAIDSparam-method}
\title{A getter for the genome slot in a RAIDSparam class}
\usage{
\S4method{genome}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{character} string with one possible value: 'HG38'.
}
\description{
A function for getting the genome slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the genome slot for the object
genome(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
