% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genome}
\alias{genome}
\title{Generic function for getting the genome slot in a class}
\usage{
genome(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the genome slot in the S4 object.
}
\description{
A generic function for getting the genome slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(genome="character"))

# Create an instance
obj <- new("MyClass", genome="123")

# Call the generic (assuming a method is implemented)
# genome(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
