% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoSource,RAIDSparam-method}
\alias{genoSource,RAIDSparam-method}
\title{A getter for the genoSource slot in a RAIDSparam class}
\usage{
\S4method{genoSource}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string corresponding to the type
of file with the genotype and the allele information of the profile that
will be provided in the 'profileFile' slot. The valid options are: "VCF",
"generic", "snp-pileup", and "bam".
}
\description{
A function for getting the genoSource slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the genoSource slot for the object
genoSource(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
