% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoSource<-,RAIDSparam-method}
\alias{genoSource<-,RAIDSparam-method}
\title{Setter function for replacement of genoSource slot in a RAIDSparam class}
\usage{
\S4method{genoSource}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string corresponding to the type
of file with the genotype and the allele information of the profile that
will be provided in the 'profileFile' slot. The valid options are: "VCF",
"generic", "snp-pileup", and "bam".}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of genoSource slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new study type to the genoSource slot in the object
genoSource(paramDemo) <- "bam"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
