% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoSource}
\alias{genoSource}
\title{Generic function for getting the genoSource slot in a class}
\usage{
genoSource(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the genoSource slot in the S4 object.
}
\description{
A generic function for getting the genoSource slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(genoSource="character"))

# Create an instance
obj <- new("MyClass", genoSource="123")

# Call the generic (assuming a method is implemented)
# genoSource(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
