% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoType,RAIDSparam-method}
\alias{genoType,RAIDSparam-method}
\title{A getter for the genoType slot in a RAIDSparam class}
\usage{
\S4method{genoType}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string representing the field
in the GDS profile file that contains all the genotype information.
}
\description{
A function for getting the genoType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the genoType slot for the object
genoType(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
