% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoType<-,RAIDSparam-method}
\alias{genoType<-,RAIDSparam-method}
\title{Setter function for replacement of genoType slot in
a RAIDSparam class}
\usage{
\S4method{genoType}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string representing the field
in the GDS profile file that contains all the genotype information.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of genoType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the genoType slot in the object
genoType(paramDemo) <- "geno.REF"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
