% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{genoType}
\alias{genoType}
\title{Generic function for getting the genoType slot in a class}
\usage{
genoType(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the genoType slot in the S4 object.
}
\description{
A generic function for getting the genoType slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(genoType="character"))

# Create an instance
obj <- new("MyClass", genoType="123")

# Call the generic (assuming a method is implemented)
# genoType(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
