% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{getGeneBlock}
\alias{getGeneBlock}
\title{Get the block number for each SNV in snp.index}
\usage{
getGeneBlock(gdsRefAnnot, snp.index, blockID)
}
\arguments{
\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened 1KG SNV Annotation GDS file. RNA specific
Default: \code{NULL}.}

\item{snp.index}{TODO}

\item{blockID}{a \code{character} string corresponding to the block
identifier in \code{gdsRefAnnot}.  RNA specific
Default: \code{NULL}}
}
\value{
TODO a \code{vector} of \code{numeric} corresponding to the
block identifier
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
