% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{getMatrixPopSynthetic}
\alias{getMatrixPopSynthetic}
\title{Generate a matrix of synthetic group per pop}
\usage{
getMatrixPopSynthetic(pRAIDS)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} object with all the RAIDS
parameters}
}
\value{
a \code{matrix} containing the sample identifiers and where
each column is the name of a subcontinental population. The number of
row corresponds to the number of samples for each subcontinental population.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## Required library
library(gdsfmt)

## TODO example

## Path to the demo pedigree file is located in this package
## <- system.file("extdata", package="RAIDS")

## Demo 1KG Reference GDS file
##fileGDS <- openfn.gds(file.path(dataDir,
##                    "PopulationReferenceDemo.gds"))

## Output VCF file that will be created (temporary)
##vcfFile <- file.path(tempdir(), "Demo_TMP_01.vcf")

## Create a VCF file with the SNV dataset present in the GDS file
## No cutoff on frequency, so all SNVs are saved
##snvListVCF(gdsReference=fileGDS, fileOut=vcfFile, offset=0L,
##                    freqCutoff=NULL)

## Close GDS file (IMPORTANT)
##closefn.gds(fileGDS)

## Remove temporary VCF file
##unlink(vcfFile, force=TRUE)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
