% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{getRef1KGPop}
\alias{getRef1KGPop}
\title{TODO}
\usage{
getRef1KGPop(gds, popName)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{popName}{TODO}
}
\value{
\code{data.frame} TODO study.annot with study.annot == study.id and
with the column popName.
}
\description{
TODO
}
\examples{

## TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
