% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRef.R
\encoding{UTF-8}
\name{getRefSuperPop2}
\alias{getRefSuperPop2}
\title{Extract the specified column from the 1KG GDS 'sample.ref' node
for the reference profiles (real ancestry assignation)}
\usage{
getRefSuperPop2(pRAIDS = pRAIDS)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{vector} of \code{character} strings representing the content
of the extracted column for the 1KG GDS 'sample.ref' node. The values
represent the known ancestry assignation. The profile
identifiers are used as names for the \code{vector}.
}
\description{
The function extract the specified column for the 'sample.ref'
node present in the Reference GDS file. The column must be present in the
\code{data.frame} saved in the 'sample.ref' node. Only the information for
the reference profiles is returned. The values
represent the known ancestry assignation.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
