% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{getTableSNV}
\alias{getTableSNV}
\title{TODO}
\usage{
getTableSNV(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  minCov = 10,
  minProb = 0.999,
  eProb = 0.001
)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{gdsSample}{TODO}

\item{sampleCurrent}{A \code{string} corresponding to
the sample.id
use in LDpruning}

\item{study.id}{A \code{string} corresponding to the study
use in LDpruning}

\item{minCov}{an \code{integer} default 10}

\item{minProb}{an \code{numeric} betweeen 0 and 1}

\item{eProb}{an \code{numeric} betweeen 0 and 1 probability of sequencing error}
}
\value{
data.frame snp.pos TODO.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
