% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{groupChrPruning}
\alias{groupChrPruning}
\title{Merge the pruning files by chromosome in one file}
\usage{
groupChrPruning(pathPrunedGDS, filePref, fileOut)
}
\arguments{
\item{pathPrunedGDS}{a \code{character} string representing the path where
the pruned files for each chromosome are located.
The path must exists.}

\item{filePref}{TODO}

\item{fileOut}{a \code{character} string representing name of the output
file that will be created. THe file will contain the information for all
pruned chromosome. The file must have a ".rds" extension.}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
