% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{identifyRelative}
\alias{identifyRelative}
\title{Identify genetically unrelated patients in GDS 1KG file}
\usage{
identifyRelative(gds, maf = 0.05, thresh = 2^(-11/2), fileIBD, filePart)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{maf}{a single \code{numeric} representing the threshold for the minor
allele frequency. Only the SNPs with ">= maf" will be used.
Default: \code{0.05}.}

\item{thresh}{a single \code{numeric} representing the threshold value used
to decide if a pair of individuals is ancestrally divergent.
Default: \code{2^(-11/2)}.}

\item{fileIBD}{a \code{character} string representing the path and file
name of the RDS file that will be created. The RDS file will contain the
kinship coefficient between the patients.
The extension of the file must be '.rds'.}

\item{filePart}{a \code{character} string representing the path and file
name of the RDS file that will be created. The RDS file will contain the
information about the 1KG patients that are unrelated.
The extension of the file must be '.rds'.}
}
\value{
\code{NULL} invisibly.
}
\description{
The function identify patients that are genetically related in
the 1KG file. It generates a first RDS file with the list of unrelated
patient. It also generates a second RDS file with the kinship coefficient
between the patients.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
