% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{inferenceType,RAIDSparam-method}
\alias{inferenceType,RAIDSparam-method}
\title{A getter for the inferenceType slot in a RAIDSparam class}
\usage{
\S4method{inferenceType}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string representing the genotype
ancestry inference method.
}
\description{
A function for getting the inferenceType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the inferenceType slot for the object
inferenceType(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
