% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{inferenceType<-,RAIDSparam-method}
\alias{inferenceType<-,RAIDSparam-method}
\title{Setter function for replacement of inferenceType slot in
a RAIDSparam class}
\usage{
\S4method{inferenceType}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string representing the genotype
ancestry inference method. The two possible values are 'PCAknn' and
'haploAdmixture'.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of inferenceType slot in a
\code{RAIDSparam} class.
}
\examples{

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the inferenceType slot in the object
inferenceType(paramDemo) <- "haploAdmixture"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
