% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{inferenceType}
\alias{inferenceType}
\title{Generic function for getting the inferenceType slot in a class}
\usage{
inferenceType(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the inferenceType slot in the S4 object.
}
\description{
A generic function for getting the inferenceType slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(inferenceType="character"))

# Create an instance
obj <- new("MyClass", inferenceType="123")

# Call the generic (assuming a method is implemented)
# inferenceType(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
