% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{kList,RAIDSparam-method}
\alias{kList,RAIDSparam-method}
\title{A getter for the kList slot in a RAIDSparam class}
\usage{
\S4method{kList}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{K} parameter. The \emph{K} parameter
represents the number of neighbors used in the K-nearest neighbor analysis.
}
\description{
A function for getting the kList slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the kList slot for the object
kList(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
