% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{kList<-,RAIDSparam-method}
\alias{kList<-,RAIDSparam-method}
\title{Setter function for replacement of kList slot in
a RAIDSparam class}
\usage{
\S4method{kList}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{K} parameter. The \emph{K} parameter
represents the number of neighbors used in the K-nearest neighbor analysis.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of kList slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the kList slot in the object
kList(paramDemo) <-  c(2L, 3L, 4L, 5L)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
