% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{kList}
\alias{kList}
\title{Generic function for getting the kList slot in a class}
\usage{
kList(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the kList slot in the S4 object.
}
\description{
A generic function for getting the kList slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(kList="character"))

# Create an instance
obj <- new("MyClass", kList="123")

# Call the generic (assuming a method is implemented)
# kList(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
