% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{listPos,RAIDSparam-method}
\alias{listPos,RAIDSparam-method}
\title{A getter for the listPos slot in a RAIDSparam class}
\usage{
\S4method{listPos}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{integer} specifying the number of
threads to be used.
}
\description{
A function for getting the listPos slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the listPos slot for the object
listPos(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
