% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{listPos<-,RAIDSparam-method}
\alias{listPos<-,RAIDSparam-method}
\title{Setter function for replacement of listPos slot in
a RAIDSparam class}
\usage{
\S4method{listPos}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a \code{data.frame} containing 2 columns named:
"snp.chromosome" and "snp.position". The first column, called
"snp.chromosome", contains the name of the chromosome where the SNV
is located. The second column, called "snp.position", contains the
position of the SNV on the chromosome.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of listPos slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new values to the listPos slot in the object
listPos(paramDemo) <-data.frame(snp.chromosome=c("1", "1"), 
    snp.position=c(12232, 433333))

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
