% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{listPos}
\alias{listPos}
\title{Generic function for getting the listPos slot in a class}
\usage{
listPos(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the listPos slot in the S4 object.
}
\description{
A generic function for getting the listPos slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(listPos="character"))

# Create an instance
obj <- new("MyClass", listPos="123")

# Call the generic (assuming a method is implemented)
# listPos(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
