% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{minCov,RAIDSparam-method}
\alias{minCov,RAIDSparam-method}
\title{A getter for the minCov slot in a RAIDSparam class}
\usage{
\S4method{minCov}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{integer} representing the minimum
required coverage.
}
\description{
A function for getting the minCov slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the minCov slot for the object
minCov(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
