% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{minCov}
\alias{minCov}
\title{Generic function for getting the minCov slot in a class}
\usage{
minCov(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the minCov slot in the S4 object.
}
\description{
A generic function for getting the minCov slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(minCov="character"))

# Create an instance
obj <- new("MyClass", minCov="123")

# Call the generic (assuming a method is implemented)
# minCov(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
