% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{minProb,RAIDSparam-method}
\alias{minProb,RAIDSparam-method}
\title{A getter for the minProb slot in a RAIDSparam class}
\usage{
\S4method{minProb}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
}
\description{
A function for getting the minProb slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the minProb slot for the object
minProb(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
