% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{minProb}
\alias{minProb}
\title{Generic function for getting the minProb slot in a class}
\usage{
minProb(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the minProb slot in the S4 object.
}
\description{
A generic function for getting the offset slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(minProb="character"))

# Create an instance
obj <- new("MyClass", minProb="123")

# Call the generic (assuming a method is implemented)
# minProb(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
