% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{nbSim,RAIDSparam-method}
\alias{nbSim,RAIDSparam-method}
\title{A getter for the nbSim slot in a RAIDSparam class}
\usage{
\S4method{nbSim}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a positive \code{integer} representing TODO.
}
\description{
A function for getting the nbSim slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the nbSim slot for the object
nbSim(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
