% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{nbSim<-,RAIDSparam-method}
\alias{nbSim<-,RAIDSparam-method}
\title{Setter function for replacement of nbSim slot in
a RAIDSparam class}
\usage{
\S4method{nbSim}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a positive \code{integer} representing TODO.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of nbSim slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the nbSim slot in the object
nbSim(paramDemo) <- 33L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
