% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{nbSim}
\alias{nbSim}
\title{Generic function for getting the nbSim slot in a class}
\usage{
nbSim(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the nbSim slot in the S4 object.
}
\description{
A generic function for getting the nbSim slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(nbSim="character"))

# Create an instance
obj <- new("MyClass", nbSim="123")

# Call the generic (assuming a method is implemented)
# nbSim(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
