% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{np,RAIDSparam-method}
\alias{np,RAIDSparam-method}
\title{A getter for the np slot in a RAIDSparam class}
\usage{
\S4method{np}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{integer} specifying the number of
threads to be used.
}
\description{
A function for getting the np slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the np slot for the object
np(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
