% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{np}
\alias{np}
\title{Generic function for getting the np slot in a class}
\usage{
np(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the np slot in the S4 object.
}
\description{
A generic function for getting the np slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(np="character"))

# Create an instance
obj <- new("MyClass", np="123")

# Call the generic (assuming a method is implemented)
# np(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
