% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{offset<-,RAIDSparam-method}
\alias{offset<-,RAIDSparam-method}
\title{Setter function for replacement of offset slot in
a RAIDSparam class}
\usage{
\S4method{offset}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{integer} that is added to the SNP position to
switch from 0-based to 1-based coordinate when needed (or reverse).}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of offset slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the offset slot in the object
offset(paramDemo) <- 0L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
