% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{paramAncestry<-,RAIDSparam-method}
\alias{paramAncestry<-,RAIDSparam-method}
\title{Setter function for replacement of paramAncestry slot in a RAIDSparam class}
\usage{
\S4method{paramAncestry}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{vector} of positive \code{integer} values
representing the length of the chromosomes.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of paramAncestry slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new chrInfo to the paramAncestry slot in the object
paramAncestry(paramDemo) <- list(ScanBamParam=NULL, PileupParam=NULL, 
    yieldSize=1000023)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
