% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{paramAncestry}
\alias{paramAncestry}
\title{Generic function for getting the paramAncestry slot in a class}
\usage{
paramAncestry(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the paramAncestry slot in the S4 object.
}
\description{
A generic function for getting the paramAncestry slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(paramAncestry="character"))

# Create an instance
obj <- new("MyClass", paramAncestry="123")

# Call the generic (assuming a method is implemented)
# paramAncestry(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
