% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pathProfileGDS,RAIDSparam-method}
\alias{pathProfileGDS,RAIDSparam-method}
\title{A getter for the pathProfileGDS slot in a RAIDSparam class}
\usage{
\S4method{pathProfileGDS}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or a single \code{character} string representing the
path to the directory where the GDS Profile files will be created.
}
\description{
A function for getting the pathProfileGDS slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pathProfileGDS slot for the object
pathProfileGDS(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
