% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pathProfileGDS<-,RAIDSparam-method}
\alias{pathProfileGDS<-,RAIDSparam-method}
\title{Setter function for replacement of pathProfileGDS slot in a RAIDSparam class}
\usage{
\S4method{pathProfileGDS}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a single \code{character} string
representing the path to the directory where the GDS Profile files will
be created. If specified, the directory must exist.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of pathProfileGDS slot in a
\code{RAIDSparam} class.
}
\examples{

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the existing path to the pathProfileGDS slot in the object
pathProfileGDS(paramDemo) <- file.path(dataDir)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
