% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pathProfileGDS}
\alias{pathProfileGDS}
\title{Generic function for getting the pathProfileGDS slot in a class}
\usage{
pathProfileGDS(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the pathProfileGDS slot in the S4 object.
}
\description{
A generic function for getting the pathProfileGDS slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(pathProfileGDS="character"))

# Create an instance
obj <- new("MyClass", pathProfileGDS="123")

# Call the generic (assuming a method is implemented)
# pathProfileGDS(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
