% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAalgorithm,RAIDSparam-method}
\alias{PCAalgorithm,RAIDSparam-method}
\title{A getter for the PCAalgorithm slot in a RAIDSparam class}
\usage{
\S4method{PCAalgorithm}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{character} string representing the algorithm to use
with the \link[SNPRelate]{snpgdsPCA} function.
}
\description{
A function for getting the PCAalgorithm slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the PCAalgorithm slot for the object
PCAalgorithm(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
