% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAalgorithm<-,RAIDSparam-method}
\alias{PCAalgorithm<-,RAIDSparam-method}
\title{Setter function for replacement of PCAalgorithm slot in
a RAIDSparam class}
\usage{
\S4method{PCAalgorithm}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{character} string representing the algorithm to
use with the \link[SNPRelate]{snpgdsPCA} function. The algorithm must be
implemented and available to the \link[SNPRelate]{snpgdsPCA} function.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of PCAalgorithm slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the PCAalgorithm slot in the object
PCAalgorithm(paramDemo) <-  "randomized"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
