% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAalgorithm}
\alias{PCAalgorithm}
\title{Generic function for getting the PCAalgorithm slot in a class}
\usage{
PCAalgorithm(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the PCAalgorithm slot in the S4 object.
}
\description{
A generic function for getting the PCAalgorithm slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(PCAalgorithm="character"))

# Create an instance
obj <- new("MyClass", PCAalgorithm="123")

# Call the generic (assuming a method is implemented)
# PCAalgorithm(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
