% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pcaList,RAIDSparam-method}
\alias{pcaList,RAIDSparam-method}
\title{A getter for the pcaList slot in a RAIDSparam class}
\usage{
\S4method{pcaList}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{vector} of positive \code{integer} representing the values
tested for the  \emph{D} parameter. The \emph{D} parameter represents the
number of dimensions used in the PCA analysis.
}
\description{
A function for getting the pcaList slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pcaList slot for the object
pcaList(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
