% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pcaList<-,RAIDSparam-method}
\alias{pcaList<-,RAIDSparam-method}
\title{Setter function for replacement of pcaList slot in
a RAIDSparam class}
\usage{
\S4method{pcaList}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{D} parameter. The \emph{D} parameter
represents the number of dimensions used in the PCA analysis.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of pcaList slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the pcaList slot in the object
pcaList(paramDemo) <-  c(2L, 3L, 4L, 5L)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
