% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pcaList}
\alias{pcaList}
\title{Generic function for getting the pcaList slot in a class}
\usage{
pcaList(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the pcaList slot in the S4 object.
}
\description{
A generic function for getting the pcaList slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(pcaList="character"))

# Create an instance
obj <- new("MyClass", pcaList="123")

# Call the generic (assuming a method is implemented)
# pcaList(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
