% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAmissingRate,RAIDSparam-method}
\alias{PCAmissingRate,RAIDSparam-method}
\title{A getter for the PCAmissingRate slot in a RAIDSparam class}
\usage{
\S4method{PCAmissingRate}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a positive \code{numeric} representing the maximum missing rate
accepted to use SNPs in the PCA analysis done with the
\link[SNPRelate]{snpgdsPCA} function. If \code{NaN}, no missing threshold.
}
\description{
A function for getting the PCAmissingRate slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the PCAmissingRate slot for the object
PCAmissingRate(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
