% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAmissingRate<-,RAIDSparam-method}
\alias{PCAmissingRate<-,RAIDSparam-method}
\title{Setter function for replacement of PCAmissingRate slot in
a RAIDSparam class}
\usage{
\S4method{PCAmissingRate}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a positive \code{numeric} representing the maximum missing
rate accepted to use SNPs in the PCA analysis done with
the \link[SNPRelate]{snpgdsPCA} function. If \code{NaN}, no
missing threshold.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of PCAmissingRate slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the PCAmissingRate slot in the object
PCAmissingRate(paramDemo) <-  0.03322

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
