% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{PCAmissingRate}
\alias{PCAmissingRate}
\title{Generic function for getting the PCAmissingRate slot in a class}
\usage{
PCAmissingRate(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the PCAmissingRate slot in the S4 object.
}
\description{
A generic function for getting the PCAmissingRate slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(PCAmissingRate="character"))

# Create an instance
obj <- new("MyClass", PCAmissingRate="123")

# Call the generic (assuming a method is implemented)
# PCAmissingRate(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
