% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy,RAIDSparam-method}
\alias{pedStudy,RAIDSparam-method}
\title{A getter for the pedStudy slot in a RAIDSparam class}
\usage{
\S4method{pedStudy}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{data.frame}.
}
\description{
A function for getting the pedStudy slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pedStudy slot for the object
pedStudy(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
