% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy<-}
\alias{pedStudy<-}
\title{Generic function for replacement of pedStudy slot in a class}
\usage{
pedStudy(x) <- value
}
\arguments{
\item{x}{a S4 object.}

\item{value}{the new value to assign or update.}
}
\value{
the modified S4 object when the new value is valid.
}
\description{
A generic function for replacement of pedStudy slot in a
S4 object.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
