% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy}
\alias{pedStudy}
\title{Generic function for getting the pedStudy slot in a class}
\usage{
pedStudy(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the pedStudy slot in the S4 object.
}
\description{
A generic function for getting the pedStudy slot in a
S4 object.
}
\examples{
# Define a dummy class to show usage
setClass("MyClass", slots = list(studyDF="character"))

# Create an instance
obj <- new("MyClass", studyDF="123")

# Call the generic (assuming a method is implemented)
# studyDF(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
