% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phase,RAIDSparam-method}
\alias{phase,RAIDSparam-method}
\title{A getter for the phase slot in a RAIDSparam class}
\usage{
\S4method{phase}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{logical}  TODO
}
\description{
A function for getting the phase slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the phase slot for the object
phase(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
