% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phase<-,RAIDSparam-method}
\alias{phase<-,RAIDSparam-method}
\title{Setter function for replacement of phase slot in
a RAIDSparam class}
\usage{
\S4method{phase}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{logical} TODO}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of phase slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the phase slot in the object
phase(paramDemo) <-  TRUE

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
