% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phase}
\alias{phase}
\title{Generic function for getting the phaseType slot in a class}
\usage{
phase(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the phase slot in the S4 object.
}
\description{
A generic function for getting the phase slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(phase="character"))

# Create an instance
obj <- new("MyClass", phase="123")

# Call the generic (assuming a method is implemented)
# phase(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
