% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phaseType,RAIDSparam-method}
\alias{phaseType,RAIDSparam-method}
\title{A getter for the phaseType slot in a RAIDSparam class}
\usage{
\S4method{phaseType}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string TODO
}
\description{
A function for getting the phaseType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the phaseType slot for the object
phaseType(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
