% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phaseType<-,RAIDSparam-method}
\alias{phaseType<-,RAIDSparam-method}
\title{Setter function for replacement of phaseType slot in
a RAIDSparam class}
\usage{
\S4method{phaseType}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string TODO}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of phaseType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the phaseType slot in the object
phaseType(paramDemo) <- "test"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
