% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{phaseType}
\alias{phaseType}
\title{Generic function for getting the phaseType slot in a class}
\usage{
phaseType(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the phaseType slot in the S4 object.
}
\description{
A generic function for getting the phaseType slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(phaseType="character"))

# Create an instance
obj <- new("MyClass", phaseType="123")

# Call the generic (assuming a method is implemented)
# phaseType(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
