% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{posOffset}
\alias{posOffset}
\title{Generic function for getting the posOffset slot in a class}
\usage{
posOffset(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the posOffset slot in the S4 object.
}
\description{
A generic function for getting the offposOffsetset slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(posOffset="character"))

# Create an instance
obj <- new("MyClass", posOffset="123")

# Call the generic (assuming a method is implemented)
# posOffset(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
