% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pRecomb,RAIDSparam-method}
\alias{pRecomb,RAIDSparam-method}
\title{A getter for the pRecomb slot in a RAIDSparam class}
\usage{
\S4method{pRecomb}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{numeric} between \code{0} and \code{1}
that represents the frequency of phase switching in the synthetic profiles.
}
\description{
A function for getting the pRecomb slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pRecomb slot for the object
pRecomb(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
