% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pRecomb}
\alias{pRecomb}
\title{Generic function for getting the pRecomb slot in a class}
\usage{
pRecomb(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the pRecomb slot in the S4 object.
}
\description{
A generic function for getting the pRecomb slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(pRecomb="character"))

# Create an instance
obj <- new("MyClass", pRecomb="123")

# Call the generic (assuming a method is implemented)
# pRecomb(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
