% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{prefix,RAIDSparam-method}
\alias{prefix,RAIDSparam-method}
\title{A getter for the prefix slot in a RAIDSparam class}
\usage{
\S4method{prefix}{RAIDSparam}(x)

\S4method{prefix}{RAIDSparam}(x)

\S4method{prefix}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string representing TODO.

a single \code{character} string representing TODO.

a single \code{character} string representing TODO.
}
\description{
A function for getting the prefix slot in a
\code{RAIDSparam} class.

A function for getting the prefix slot in a
\code{RAIDSparam} class.

A function for getting the prefix slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the prefix slot for the object
prefix(paramDemo)


## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the prefix slot for the object
prefix(paramDemo)


## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the prefix slot for the object
prefix(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
