% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{prefix<-,RAIDSparam-method}
\alias{prefix<-,RAIDSparam-method}
\title{Setter function for replacement of prefix slot in
a RAIDSparam class}
\usage{
\S4method{prefix}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string representing TODO.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of prefix slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the prefix slot in the object
prefix(paramDemo) <- "test"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
