% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{prefix}
\alias{prefix}
\title{Generic function for getting the prefix slot in a class}
\usage{
prefix(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the prefix slot in the S4 object.
}
\description{
A generic function for getting the prefix slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(prefix="character"))

# Create an instance
obj <- new("MyClass", prefix="123")

# Call the generic (assuming a method is implemented)
# prefix(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
