% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{prepPed1KG}
\alias{prepPed1KG}
\title{Prepare the pedigree file using pedigree information from 1KG}
\usage{
prepPed1KG(pedFile, PATHGENO = file.path("data", "sampleGeno"), batch.v = 0L)
}
\arguments{
\item{pedFile}{a \code{character} string representing the path and
file name of the pedigree file from 1KG. The file must exist.}

\item{PATHGENO}{a \code{character} string representing the path where
the 1K genotyping files for each sample are located. Only the samples with
associated genotyping files are retained in the creation of the final
\code{data.frame}. The name of the genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file.
Default: \code{"./data/sampleGeno"}.}

\item{batch.v}{a\code{integer} that uniquely identifies the source of the
pedigree information. The 1KG is usually \code{0L}. Default: \code{0L}.}
}
\value{
A \code{data.frame} containing the needed pedigree information
from 1K.
}
\description{
Using the pedigree file from 1KG, this function extracts
needed information and formats it into a \code{data.frame} so in can
be used in following steps of the ancestry inference process. The
function also requires that the genotyping files associated to each
sample be available in a specified directory.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## Demo pedigree file
pedDemoFile <- file.path(data.dir, "PedigreeDemo.ped")

## Create a data.frame containing the information of the retained
## samples (samples with existing genotyping files)
prepPed1KG(pedFile=pedDemoFile, PATHGENO=data.dir, batch.v=0L)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
