% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{prepPedSynthetic1KG}
\alias{prepPedSynthetic1KG}
\title{TODO}
\usage{
prepPedSynthetic1KG(gds, gdsSample, study.id, popName)
}
\arguments{
\item{gds}{an object of class
\code{\link[gdsfmt:gds.class]{gdsfmt::gds.class}}, the opened 1 KG GDS file.}

\item{gdsSample}{an object of class
\code{\link[gdsfmt:gds.class]{gdsfmt::gds.class}}, the opened GDS Sample
file.}

\item{study.id}{a \code{character} string}

\item{popName}{a \code{character} string representing the name of the
column from the \code{data.frame} stored in the "sample.annot" node of the
1KG GDS file. The column must be present in the \code{data.frame}.}
}
\value{
\code{data.frame} TODO study.annot with study.annot == study.id and
with the column popName.
}
\description{
TODO
}
\examples{

## TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
