% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{prepPMatrix}
\alias{prepPMatrix}
\title{Generate the row for the ADMIXTURE P-matrax}
\usage{
prepPMatrix(anchor, snpId, matGr = NULL, pRAIDS)
}
\arguments{
\item{anchor}{a \code{vector} of \code{character} representing the
ancestral group for each profile from the reference
where sample.ref == 1. If the admixture proportion is higher than
the threshold for one ancestral group, the value is the name
of the ancestral group; else "-".}

\item{snpId}{a \code{vector} of \code{character} representing the list
of snp.id from gdsReference to keep.}

\item{matGr}{a \code{matrix} containing the sample identifiers and where
each column is the name of a subcontinental population. The number of
row corresponds to the number of samples for each subcontinental population.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters.}
}
\value{
a \code{list} TODO.
}
\description{
This function write with P-matrix
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
