% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{prepSynthetic}
\alias{prepSynthetic}
\title{TODO}
\usage{
prepSynthetic(
  gdsSampleFile,
  listSampleRef,
  data.id.profile,
  studyDF,
  nbSim = 1,
  prefId = "",
  pRecomb = 0.01,
  minProb = 0.999,
  seqError = 0.001
)
}
\arguments{
\item{gdsSampleFile}{the path of an object of class \code{gds} related to
the sample}

\item{listSampleRef}{a \code{string} TODO}

\item{data.id.profile}{a \code{string} TODO}

\item{studyDF}{TODO}

\item{nbSim}{a \code{integer} TODO. Default: \code{1}.}

\item{prefId}{a \code{string} TODO}

\item{pRecomb}{a \code{numeric} between 0 and 1 TODO. Default: \code{0.01}.}

\item{minProb}{a \code{numeric} TODO. Default: \code{0.999}.}

\item{seqError}{a \code{numeric} TODO. Default: \code{0.001}.}
}
\value{
TODO a \code{vector} of \code{character} string
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
