% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{processBlockChr}
\alias{processBlockChr}
\title{TODO}
\usage{
processBlockChr(snpKeep, pathBlock, superPop, chr)
}
\arguments{
\item{snpKeep}{TODO}

\item{pathBlock}{TODO}

\item{superPop}{TODO}

\item{chr}{TODO}
}
\value{
the a \code{array} with the sample from pedDF keept
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
