% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{profileFile<-,RAIDSparam-method}
\alias{profileFile<-,RAIDSparam-method}
\title{Setter function for replacement of profileFile slot in a RAIDSparam class}
\usage{
\S4method{profileFile}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a \code{character} string representing the
path to the file
with genotype and the allele information of the profile. The format of
the accepted file is determined by the value in the 'genoSource' slot. A
profile would have a file with extension "vcf.gz" when 'genoSource' slot
is "VCF". If 'genoSource' is "generic" or "snp-pileup", then ".txt.gz". If
'genoSource' is "bam", then ".bam" (the file needs to be indexed with
an existing corresponding ".bai" file).}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of profileFile slot in a
\code{RAIDSparam} class.
}
\examples{

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Adjust the genoSource slot to the updated format file
genoSource(paramDemo) <- "generic"

## Assign the existing profile file to the profileFile slot in the object
profileFile(paramDemo) <- file.path(dataDir, "tests/ex1.txt.gz")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
