% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{profileFile}
\alias{profileFile}
\title{Generic function for getting the profileFile slot in a class}
\usage{
profileFile(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the profileFile slot in the S4 object.
}
\description{
A generic function for getting the profileFile slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(profileFile="character"))

# Create an instance
obj <- new("MyClass", profileFile="123")

# Call the generic (assuming a method is implemented)
# profileFile(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
