% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{projectSample2PCA}
\alias{projectSample2PCA}
\title{TODO}
\usage{
projectSample2PCA(gds, listPCA, sample.current, np = 1)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{listPCA}{a \code{list}  with with two objects
pca.unrel -> \code{snpgdsPCAClass}
and a snp.load -> \code{snpgdsPCASNPLoading}}

\item{sample.current}{the sample.id to project in the PCA}

\item{np}{a single positive \code{integer} representing the number of
threads. Default: \code{1}.}
}
\value{
a \code{snpgdsPCAClass} object, a \code{list} that contains:
\itemize{
   \item{sample.id} {the sample ids used in the analysis}
   \item{snp.id} {the SNP ids used in the analysis}
   \item{eigenvalues} {eigenvalues}
   \item{eigenvect} {eigenvactors, “# of samples” x “eigen.cnt”}
   \item{TraceXTX} {the trace of the genetic covariance matrix}
   \item{Bayesian} {whether use bayerisan normalization}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")
## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
