% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{pruning1KG.Chr}
\alias{pruning1KG.Chr}
\title{Find the pruned snv in 1KG by chr}
\usage{
pruning1KG.Chr(
  gds,
  method = "corr",
  listSamples = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE,
  chr = NULL,
  minAF = NULL,
  outPref = "pruned_1KG",
  keepObj = FALSE
)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{method}{}

\item{listSamples}{}

\item{slide.max.bp.v}{}

\item{ld.threshold.v}{}

\item{np}{}

\item{verbose.v}{}

\item{chr}{}

\item{minAF}{}

\item{outPref}{}

\item{keepObj}{}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
