% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pruningMethod<-,RAIDSparam-method}
\alias{pruningMethod<-,RAIDSparam-method}
\title{Setter function for replacement of pruningMethod slot in
a RAIDSparam class}
\usage{
\S4method{pruningMethod}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime", and "composite".}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of pruningMethod slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the pruningMethod slot in the object
pruningMethod(paramDemo) <- "r"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
