% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pruningMethod}
\alias{pruningMethod}
\title{Generic function for getting the pruningMethod slot in a class}
\usage{
pruningMethod(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the pruningMethod slot in the S4 object.
}
\description{
A generic function for getting the pruningMethod slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(pruningMethod="character"))

# Create an instance
obj <- new("MyClass", pruningMethod="123")

# Call the generic (assuming a method is implemented)
# pruningMethod(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
