% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{readGenoVCF3}
\alias{readGenoVCF3}
\title{Read a VCF file with the genotypes use for the ancestry call}
\usage{
readGenoVCF3(pRAIDS)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{data.frame} containing at least:
\describe{
\item{Chromosome}{ a \code{numeric} representing the name of
the chromosome}
\item{Position}{ a \code{numeric} representing the position on the
chromosome}
\item{ref}{ a \code{character} string representing the reference nucleotide}
\item{alt}{ a \code{character} string representing the alternative
nucleotide}
\item{index}{ a \code{integer} index of the data.frame}
\item{refIndex}{ a \code{integer} index of the reference SNV}
\item{geno}{ a \code{character} string representing the genotype}
\item{phase}{ a \code{character} representing the phase state}
\item{g}{ a \code{integer} representing the genotype state}
\item{p}{ a \code{integer} representing the phase state}
}
}
\description{
The function reads VCF file and returns a data frame
containing the information about the read counts for the SNVs present in
the file.
}
\examples{


## Directory where demo SNP-pileup file
dataDir <- system.file("extdata/example/snpPileup", package="RAIDS")

## The SNP-pileup file
snpPileupFile <- file.path(dataDir, "ex1.vcf.gz")

info <- RAIDS:::readGenoVCF(fileName=snpPileupFile)
head(info)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
