% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{readSNVFileGeneric}
\alias{readSNVFileGeneric}
\title{TODO}
\usage{
readSNVFileGeneric(fileName, offset = 0L)
}
\arguments{
\item{fileName}{File name with the path to a
csv with the columns:
Chromosome,Position,Ref,Alt,count,File1R,File1A}

\item{offset}{TODO}
}
\value{
the a \code{data.frame} containing at least:
\itemize{
\item{Chromosome} {TODO}
\item{Position} {TODO}
\item{Ref}
\item{Alt}
\item{File1R} {deep of the reference allele}
\item{File1A} {deep of the alternative allele}
\item{count} {Total deep at the position}
}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
