% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{readSNVPileupFile}
\alias{readSNVPileupFile}
\title{TODO}
\usage{
readSNVPileupFile(fileName, offset = 0L)
}
\arguments{
\item{fileName}{Output from snp-pileup
must csv with the columns:
Chromosome,Position,Ref,Alt,File1R,File1A,File1E,File1D}

\item{offset}{TODO}
}
\value{
the a \code{data.frame} containing at least:
\itemize{
\item{Chromosome} {TODO}
\item{Position} {TODO}
\item{File1R} {TODO}
\item{File1A} {TODO}
\item{count} {TODO}
}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
