% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{reference<-,RAIDSparam-method}
\alias{reference<-,RAIDSparam-method}
\title{Setter function for replacement of reference slot in a RAIDSparam class}
\usage{
\S4method{reference}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string with two possible values:
'1KGv1.0', '1k_hgdpV0.1'. It specifies the type of inference.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of reference slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new reference to the reference slot in the object
reference(paramDemo) <- "1k_hgdpV0.1"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
