% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{reference}
\alias{reference}
\title{Generic function for getting the reference slot in a class}
\usage{
reference(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the reference slot in the S4 object.
}
\description{
A generic function for getting the reference slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(reference="character"))

# Create an instance
obj <- new("MyClass", reference="123")

# Call the generic (assuming a method is implemented)
# reference(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
