% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{sampleRef,RAIDSparam-method}
\alias{sampleRef,RAIDSparam-method}
\title{A getter for the inferenceType slot in a RAIDSparam class}
\usage{
\S4method{sampleRef}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or a \code{vector} of \code{character} strings that
represent a subset of sample identifiers from the GDS Reference File.
Those samples will be used for the ancestry inference. If \code{NULL}, all
the samples are retained.
}
\description{
A function for getting the sampleRef slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the sampleRef slot for the object
sampleRef(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
