% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{sampleRef}
\alias{sampleRef}
\title{Generic function for getting the sampleRef slot in a class}
\usage{
sampleRef(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the sampleRef slot in the S4 object.
}
\description{
A generic function for getting the sampleRef slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(sampleRef="character"))

# Create an instance
obj <- new("MyClass", sampleRef="123")

# Call the generic (assuming a method is implemented)
# sampleRef(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
