% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic2.R
\encoding{UTF-8}
\name{selectForSynthetic}
\alias{selectForSynthetic}
\title{Random selection of a specific number of reference profiles in each
subcontinental population present in the refrence GDS file.
( call select1KGPopForSynthetic or selectHGDP1kgForSynthetic)}
\usage{
selectForSynthetic(pRAIDS, nbProfiles = 30L)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}

\item{nbProfiles}{a single positive \code{integer} representing the number
of samples that will be selected for each subcontinental population present
in the 1KG GDS file. If the number of samples in a specific subcontinental
population is smaller than the \code{nbProfiles}, the number of samples
selected in this
subcontinental population will correspond to the size of this population.}
}
\value{
a \code{data.frame} containing those columns:
\describe{
\item{sample.id}{ a \code{character} string representing the sample
identifier. }
\item{pop.group}{ a \code{character} string representing the
subcontinental population assigned to the sample. }
\item{superPop}{ a \code{character} string representing the
super-population assigned to the sample. }
}
}
\description{
The function randomly selects a fixed number of reference
for each subcontinental population present in the 1KG GDS file. When a
subcontinental population has less samples than the fixed number, all
samples from the subcontinental population are selected.
}
\examples{
## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata", package="RAIDS")


#################################################################
## The 1KG GDS file and the 1KG SNV Annotation GDS file
## need to be located in the same directory
## Note that the 1KG GDS file used for this example is a
## simplified version and CANNOT be used for any real analysis
#################################################################
path1KG <- file.path(dataDir, "tests")

fileReferenceGDS <- file.path(path1KG, "ex1_good_small_1KG.gds")
fileAnnotGDS <- file.path(path1KG, "ex1_good_small_1KG_Annot.gds")

#################################################################
## The Sample SNP pileup files (one per sample) need
## to be located in the same directory.
#################################################################
demoProfileEx1 <- file.path(dataDir, "example", "snpPileup", "ex1.txt.gz")

#################################################################
## The path where the Profile GDS Files (one per sample)
## will be created need to be specified.
#################################################################
pathProfileGDS <- file.path(tempdir(), "out.tmp")

####################################################################
## Fix seed to ensure reproducible results
####################################################################
set.seed(3043)

pRAIDS <- RAIDS:::paramRAIDS(profileFile=demoProfileEx1,
    genoSource="snp-pileup",
    pathProfileGDS=pathProfileGDS,
    fileReferenceGDS=fileReferenceGDS,
    fileReferenceAnnotGDS=fileAnnotGDS
    )

## The number of samples needed by subcontinental population
## The number is small for demonstration purpose
nbProfiles <- 5L

fileGDS <- file.path(dataDir, "PopulationReferenceDemo.gds")

## Extract a selected number of random samples
## for each subcontinental population
## In the 1KG GDS Demo file, there is one subcontinental population
dataR <- selectForSynthetic(pRAIDS=pRAIDS, nbProfiles=nbProfiles)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
