% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic2.R
\encoding{UTF-8}
\name{selectHGDP1kgForSynthetic}
\alias{selectHGDP1kgForSynthetic}
\title{Random selection of a specific number of reference profiles in each
subcontinental population present in the 1KG GDS file ( same as select1KGPop
but the function doesn't need gds object as parameters but the file name
of the referenceGDS )}
\usage{
selectHGDP1kgForSynthetic(pRAIDS, nbProfiles = 30L)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}

\item{nbProfiles}{a single positive \code{integer} representing the number
of samples that will be selected for each subcontinental population present
in the 1KG GDS file. If the number of samples in a specific subcontinental
population is smaller than the \code{nbProfiles}, the number of samples
selected in this
subcontinental population will correspond to the size of this population.}
}
\value{
a \code{data.frame} containing those columns:
\describe{
\item{sample.id}{ a \code{character} string representing the sample
identifier. }
\item{pop.group}{ a \code{character} string representing the
subcontinental population assigned to the sample. }
\item{superPop}{ a \code{character} string representing the
super-population assigned to the sample. }
}
}
\description{
The function randomly selects a fixed number of reference
for each subcontinental population present in the 1KG GDS file. When a
subcontinental population has less samples than the fixed number, all
samples from the subcontinental population are selected.
}
\examples{
## TODO change the example




}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
