% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{selParaPCAUpQuartile}
\alias{selParaPCAUpQuartile}
\title{TODO}
\usage{
selParaPCAUpQuartile(
  matKNN.All,
  pedCall,
  refCall,
  predCall,
  listCall,
  kList = 3:15,
  pcaList = 2:15
)
}
\arguments{
\item{matKNN.All}{TODO see it is rbind matKNN of the
computeKNNSuperPoprSynthetic return from group synthetic data}

\item{pedCall}{TODO see return of prepPedSynthetic1KG}

\item{refCall}{TODO column name in pedCall with the call}

\item{predCall}{TODO column name in matKNN with the call}

\item{listCall}{TODO array of the possible call}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
