% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{seqError,RAIDSparam-method}
\alias{seqError,RAIDSparam-method}
\title{A getter for the seqError slot in a RAIDSparam class}
\usage{
\S4method{seqError}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error.
}
\description{
A function for getting the seqError slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the seqError slot for the object
seqError(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
