% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{seqError}
\alias{seqError}
\title{Generic function for getting the seqError slot in a class}
\usage{
seqError(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the seqError slot in the S4 object.
}
\description{
A generic function for getting the seqError slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(seqError="character"))

# Create an instance
obj <- new("MyClass", seqError="123")

# Call the generic (assuming a method is implemented)
# seqError(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
