% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{seqErrorSyn,RAIDSparam-method}
\alias{seqErrorSyn,RAIDSparam-method}
\title{A getter for the seqErrorSyn slot in a RAIDSparam class}
\usage{
\S4method{seqErrorSyn}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error for synthetic.
}
\description{
A function for getting the seqErrorSyn slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the seqErrorSyn slot for the object
seqErrorSyn(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
