% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{seqErrorSyn<-,RAIDSparam-method}
\alias{seqErrorSyn<-,RAIDSparam-method}
\title{Setter function for replacement of seqErrorSyn slot in
a RAIDSparam class}
\usage{
\S4method{seqErrorSyn}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error for synthetic.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of seqErrorSyn slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the seqErrorSyn slot in the object
seqErrorSyn(paramDemo) <- 0.03

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
