% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{seqErrorSyn}
\alias{seqErrorSyn}
\title{Generic function for getting the seqErrorSyn slot in a class}
\usage{
seqErrorSyn(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the seqErrorSyn slot in the S4 object.
}
\description{
A generic function for getting the seqErrorSyn slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(seqErrorSyn="character"))

# Create an instance
obj <- new("MyClass", seqErrorSyn="123")

# Call the generic (assuming a method is implemented)
# seqErrorSyn(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
