% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{slideWindowMaxBP,RAIDSparam-method}
\alias{slideWindowMaxBP,RAIDSparam-method}
\title{A getter for the slideWindowMaxBP slot in a RAIDSparam class}
\usage{
\S4method{slideWindowMaxBP}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{integer} that represents the maximum
basepairs (bp) in the sliding window. This parameter is used for
the LD pruning done by the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.
}
\description{
A function for getting the slideWindowMaxBP slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the slideWindowMaxBP slot for the object
slideWindowMaxBP(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
