% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{slideWindowMaxBP<-,RAIDSparam-method}
\alias{slideWindowMaxBP<-,RAIDSparam-method}
\title{Setter function for replacement of slideWindowMaxBP slot in
a RAIDSparam class}
\usage{
\S4method{slideWindowMaxBP}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single positive \code{integer} that represents the maximum
basepairs (bp) in the sliding window. This parameter is used
for the LD pruning done by the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of slideWindowMaxBP slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the slideWindowMaxBP slot in the object
slideWindowMaxBP(paramDemo) <- 1000L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
