% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{slideWindowMaxBP}
\alias{slideWindowMaxBP}
\title{Generic function for getting the slideWindowMaxBP slot in a class}
\usage{
slideWindowMaxBP(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the slideWindowMaxBP slot in the S4 object.
}
\description{
A generic function for getting the slideWindowMaxBP slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(slideWindowMaxBP="character"))

# Create an instance
obj <- new("MyClass", slideWindowMaxBP="123")

# Call the generic (assuming a method is implemented)
# slideWindowMaxBP(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
