% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{snvListVCF}
\alias{snvListVCF}
\title{Generate a vcf for the SNP with at least freqCutoff for all frequency
for at least one super population}
\usage{
snvListVCF(gds, fileOUT, offset = 0, freqCutoff = NULL)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{fileOUT}{TODO}

\item{offset}{TODO. Default: \code{0}.}

\item{freqCutoff}{TODO. Default: \code{NULL}.}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
