% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{snvListVCFRef}
\alias{snvListVCFRef}
\title{Generate a VCF with the information from the SNPs that pass
a cut-off threshold}
\usage{
snvListVCFRef(pRAIDS, fileOut, freqCutoff = NULL)
}
\arguments{
\item{pRAIDS}{a TODO.}

\item{fileOut}{a \code{character} string representing the path and file
name of the VCF file that will be created wit the retained SNP information.
The file should have the ".vcf" extension.}

\item{freqCutoff}{a single positive \code{numeric} specifying the cut-off to
keep a SNP. If \code{NULL}, all SNPs are retained. Default: \code{NULL}.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
