% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{specificSNV<-,RAIDSparam-method}
\alias{specificSNV<-,RAIDSparam-method}
\title{Setter function for replacement of specificSNV slot in
a RAIDSparam class}
\usage{
\S4method{specificSNV}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a \code{data.frame} containing 2 columns. The
first column, called "snp.chromosome" contains the name of the chromosome
wherethe SNV is located. The second column, called "snp.position", contains
the position of the SNV on the chromosome. Optionally, the column "snvKeep"
contains the SNV index in the reference or -1 if not in the reference.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of specificSNV slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the specificSNV slot in the object
specificSNV(paramDemo) <- data.frame(snp.chromosome=c("1", "1"), 
    snp.position=c(13333, 2222213))

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
