% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{specificSNV<-}
\alias{specificSNV<-}
\title{Generic function for replacement of specificSNV slot in a class}
\usage{
specificSNV(x) <- value
}
\arguments{
\item{x}{a S4 object.}

\item{value}{the new value to assign or update.}
}
\value{
the modified S4 object when the new value is valid.
}
\description{
A generic function for replacement of specificSNV
slot in a S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(specificSNV="character"))

# Create an instance
obj <- new("MyClass", specificSNV="123")

# Call the generic (assuming a method is implemented)
# specificSNV(obj) <- "333"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
