% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{specificSNV}
\alias{specificSNV}
\title{Generic function for getting the specificSNV slot in a class}
\usage{
specificSNV(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the specificSNV slot in the S4 object.
}
\description{
A generic function for getting the specificSNV slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(specificSNV="character"))

# Create an instance
obj <- new("MyClass", specificSNV="123")

# Call the generic (assuming a method is implemented)
# specificSNV(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
