% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDF,RAIDSparam-method}
\alias{studyDF,RAIDSparam-method}
\title{Getter for the studyDF slot in a RAIDSparam class}
\usage{
\S4method{studyDF}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{data.frame}.
}
\description{
A function for getting the studyDF slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the studyDF slot for the object
studyDF(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
