% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDF<-,RAIDSparam-method}
\alias{studyDF<-,RAIDSparam-method}
\title{Replacement of studyDF slot in a \code{RAIDSparam} object}
\usage{
\S4method{studyDF}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{data.frame} containing the information
about the study associated with the analyzed sample(s). The
\code{data.frame} must have those 3 columns: "study.id", "study.desc",
"study.platform". All columns must be character strings (no factors).}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of the studyDF slot in a
\code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
