% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDF}
\alias{studyDF}
\title{Generic function for getting the studyDF slot in a class}
\usage{
studyDF(x)
}
\description{
A generic function for getting the studyDF slot in a
class. The function is implemented for the \code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
