% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDFSyn,RAIDSparam-method}
\alias{studyDFSyn,RAIDSparam-method}
\title{A getter for the studyDFSyn slot in a RAIDSparam class}
\usage{
\S4method{studyDFSyn}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{data.frame}.
}
\description{
A function for getting the studyDFSyn slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the studyDFSyn slot for the object
studyDFSyn(paramDemo)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
