% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDFSyn<-,RAIDSparam-method}
\alias{studyDFSyn<-,RAIDSparam-method}
\title{Replacement of studyDFSyn slot in a \code{RAIDSparam} object}
\usage{
\S4method{studyDFSyn}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{data.frame} containing the information
about the synthetic data for the analyzed sample(s). The \code{data.frame}
must have those 3 columns: "study.id", "study.desc", "study.platform". All
columns must be character strings (no factors).}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of the studyDFSyn slot in a
\code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
