% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDFSyn<-}
\alias{studyDFSyn<-}
\title{Generic function for replacement of studyDFSyn slot in a class}
\usage{
studyDFSyn(x) <- value
}
\description{
A generic function for replacement of studyDFSyn slot in a
class. The function is implemented for the \code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
