% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\name{studyDFSyn}
\alias{studyDFSyn}
\title{Generic function for getting the studyDFSyn slot in a class}
\usage{
studyDFSyn(x)
}
\description{
Generic function for getting the studyDFSyn slot in a class
}
