% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyType,RAIDSparam-method}
\alias{studyType,RAIDSparam-method}
\title{A getter for the studyType slot in a RAIDSparam class}
\usage{
\S4method{studyType}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a a single \code{character} string representing the type
of study.
}
\description{
A function for getting the studyType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the studyType slot for the object
studyType(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
