% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyType<-,RAIDSparam-method}
\alias{studyType<-,RAIDSparam-method}
\title{Setter function for replacement of studyType slot in a RAIDSparam class}
\usage{
\S4method{studyType}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{character} string representing the type of
study. The possible choices are: "LD" and "GeneAware". The study type
affects how the allelic fraction is estimated.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of studyType slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign the new study type to the studyType slot in the object
studyType(paramDemo) <- "GeneAware"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
