% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{syntheticRefDF,RAIDSparam-method}
\alias{syntheticRefDF,RAIDSparam-method}
\title{A getter for the syntheticRefDF slot in a RAIDSparam class}
\usage{
\S4method{syntheticRefDF}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or a \code{data.frame} containing a subset of
reference profiles for each sub-population present in the Reference GDS
file.
}
\description{
A function for getting the syntheticRefDF slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the syntheticRefDF slot for the object
syntheticRefDF(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
