% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{syntheticRefDF<-,RAIDSparam-method}
\alias{syntheticRefDF<-,RAIDSparam-method}
\title{Setter function for replacement of syntheticRefDF slot in
a RAIDSparam class}
\usage{
\S4method{syntheticRefDF}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a \code{data.frame} containing a subset of
reference profiles for each sub-population present in the Reference GDS
file. The \code{data.frame} must have those columns:
\describe{
\item{sample.id}{ a \code{character} string representing the sample
identifier. }
\item{pop.group}{ a \code{character} string representing the
subcontinental population assigned to the sample. }
\item{superPop}{ a \code{character} string representing the
super-population assigned to the sample. }
}}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of syntheticRefDF slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new values to the syntheticRefDF slot in the object
syntheticRefDF(paramDemo) <- data.frame(sample.id=c("A", "B"), 
    pop.group=c("CEU", "CEU"), superPop=c("EUR", "EUR"))

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
