% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{syntheticRefDF}
\alias{syntheticRefDF}
\title{Generic function for getting the syntheticRefDF slot in a class}
\usage{
syntheticRefDF(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the syntheticRefDF slot in the S4 object.
}
\description{
A generic function for getting the syntheticRefDF slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(syntheticRefDF="character"))

# Create an instance
obj <- new("MyClass", syntheticRefDF="123")

# Call the generic (assuming a method is implemented)
# syntheticRefDF(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
