% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{tableBlockAF}
\alias{tableBlockAF}
\title{TODO}
\usage{
tableBlockAF(snp.pos)
}
\arguments{
\item{snp.pos}{For a specific chromosome a \code{data.frame} with lap for
the SNV dataset with
coverage > \code{minCov}.}
}
\value{
TODO a \code{data.frame} with the information related to allelic
fraction for each block gene
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
