% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{thresholdLD,RAIDSparam-method}
\alias{thresholdLD,RAIDSparam-method}
\title{A getter for the thresholdLD slot in a RAIDSparam class}
\usage{
\S4method{thresholdLD}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single positive \code{numeric} value that represents
the LD threshold used in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.
}
\description{
A function for getting the thresholdLD slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the thresholdLD slot for the object
thresholdLD(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
