% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{thresholdLD<-,RAIDSparam-method}
\alias{thresholdLD<-,RAIDSparam-method}
\title{Setter function for replacement of thresholdLD slot in
a RAIDSparam class}
\usage{
\S4method{thresholdLD}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single positive \code{numeric} value that represents
the LD threshold used in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of thresholdLD slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the thresholdLD slot in the object
thresholdLD(paramDemo) <- 1000L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
