% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{thresholdLD}
\alias{thresholdLD}
\title{Generic function for getting the thresholdLD slot in a class}
\usage{
thresholdLD(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the thresholdLD slot in the S4 object.
}
\description{
A generic function for getting the thresholdLD slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(thresholdLD="character"))

# Create an instance
obj <- new("MyClass", thresholdLD="123")

# Call the generic (assuming a method is implemented)
# thresholdLD(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
