% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validateAdd1KG2SampleGDS}
\alias{validateAdd1KG2SampleGDS}
\title{Validate input parameters for add1KG2SampleGDS() function}
\usage{
validateAdd1KG2SampleGDS(gds, gdsSampleFile, sampleCurrent, study.id)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{gdsSampleFile}{a \code{character} string representing the path and
file name of the GDS Sample file. The GDS Sample file must exist.}

\item{sampleCurrent}{a \code{character} string corresponding to the sample
identifier associated to the current list of pruned SNVs.}

\item{study.id}{a \code{character} string corresponding to the study
identifier associated to the current list of pruned SNVs.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates the input parameters for the
\code{\link{add1KG2SampleGDS}} function.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## TODO
RAIDS:::validateAdd1KG2SampleGDS(gds=gds1KG,
    gdsSampleFile=file.path(data.dir, "GDS_Sample_with_study_demo.gds"),
    sampleCurrent="Sample01", study.id="Synthetic")


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
