% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validateAppendStudy2GDS1KG}
\alias{validateAppendStudy2GDS1KG}
\title{Validate input parameters for appendStudy2GDS1KG() function}
\usage{
validateAppendStudy2GDS1KG(
  pathGeno,
  filePedRDS,
  fileNameGDS,
  batch,
  studyDF,
  listSamples,
  pathProfileGDS,
  genoSource,
  verbose
)
}
\arguments{
\item{pathGeno}{a \code{character} string representing the path to the
directory containing the output of SNP-pileup, a VCF Sample file, for
each sample.}

\item{filePedRDS}{a \code{character} string representing the path to the
RDS file that contains the information about the sample to analyse.}

\item{fileNameGDS}{a \code{character} string representing the file name of
the 1KG GDS file. The file must exist.}

\item{batch}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.}

\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "studyID", "study.desc", "study.platform". All columns
must be in \code{character} strings.}

\item{listSamples}{a \code{vector} of \code{character} string corresponding
to the sample identifiers that will have a GDS Sample file created. The
sample identifiers must be present in the "Name.ID" column of the RDS file
passed to the \code{filePedRDS} parameter.
If \code{NULL}, all samples in the \code{filePedRDS} are selected.}

\item{pathProfileGDS}{a \code{character} string representing the path to
the directory where the GDS Sample files will be created.}

\item{genoSource}{a \code{character} string with two possible values:
'snp-pileup' or 'generic'. It specifies if the genotype files
are generate by snp-pileup (Facets) or generic format csv.}

\item{verbose}{a \code{logical} indicating if message information should be
printed.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates the input parameters for the
\code{\link{appendStudy2GDS1KG}} function.
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

gds1KG <- file.path(dataDir, "1KG_Demo.gds")
ped <- file.path(dataDir, "unrelatedPatientsInfo_Demo.rds")

## The data.frame containing the information about the study
## The 3 mandatory columns: "studyID", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyInfo <- data.frame(study.id="Pancreatic.WES",
                study.desc="Pancreatic study",
                study.platform="WES",
                stringsAsFactors=FALSE)

## The validatiion should be successful
RAIDS:::validateAppendStudy2GDS1KG(pathGeno=dataDir,
    filePedRDS=ped, fileNameGDS=gds1KG,
    batch=1L, studyDF=studyInfo, listSamples=c("HC01", "HC02"),
    pathProfileGDS=dataDir, genoSource="snp-pileup", verbose=TRUE)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
