% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{validateCharacterString}
\alias{validateCharacterString}
\title{Validate that the input parameter is a character sting}
\usage{
validateCharacterString(value, name)
}
\arguments{
\item{value}{a \code{character} string that will be validated.}

\item{name}{a \code{character} string that represents the name of the
parameter that is tested.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates that the input parameter is
a character string (vector of 1 entry). If the parameter is not a
character string, the function generates
an error with a specific message.
}
\examples{

## The validation should be successful
RAIDS:::validateCharacterString(value="hi", name="test")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
