% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validateComputePoolSyntheticAncestry}
\alias{validateComputePoolSyntheticAncestry}
\title{Validate the parameters of the computePoolSyntheticAncestry ()
function}
\usage{
validateComputePoolSyntheticAncestry(
  gdsReference,
  profileGDS,
  profileAnaID,
  dataRef,
  spRef,
  studyIDSyn,
  np,
  listCatPop,
  fieldPopIn1KG,
  fieldPopInfAnc,
  kList,
  pcaList,
  algorithm,
  eigenCnt,
  missingRate
)
}
\arguments{
\item{gdsReference}{an object of class \link[gdsfmt]{gds.class}
(a GDS file), the opened 1KG GDS file.}

\item{profileGDS}{an object of class \link[gdsfmt]{gds.class} (a GDS file),
an opened Profile GDS file.}

\item{profileAnaID}{TODO}

\item{dataRef}{a \code{data.frame} TODO}

\item{spRef}{TODO}

\item{studyIDSyn}{a \code{character} string corresponding to the study
identifier. The study identifier must be present in the GDS Sample file.}

\item{np}{a single positive \code{integer} representing the number of
threads. Default: \code{1L}.}

\item{listCatPop}{a \code{vector} of \code{character} string
representing the list of possible ancestry assignations. Default:
\code{("EAS", "EUR", "AFR", "AMR", "SAS")}.}

\item{fieldPopIn1KG}{a \code{character} string representing the TODO}

\item{fieldPopInfAnc}{a \code{character} string representing the name of
the column that will contain the inferred ancestry for the specified
dataset. Default: \code{"SuperPop"}.}

\item{kList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _K_ parameter. The _K_ parameter represents the
number of neighbors used in the K-nearest neighbors analysis. If
\code{NULL}, the value \code{seq(2,15,1)} is assigned.
Default: \code{seq(2,15,1)}.}

\item{pcaList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _D_ parameter. The _D_ parameter represents the
number of dimensions used in the PCA analysis.  If \code{NULL},
the value \code{seq(2,15,1)} is assigned.
Default: \code{seq(2,15,1)}.}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.}

\item{missing.rate}{a \code{numeric} value representing the threshold
missing rate at with the SNVs are discarded; the SNVs are retained in the
\link[SNPRelate]{snpgdsPCA}
with "<= missing.rate" only; if \code{NaN}, no missing threshold.}
}
\value{
The integer \code{0L} when successful.
}
\description{
The function validates the input parameters for the
\code{\link{computePoolSyntheticAncestry}} function.
When a parameter is not as expected, an error message is generated.
}
\examples{


## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG_GDS.gds")
fileProfileGDS <- file.path(dataDir, "ex1_demo.gds")

## Open GDS files
gds1KG <- snpgdsOpen(fileGDS)
gdsProfile <- openfn.gds(fileProfileGDS)

dataRef <- data.frame(test=c(1,2), stringAsFactro=FALSE)

## The function returns 0L when all parameters are valid
RAIDS:::validateComputePoolSyntheticAncestry(gdsReference=gds1KG,
    profileGDS=gdsProfile, profileAnaID="SampleID",
    dataRef=dataRef, spRef=NULL,  studyIDSyn="MyStudy",
    np=1L, listCatPop=c("EAS", "EUR", "AFR", "AMR", "SAS"),
    fieldPopIn1KG="SuperPop", fieldPopInfAnc="SuperPop",
    kList=seq(2,15,1), pcaList=seq(2,15,1),
    algorithm="exact", eigenCnt=32L, missingRate=0.025)

## Close GDS files (it is important to always close the GDS files)
closefn.gds(gds1KG)
closefn.gds(gdsProfile)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
