% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validateEstimateAllelicFraction}
\alias{validateEstimateAllelicFraction}
\title{Validate input parameters for estimateAllelicFraction() function}
\usage{
validateEstimateAllelicFraction(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  chrInfo,
  studyType,
  minCov,
  minProb,
  eProb,
  cutOffLOH,
  cutOffHomoScore,
  wAR,
  cutOffAR,
  gdsRefAnnot,
  block.id
)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the 1KG GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the GDS Sample file.}

\item{sampleCurrent}{a \code{character} string corresponding to
the sample identifier as used in \code{\link{pruningSample}} function.}

\item{study.id}{a \code{character} string corresponding to the name of
the study as
used in \code{\link{pruningSample}} function.}

\item{chrInfo}{a \code{vector} of \code{integer} values representing
the length of the chromosomes.}

\item{studyType}{a \code{character} string representing the type of study.
The possible choices are: "DNA" and "RNA". The type of study affects the
way the estimation of the allelic fraction is done. Default: \code{"DNA"}.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage.}

\item{minProb}{a single \code{numeric} between 0 and 1 representing TODO.}

\item{eProb}{a single \code{numeric} between 0 and 1 representing the
probability of sequencing error.}

\item{cutOffLOH}{a single \code{numeric} representing the cutoff, in log,
for the homozygote score to assign a region as LOH.}

\item{cutOffHomoScore}{a single \code{numeric} representing the cutoff, in
log, that the SNVs in a block are called homozygote by error.}

\item{wAR}{a single positive \code{integer} representing the size-1 of
the window used to compute an empty box.}

\item{cutOffAR}{a single \code{numeric} representing the cutoff, in
log score, that the SNVs in a gene are allelic fraction different from 0.5.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the1 1KG SNV Annotation GDS file.
 **This parameter is RNA specific.**}

\item{block.id}{a \code{character} string corresponding to the block
identifier in \code{gdsRefAnnot}. **This parameter is RNA specific.**}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function validates the input parameters for the
\code{\link{estimateAllelicFraction}} function.
}
\examples{

## Required libraries
library(BSgenome.Hsapiens.UCSC.hg38)

## Directory where demo GDS files are located
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## The GDS Sample (opened)
gdsSample <- openfn.gds(file.path(data.dir,
                    "GDS_Sample_with_study_demo.gds"), readonly=TRUE)

## Get chromosome length information
chrInfo <- integer(25L)
for(i in seq_len(22L)){ chrInfo[i] <- length(Hsapiens[[paste0("chr", i)]])}
chrInfo[23] <- length(Hsapiens[["chrX"]])
chrInfo[24] <- length(Hsapiens[["chrY"]])
chrInfo[25] <- length(Hsapiens[["chrM"]])


## The validatiion should be successful
RAIDS:::validateEstimateAllelicFraction(gds=gds1KG, gdsSample=gdsSample,
    sampleCurrent="Sample01", study.id="Synthetic", chrInfo=chrInfo,
    studyType="DNA", minCov=10L, minProb=0.03, eProb=0.002, cutOffLOH=10,
    cutOffHomoScore=11, wAR=2, cutOffAR=10, gdsRefAnnot=gds1KG,
    block.id="1")

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)
closefn.gds(gdsfile=gdsSample)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
