% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{validateGetTableSNV}
\alias{validateGetTableSNV}
\title{Validate input parameters for getTableSNV() function}
\usage{
validateGetTableSNV(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  minCov,
  minProb,
  eProb,
  verbose
)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}} (a GDS file),
the opened 1KG GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened GDS Sample file.}

\item{sampleCurrent}{a \code{character} string corresponding to
the sample identifier used in \code{\link{pruningSample}} function.}

\item{study.id}{a \code{character} string corresponding to the study
identifier used in \code{\link{pruningSample}} function.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage.}

\item{minProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.}

\item{eProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error.}

\item{verbose}{a \code{logicial} indicating if messages should be printed
when the function is running.}
}
\value{
The integer \code{0L} when the function is successful.
}
\description{
This function validates the input parameters for the
\code{\link{getTableSNV}} function.
}
\examples{

## Directory where demo GDS files are located
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## The GDS Sample file
gdsSample <- openfn.gds(file.path(data.dir,
                         "GDS_Sample_with_study_demo.gds"), readonly=TRUE)

## The validation should be successful
RAIDS:::validateGetTableSNV(gds=gds1KG, gdsSample=gdsSample,
     sampleCurrent="A101TCGA", study.id="TCGA", minCov=10L,
     minProb=0.998, eProb=0.002, verbose=TRUE)

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)
closefn.gds(gdsfile=gdsSample)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
