% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{validateLogical}
\alias{validateLogical}
\title{Validate that the input parameter is a logical}
\usage{
validateLogical(logical, name)
}
\arguments{
\item{logical}{a \code{logical} that will be validated.}

\item{name}{a \code{character} string that represents the name of the
parameter that is tested.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates that the input parameter is
a logical. If the parameter is not a logicial, the function generates
an error with a specific message.
}
\examples{

## The validation should be successful
RAIDS:::validateLogical(logical=TRUE, name="test")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
