% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersRAIDS_internal.R
\encoding{UTF-8}
\name{validateParamRAIDS_subpart02}
\alias{validateParamRAIDS_subpart02}
\title{Validate the second subsection of the parametersRAIDS object}
\usage{
validateParamRAIDS_subpart02(parameters)
}
\arguments{
\item{parameters}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{TRUE} when all the parameters tested in the object are valid;
otherwise \code{FALSE}
}
\description{
This function validates a subsection of the parameters
present in the \code{parametersRAIDS} object. The validated parameters are:
"specificSNV", "phase", "PCAalgorithm", "PCAmissingRate" TODO.
}
\examples{

## A data frame with the retained SNVs
specificSNV <-  data.frame("snp.chromosome"=c("chr1", "chr1"),
                            "snp.position"=c(12012, 14222))

## Create an object of class 'parametersRAIDS' with most parameters filled
## default values
parameterAll <- paramRAIDS(genoSource='generic', specificSNV=specificSNV,
        verbose=TRUE)

## Return TRUE when the tested parameters are valid
RAIDS:::validateParamRAIDS_subpart02(parameters=parameterAll)

}
\author{
Pascal Belleau and Astrid Deschênes
}
\keyword{internal}
