% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersRAIDS_internal.R
\encoding{UTF-8}
\name{validateParamRAIDS_subpart03}
\alias{validateParamRAIDS_subpart03}
\title{Validate the third subsection of the parametersRAIDS object}
\usage{
validateParamRAIDS_subpart03(parameters)
}
\arguments{
\item{parameters}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{TRUE} when all the parameters tested in the object are valid;
otherwise \code{FALSE}
}
\description{
This function validates a subsection of the parameters
present in the \code{parametersRAIDS} object. The validated parameters are:
"eigenCount", "kList", "pcaList", "fieldPopInRef", "fieldPopInfAnc",
"fieldSubPop", and "verbose".
}
\examples{

## Create an object of class 'parametersRAIDS' with most parameters filled
## default values
parameterAll <- paramRAIDS(genoSource='generic', verbose=TRUE)

## Return TRUE when the tested parameters are valid
RAIDS:::validateParamRAIDS_subpart03(parameters=parameterAll)

}
\author{
Pascal Belleau and Astrid Deschênes
}
\keyword{internal}
