% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersRAIDS_internal.R
\encoding{UTF-8}
\name{validateParamRAIDS_subpart04}
\alias{validateParamRAIDS_subpart04}
\title{Validate the fourth subsection of the parametersRAIDS object}
\usage{
validateParamRAIDS_subpart04(parameters)
}
\arguments{
\item{parameters}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{TRUE} when all the parameters tested in the object are valid;
otherwise \code{FALSE}
}
\description{
This function validates a subsection of the parameters
present in the \code{parametersRAIDS} object. The validated parameters are:
"offset", "minCov", "minProb", ...
}
\examples{

## Create an object of class 'parametersRAIDS' with most parameters filled
## default values
parameterAll <- paramRAIDS(minCov=20L, minProb=0.20, offset=-1L)

## Return TRUE when the tested parameters are valid
RAIDS:::validateParamRAIDS_subpart04(parameters=parameterAll)

}
\author{
Pascal Belleau and Astrid Deschênes
}
\keyword{internal}
