% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validatePrepPed1KG}
\alias{validatePrepPed1KG}
\title{Validate input parameters for prepPed1KG() function}
\usage{
validatePrepPed1KG(pedFile, PATHGENO, batch.v)
}
\arguments{
\item{pedFile}{a \code{character} string representing the path and
file name of the pedigree file from 1KG. The file must exist.}

\item{PATHGENO}{a \code{character} string representing the path where
the 1K genotyping files for each sample are located. Only the samples with
associated genotyping files are retained in the creation of the final
\code{data.frame}. The name of the genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file.
Default: \code{"./data/sampleGeno"}.}

\item{batch.v}{a\code{integer} that uniquely identifies the source of the
pedigree information. The 1KG is usually \code{0L}. Default: \code{0L}.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates the input parameters for the
\code{\link{prepPed1KG}} function.
}
\examples{

## Directory where demo GDS files are located
data.dir <- system.file("extdata", package="RAIDS")

## Demo pedigree file
pedDemoFile <- file.path(data.dir, "PedigreeDemo.ped")

## The validation should be successful
RAIDS:::validatePrepPed1KG(pedFile=pedDemoFile,
     PATHGENO="data.dir", batch.v=1)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
