% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{validatePruningSample}
\alias{validatePruningSample}
\title{Validate input parameters for pruningSample() function}
\usage{
validatePruningSample(
  gds,
  method,
  sampleCurrent,
  study.id,
  listSNP,
  slide.max.bp.v,
  ld.threshold.v,
  np,
  verbose.v,
  chr,
  minAF.SuperPop,
  keepGDSpruned,
  PATHSAMPLEGDS,
  keepFile,
  PATHPRUNED,
  outPref
)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
1 KG GDS file.}

\item{method}{a \code{character} string that represents the method that will
be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function.}

\item{sampleCurrent}{a \code{character} string
corresponding to the sample identifier used in LD pruning done by the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. A GDS Sample file
corresponding to the sample identifier must exist and be located in the
\code{PATHSAMPLEGDS} directory.}

\item{study.id}{a \code{character} string corresponding to the study
identifier used in the \code{\link[SNPRelate]{snpgdsLDpruning}} function.
The study identifier must be present in the GDS Sample file.}

\item{listSNP}{a \code{vector} of SNVs identifiers specifying selected to
be passed the the pruning function;
if \code{NULL}, all SNVs are used in the
\code{\link[SNPRelate]{snpgdsLDpruning}} function.}

\item{slide.max.bp.v}{a single positive \code{integer} that represents
the maximum basepairs (bp) in the sliding window. This parameter is used
for the LD pruning done in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.}

\item{ld.threshold.v}{a single \code{numeric} value that represents the LD
threshold used in the \code{\link[SNPRelate]{snpgdsLDpruning}} function.}

\item{np}{a single positive \code{integer} specifying the number of
threads to be used.}

\item{verbose.v}{a \code{logicial} indicating if information is shown
during the process in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.}

\item{chr}{a \code{character} string representing the chromosome where the
selected SNVs should belong. Only one chromosome can be handled. If
\code{NULL}, the chromosome is not used as a filtering criterion.}

\item{minAF.SuperPop}{a single positive \code{numeric} representing the
minimum allelic frequency used to select the SNVs. If \code{NULL}, the
allelic frequency is not used as a filtering criterion.}

\item{keepGDSpruned}{a \code{logicial} indicating if the information about
the pruned SNVs should be added to the GDS Sample file.}

\item{PATHSAMPLEGDS}{a \code{character} string representing the directory
where the GDS Sample file will be created. The directory must exist.}

\item{keepFile}{a \code{logical} indicating if RDS files containing the
information about the pruned SNVs must be created.}

\item{PATHPRUNED}{a \code{character} string representing an existing
directory. The directory must exist.}

\item{outPref}{a \code{character} string that represents the prefix of the
RDS files that will be generated. The RDS files are only generated when
the parameter \code{keepFile}=\code{TRUE}.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates the input parameters for the
\code{\link{pruningSample}} function.
}
\examples{

## Directory where demo GDS files are located
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## The validation should be successful
RAIDS:::validatePruningSample(gds=gds1KG, method="corr",
     sampleCurrent="TGCA_01", study.id="TCGA",
     listSNP=c("sr10103", "sr10202"), slide.max.bp.v=1000L,
     ld.threshold.v=0.008, np=1L, verbose.v=TRUE, chr=1,
     minAF.SuperPop=0.002, keepGDSpruned=TRUE, PATHSAMPLEGDS=data.dir,
     keepFile=FALSE, PATHPRUNED=data.dir, outPref="test")

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
