% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{validateSingleRatio}
\alias{validateSingleRatio}
\title{Validate that the input parameter is a single positive numeric
between zero and one (included)}
\usage{
validateSingleRatio(value, name)
}
\arguments{
\item{value}{a single positive \code{numeric} that will be validated.}

\item{name}{a \code{character} string that represents the name of the
parameter that is tested.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function validates that the input parameter is
a single numeric between zero and one (included). If the parameter is not,
the function generates an error with a specific message.
}
\examples{

## The validation should be successful
RAIDS:::validateSingleRatio(value=0.02, name="test")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
