% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_internal.R
\encoding{UTF-8}
\name{validateSyntheticGeno}
\alias{validateSyntheticGeno}
\title{Validate input parameters for syntheticGeno() function}
\usage{
validateSyntheticGeno(
  gds,
  gdsRefAnnot,
  gdsSampleFile,
  data.id.profile,
  listSampleRef,
  nbSim,
  prefId,
  pRecomb,
  minProb,
  seqError
)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the 1KG GDS file.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the1 1KG SNV Annotation GDS file.}

\item{gdsSampleFile}{a \code{character} string representing the file name of
the GDS Sample file containing the information about the sample.
The file must exist.}

\item{data.id.profile}{a \code{character} string representing the unique
identifier of the cancer sample.}

\item{listSampleRef}{a \code{vector} of \code{character} strings
representing the sample identifiers of the 1KG selected reference samples.}

\item{nbSim}{a single positive \code{integer} representing the number of
simulations that will be generated per sample + 1KG reference combination.}

\item{prefId}{a \code{character} string that represent the prefix that will
be added to the name of the synthetic profiles generated by the function.}

\item{pRecomb}{a single positive \code{numeric} between 0 and 1 that
represents the frequency of phase switching in the synthetic profiles.}

\item{minProb}{a single positive \code{numeric} between 0 and 1 that
represents the probability that the genotype is correct.}

\item{seqError}{a single positive \code{numeric} between 0 and 1
representing the sequencing error rate.}
}
\value{
The integer \code{0L} when the function is successful.
}
\description{
This function validates the input parameters for the
\code{\link{syntheticGeno}} function.
}
\examples{

## Directory where demo GDS files are located
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## The 1KG GDS Annotation file (opened)
gds1KGAnnot <- openfn.gds(file.path(data.dir, "gdsAnnot1KG.gds"),
                             readonly=TRUE)

## The GDS Sample file
gdsSample <- file.path(data.dir, "GDS_Sample_with_study_demo.gds")

## The validation should be successful
RAIDS:::validateSyntheticGeno(gds=gds1KG, gdsRefAnnot=gds1KGAnnot,
     gdsSampleFile=gdsSample, data.id.profile="A101TCGA",
     listSampleRef="A101TCGA", nbSim=1L, prefId="TCGA", pRecomb=0.02,
     minProb=0.999, seqError=0.002)

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)
closefn.gds(gdsfile=gds1KGAnnot)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
