% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{verbose,RAIDSparam-method}
\alias{verbose,RAIDSparam-method}
\title{A getter for the verbose slot in a RAIDSparam class}
\usage{
\S4method{verbose}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{logical} indicating if message information should be
printed.
}
\description{
A function for getting the verbose slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the verbose slot for the object
verbose(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
