% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{verbose<-,RAIDSparam-method}
\alias{verbose<-,RAIDSparam-method}
\title{Setter function for replacement of verbose slot in
a RAIDSparam class}
\usage{
\S4method{verbose}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single \code{logical} indicating if message information
should be printed.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of verbose slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign new value to the verbose slot in the object
verbose(paramDemo) <-  TRUE

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
