% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{verbose}
\alias{verbose}
\title{Generic function for getting the verbose slot in a class}
\usage{
verbose(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the verbose slot in the S4 object.
}
\description{
A generic function for getting the verbose slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(verbose="character"))

# Create an instance
obj <- new("MyClass", verbose="123")

# Call the generic (assuming a method is implemented)
# verbose(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
