% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{writeFamProfile}
\alias{writeFamProfile}
\title{Generate a plink fam file}
\usage{
writeFamProfile(
  fileOut,
  listProfile,
  listRM = NULL,
  profileOnly = FALSE,
  pRAIDS
)
}
\arguments{
\item{fileOut}{a \code{character} string representing the path and
the file name of the new fam file.}

\item{listProfile}{a \code{vector} of \code{character} representing the list
of profile from gdsProfile to keep}

\item{listRM}{a TODO. Default: \code{NULL}.}

\item{profileOnly}{a \code{integer} TODO. Default: \code{FALSE}.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
