% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{writePMatrix}
\alias{writePMatrix}
\title{Generate the row for the ADMIXTURE P-matrax}
\usage{
writePMatrix(pathOut, rowF, pRAIDS)
}
\arguments{
\item{pathOut}{a \code{character} representing the output
path}

\item{rowF}{a \code{list} containing allele frequencies for each
population and and for a subset if intersect quasi-no admix subset
intersect the synthetic}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters.}
}
\value{
a \code{array} of the frequency matrix.
}
\description{
This function write with P-matrix
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
