% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{add1KG2SampleGDS}
\alias{add1KG2SampleGDS}
\title{TODO}
\usage{
add1KG2SampleGDS(gds, gdsSampleFile, sampleCurrent, study.id)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{gdsSampleFile}{the path of an object of class \code{gds} related to
the sample}

\item{sampleCurrent}{A \code{string} corresponding to
the sample.id
use in LDpruning}

\item{study.id}{A \code{string} corresponding to the study
use in LDpruning}
}
\value{
\code{0L} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
