% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{add1KG2SampleGDS}
\alias{add1KG2SampleGDS}
\title{Add the information about the pruned SNVs into the GDS Sample file}
\usage{
add1KG2SampleGDS(gds, gdsSampleFile, sampleCurrent, study.id)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{gdsSampleFile}{a \code{character} string representing the path and
file name of the GDS Sample file. The GDS Sample file must exist.}

\item{sampleCurrent}{a \code{character} string corresponding to the sample
identifier associated to the current list of pruned SNVs.}

\item{study.id}{a \code{character} string corresponding to the study
identifier associated to the current list of pruned SNVs.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
The function extracts the information about the pruned SNVs
from the 1KG GDS file and adds entries related to the pruned SNVs in
the GDS Sample file.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
