% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{addBlockFromPlink2GDS}
\alias{addBlockFromPlink2GDS}
\title{TODO contain the information from 1KG}
\usage{
addBlockFromPlink2GDS(
  gds,
  gdsOut,
  PATHBLOCK,
  superPop,
  blockName,
  blockDesc,
  verbose = FALSE
)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), TODO}

\item{gdsOut}{an object of class \code{gds} opened in writing mode.}

\item{PATHBLOCK}{TODO}

\item{superPop}{TODO}

\item{blockName}{TODO}

\item{blockDesc}{TODO}

\item{verbose}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}.}
}
\value{
\code{OL} when the function is successful.
}
\description{
TODO
}
\details{
More information about GDS file format can be found at the Bioconductor
gdsfmt website:
https://bioconductor.org/packages/gdsfmt/
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
