% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{addGDS1KGLDBlock}
\alias{addGDS1KGLDBlock}
\title{TODO}
\usage{
addGDS1KGLDBlock(gds, listBlock, blockName, blockDesc)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the sample}

\item{listBlock}{TODO}

\item{blockName}{TODO}

\item{blockDesc}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
