% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDS1KGLDBlock}
\alias{addGDS1KGLDBlock}
\title{TODO}
\usage{
addGDS1KGLDBlock(gds, listBlock, blockName, blockDesc)
}
\arguments{
\item{gds}{an object of class \code{gds} opened in writing mode.}

\item{listBlock}{TODO}

\item{blockName}{TODO}

\item{blockDesc}{TODO}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
