% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{addGDSRef}
\alias{addGDSRef}
\title{The function add an array sample.ref to the gds file.It define base
on a list of unrelated samples.}
\usage{
addGDSRef(gds, filePart)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{filePart}{a \code{list} from the function pcairPartition in GENESIS}
}
\description{
This function create the field sample.ref which is 1 when the sample
are a reference and 0 otherwise. The sample.ref is fill base of on the file filePart$unrels
from  in GENESIS TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
\keyword{internal}
