% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDSRef}
\alias{addGDSRef}
\title{Create a "sample.ref" node i a GDS file with the information about
the related/unrelated state of the reference samples}
\usage{
addGDSRef(gds, filePart)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened GDS file.}

\item{filePart}{a \code{character} string representing the path and file
name of a RDS file containing the information about the related and
unrelated samples in the reference dataset. The RDS file must exist. The
RDS file must contains a \code{vector} of \code{character} strings called
"unrels" with the name of the unrelated samples.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function creates a "sample.ref" node in the GDS file.
The node contains a vector of integers with value of 1 when
the samples are used as references and 0 otherwise.
The information used to fill the "sample.ref" node comes from the RDS file
that contains the information about the unrelated reference samples.
}
\examples{

## Locate RDS with unrelated/related status for 1KG samples
data.dir <- system.file("extdata", package="RAIDS")
rdsFilePath <- file.path(data.dir, "unrelatedPatientsInfo_Demo.rds")

## Create a temporary GDS file in an test directory
data.dir <- system.file("extdata/tests", package="RAIDS")
gdsFilePath <- file.path(data.dir, "GDS_TEMP_11.gds")

## Create and open the GDS file
GDS_file_tmp  <- createfn.gds(filename=gdsFilePath)
## Create "sample.id" node (the node must be present)
sampleIDs <- c("HG00104", "HG00109", "HG00110")
add.gdsn(node=GDS_file_tmp, name="sample.id", val=sampleIDs)

## Create  "sample.ref" node in GDS file using RDS information
RAIDS:::addGDSRef(gds=GDS_file_tmp, filePart=rdsFilePath)

## Read sample reference data.frame
read.gdsn(index.gdsn(node=GDS_file_tmp, path="sample.ref"))

## Close GDS file
closefn.gds(gdsfile=GDS_file_tmp)

## Delete the temporary GDS file
unlink(x=gdsFilePath, force=TRUE)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
