% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDSStudyPruning}
\alias{addGDSStudyPruning}
\title{Add the pruned.study entry in the GDS Sample file}
\usage{
addGDSStudyPruning(gds, pruned, sample.id)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
GDS Sample file.}

\item{pruned}{TODO}

\item{sample.id}{a \code{character} string corresponding to
the sample identifier.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function adds the pruned.study entry in the GDS
Sample file.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
