% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{addGDSStudyPruning}
\alias{addGDSStudyPruning}
\title{fill the pruned.study in the gds file}
\usage{
addGDSStudyPruning(gds, pruned, sample.id)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the sample}

\item{pruned}{TODO}

\item{sample.id}{A \code{string} corresponding to
the sample.id}
}
\value{
None
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
