% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addGDSStudyPruning}
\alias{addGDSStudyPruning}
\title{Add the pruned.study entry (pruned SNVs) in the GDS Sample file}
\usage{
addGDSStudyPruning(gds, pruned)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
GDS Sample file.}

\item{pruned}{a \code{vector} of pruned SNVs.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function adds the pruned.study entry, which contains
the pruned SNVs, in the GDS
Sample file. If a pruned.study entry is already present, the entry is
deleted and a new entry is created.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
