% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{addGDSStudyPruning}
\alias{addGDSStudyPruning}
\title{fill the pruned.study in the gds file}
\usage{
addGDSStudyPruning(gds, PATHPRUNED, listSamples, prefFile)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{PATHPRUNED}{TODO}

\item{listSamples}{A \code{vector} of \code{string} corresponding to
the sample.ids}

\item{prefFile}{\code{string} with the prefix of the pruned file}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
