% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{addGeneBlockGDSRefAnnot}
\alias{addGeneBlockGDSRefAnnot}
\title{Generate two indexes based on gene annotation for gdsAnnot1KG block in add them
to gdsAnnot1KG}
\usage{
addGeneBlockGDSRefAnnot(
  gds,
  file.gdsRefAnnot,
  winSize = 10000,
  EnsDb,
  suffixe.blockName
)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{file.gdsRefAnnot}{the filename corresponding to an object of
class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the1 1KG SNV Annotation GDS file. The function will
open it in write mode and close it after.}

\item{winSize}{a single positive \code{integer} representing the
size of the window to use to group the SNVs when the SNVs are in a
non-coding region. Default: \code{10000}.}

\item{EnsDb}{An object with the ensembl genome annotation
Default: \code{EnsDb.Hsapiens.v86}.}

\item{suffixe.blockName}{TODO ex Ensembl.Hsapiens.v86}
}
\value{
\code{OL} when the function is successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
