% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{addGeneBlockGDSRefAnnot}
\alias{addGeneBlockGDSRefAnnot}
\title{Generate two indexes based on gene annotation for gdsAnnot1KG
block and add the indexes into the
gdsAnnot1KG}
\usage{
addGeneBlockGDSRefAnnot(
  gdsReference,
  file.gdsRefAnnot,
  winSize = 10000,
  EnsDb,
  suffixe.blockName
)
}
\arguments{
\item{gdsReference}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{file.gdsRefAnnot}{the filename corresponding the 1KG SNV
Annotation GDS file. The function will
open it in write mode and close it after. The file must exist.}

\item{winSize}{a single positive \code{integer} representing the
size of the window to use to group the SNVs when the SNVs are in a
non-coding region. Default: \code{10000L}.}

\item{EnsDb}{An object with the ensembl genome annotation
Default: \code{EnsDb.Hsapiens.v86}.}

\item{suffixe.blockName}{TODO ex Ensembl.Hsapiens.v86}
}
\value{
The integer \code{OL} when the function is successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
