% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addPhase1KG2SampleGDSFromFile}
\alias{addPhase1KG2SampleGDSFromFile}
\title{TODO}
\usage{
addPhase1KG2SampleGDSFromFile(gds, PATHSAMPLEGDS, PATHGENO, fileLSNP)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{PATHSAMPLEGDS}{the path of an object of class \code{gds} related to
the sample}

\item{PATHGENO}{TODO}

\item{fileLSNP}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", 'RAIDS')

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
