% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addPhase1KG2SampleGDSFromFile}
\alias{addPhase1KG2SampleGDSFromFile}
\title{TODO}
\usage{
addPhase1KG2SampleGDSFromFile(
  gds,
  pathProfileGDS,
  pathGeno,
  fileLSNP,
  verbose = FALSE
)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{pathProfileGDS}{a \code{character} string representing the path to
the directory that contains the Profile GDS files. The directory must
exist.}

\item{pathGeno}{a \code{character} string representing the path to
the directory that contains TODO}

\item{fileLSNP}{TODO}

\item{verbose}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}.}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", 'RAIDS')

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
