% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addPhase1KG2SampleGDSFromGDS}
\alias{addPhase1KG2SampleGDSFromGDS}
\title{TODO}
\usage{
addPhase1KG2SampleGDSFromGDS(gds, gdsPhase, PATHSAMPLEGDS)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{gdsPhase}{TODO}

\item{PATHSAMPLEGDS}{the path of an object of class \code{gds} related to
the sample}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
