% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{addRef2GDS1KG}
\alias{addRef2GDS1KG}
\title{This function create the field sample.ref which is 1 when the sample
are a reference and 0 otherwise. The sample.ref is fill base of on the file filePart}
\usage{
addRef2GDS1KG(fileNameGDS, filePart)
}
\arguments{
\item{fileNameGDS}{\code{string} with the path and the name of the gds file}

\item{filePart}{file save by identifyRelative}
}
\value{
None.
}
\description{
This function create the field sample.ref which is 1 when the sample
are a reference and 0 otherwise. The sample.ref is fill base of on the file filePart$unrels
from  in GENESIS TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
