% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addStudy1Kg}
\alias{addStudy1Kg}
\title{TODO}
\usage{
addStudy1Kg(gds, gdsSampleFile)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{gdsSampleFile}{a \code{character} string representing the path and
file name of the GDS Sample file. The GDS Sample file must exist.}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
