% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addStudy1Kg}
\alias{addStudy1Kg}
\title{Append information about the 1KG samples into
the GDS Sample file}
\usage{
addStudy1Kg(gds, fileProfileGDS)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{fileProfileGDS}{a \code{character} string representing the path and
file name of the GDS Sample file. The GDS Sample file must exist.}
}
\value{
The integer \code{0L} when successful.
}
\description{
The information about the samples present in the 1KG GDS file
is added into the GDS Sample file. Only the information about the
unrelated samples
from the 1OOO Genome Study are copied into the GDS Sample file. The
information is only added to the GDS Sample file when the 1KG Study is not
already present in the GDS Sample file. The sample information for all
selected samples is appended to the GDS Sample file "study.annot" node.
The study information is appended to the GDS Sample file "study.list" node.
}
\examples{

# TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
