% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{addStudy1Kg}
\alias{addStudy1Kg}
\title{TODO}
\usage{
addStudy1Kg(gds, gdsSampleFile)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the 1000 Genomes}

\item{gdsSampleFile}{the path of an object of class \code{gds} related to
the sample}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
