% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper2_internal.R
\encoding{UTF-8}
\name{addStudyGDSProfileSyn}
\alias{addStudyGDSProfileSyn}
\title{Add information related to a specific study and specific samples
into a GDS Sample file}
\usage{
addStudyGDSProfileSyn(profileSyn, pRAIDS = pRAIDS)
}
\arguments{
\item{profileSyn}{an object of TODO.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{vector} of \code{character} strings representing the sample
identifiers that have been saved in the GDS Sample file.
}
\description{
This function add entries related to 1) a specific study and
2) specific samples into a GDS Sample file.
The study information is appended to the GDS Sample file "study.list" node
when the node is already present in the file. Otherwise, the node is
created and then, the information is added.
The sample information for all selected samples is appended to the GDS
Sample file "study.annot" node
when the node is already present in the file. Otherwise, the node is
created and then, the information is added.
}
\examples{

## Current directory
dataDir <- file.path(tempdir())

fileGDS <- file.path(system.file("extdata/tests", package="RAIDS"), "ex1_good_small_1KG.gds")

## Copy required file into current directory
file.copy(from=file.path(system.file("extdata/tests", package="RAIDS"),
                    "ex1.txt.gz"), to=dataDir)

file.copy(from=file.path(system.file("extdata/example/snpPileup", package="RAIDS"),
                    "ex1.vcf.gz"), to=dataDir)

## The data.frame containing the information about the study
## The 3 mandatory columns: "study.id", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyDF <- data.frame(study.id = "MYDATA",
                        study.desc = "Description",
                        study.platform = "PLATFORM",
                        stringsAsFactors = FALSE)

## The data.frame containing the information about the samples
## The entries should be strings, not factors (stringsAsFactors=FALSE)
samplePED <- data.frame(Name.ID=c("ex1", "ex2"),
                    Case.ID=c("Patient_h11", "Patient_h12"),
                    Diagnosis=rep("Cancer", 2),
                    Sample.Type=rep("Primary Tumor", 2),
                    Source=rep("Databank B", 2), stringsAsFactors=FALSE)
rownames(samplePED) <- samplePED$Name.ID

pRAIDS <- RAIDS:::paramRAIDS(profileFile=file.path(dataDir, "ex1.txt.gz"),
    profileFileGeno=file.path(dataDir, "ex1.vcf.gz"),
    studyDF=studyDF,
    genoSource="snp-pileup",
    pedStudy=samplePED,
    pathProfileGDS=dataDir,
    fileReferenceGDS=fileGDS
    )

## Parse the coverage to the Profile GDS file


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
