% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addStudyGDSSample}
\alias{addStudyGDSSample}
\title{This function create the gds file fields related to the study and
the sample in it.}
\usage{
addStudyGDSSample(
  gds,
  pedDF,
  batch = 1,
  listSamples = NULL,
  studyDF,
  verbose = TRUE
)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{pedDF}{a \code{data.frame} with the sample info. Must have the column
sample.id, Name.ID, sex, pop.group, superPop and batch. The unique id of
pedDF is Name.ID and the row.name is Name.ID too.}

\item{batch}{a \code{integer} corresponding}

\item{listSamples}{a \code{array} with the sample from pedDF to keep}

\item{studyDF}{a \code{data.frame} with at least the column study.id,
study.desc and study.platform}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function. Default: \code{TRUE}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
