% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addUpdateLap}
\alias{addUpdateLap}
\title{TODO}
\usage{
addUpdateLap(gds, snp.lap)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), a GDS file.}

\item{snp.lap}{TODO}
}
\value{
The integer \code{0L} when successful.
}
\description{
The function adds the information related to TODO
for a specific sample into a GDS file into
the "lap" node. If the "lap" already exists, the previous
information is erased.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
