% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addUpdateSegment}
\alias{addUpdateSegment}
\title{TODO}
\usage{
addUpdateSegment(gds, snp.seg)
}
\arguments{
\item{gds}{an object of class \code{gds} opened for the sample}

\item{snp.seg}{TODO}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
