% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{addUpdateSegment}
\alias{addUpdateSegment}
\title{Add information related to allelic fraction of a pruned SNV dataset
for a specific sample into a GDS file}
\usage{
addUpdateSegment(gds, snp.seg)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), a GDS file.}

\item{snp.seg}{TODO}
}
\value{
The integer \code{0L} when successful.
}
\description{
The function adds the information related to allelic
fraction of a pruned SNV dataset for a specific sample into a GDS file into
the "segment" node. If the "segment" already exists, the previous
information is erased.
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
