% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis.R
\encoding{UTF-8}
\name{ancestryInferencePCAKNN}
\alias{ancestryInferencePCAKNN}
\title{Run most steps leading to the ancestry inference call based on PCA
and KNN on a specific profile}
\usage{
ancestryInferencePCAKNN(pRAIDS = paramRAIDS())
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{list} containing 4 entries:
\describe{
\item{\code{pcaSample}}{ a \code{list} containing the information related
to the eigenvectors. The \code{list} contains those 3 entries:
\describe{
\item{\code{sample.id}}{ a \code{character} string representing the unique
identifier of the current profile.}
\item{\code{eigenvector.ref}}{ a \code{matrix} of \code{numeric} containing
the eigenvectors for the reference profiles.}
\item{\code{eigenvector}}{ a \code{matrix} of \code{numeric} containing the
eigenvectors for the current profile projected on the PCA from the
reference profiles.}
}
}
\item{\code{paraSample}}{ a \code{list} containing the results with
different \code{D} and \code{K} values that lead to optimal parameter
selection. The \code{list} contains those entries:
\describe{
\item{\code{dfPCA}}{ a \code{data.frame} containing statistical results
on all combined synthetic results done with a fixed value of \code{D} (the
number of dimensions). The \code{data.frame} contains those columns:
\describe{
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions).}
\item{\code{median}}{ a \code{numeric} representing the median of the
minimum AUROC obtained (within super populations) for all combination of
the fixed \code{D} value and all tested \code{K} values. }
\item{\code{mad}}{ a \code{numeric} representing the MAD of the minimum
AUROC obtained (within super populations) for all combination of the fixed
\code{D} value and all tested \code{K} values. }
\item{\code{upQuartile}}{ a \code{numeric} representing the upper quartile
of the minimum AUROC obtained (within super populations) for all
combination of the fixed \code{D} value and all tested \code{K} values. }
\item{\code{k}}{ a \code{numeric} representing the optimal \code{K} value
(the number of neighbors) for a fixed \code{D} value. }
}
}
\item{\code{dfPop}}{ a \code{data.frame} containing statistical results on
all combined synthetic results done with different values of \code{D} (the
number of dimensions) and \code{K} (the number of neighbors).
The \code{data.frame} contains those columns:
\describe{
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions).}
\item{\code{K}}{ a \code{numeric} representing the value of \code{K} (the
number of neighbors).}
\item{\code{AUROC.min}}{ a \code{numeric} representing the minimum accuracy
obtained by grouping all the synthetic results by super-populations, for
the specified values of \code{D} and \code{K}.}
\item{\code{AUROC}}{ a \code{numeric} representing the accuracy obtained
by grouping all the synthetic results for the specified values of \code{D}
and \code{K}.}
\item{\code{Accu.CM}}{ a \code{numeric} representing the value of accuracy
of the confusion matrix obtained by grouping all the synthetic results for
the specified values of \code{D} and \code{K}.}
}
}
\item{\code{dfAUROC}}{ a \code{data.frame} the summary of the results by
super-population. The \code{data.frame} contains
those columns:
\describe{
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions).}
\item{\code{K}}{ a \code{numeric} representing the value of \code{K} (the
number of neighbors).}
\item{\code{Call}}{ a \code{character} string representing the
super-population.}
\item{\code{L}}{ a \code{numeric} representing the lower value of the 95\%
confidence interval for the AUROC obtained for the fixed values of
super-population, \code{D} and \code{K}.}
\item{\code{AUROC}}{ a \code{numeric} representing  the AUROC obtained for
the fixed values of super-population, \code{D} and \code{K}.}
\item{\code{H}}{ a \code{numeric} representing the higher value of the 95\%
confidence interval for the AUROC obtained for the fixed values of
super-population, \code{D} and \code{K}.}
}
}
\item{\code{D}}{ a \code{numeric} representing the optimal \code{D} value
(the number of dimensions) for the specific profile.}
\item{\code{K}}{ a \code{numeric} representing the optimal \code{K} value
(the number of neighbors) for the specific profile.}
\item{\code{listD}}{ a \code{numeric} representing the optimal \code{D}
values (the number of dimensions) for the specific profile. More than one
\code{D} is possible.}
}
}
\item{\code{KNNSample}}{  a \code{data.frame} containing the inferred
ancestry for different values of \code{K} and \code{D}. The
\code{data.frame} contains those columns:
\describe{
\item{\code{sample.id}}{ a \code{character} string representing the unique
identifier of the current profile.}
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions) used to infer the ancestry. }
\item{\code{K}}{ a \code{numeric} representing the value of \code{K} (the
number of neighbors) used to infer the ancestry. }
\item{\code{SuperPop}}{ a \code{character} string representing the inferred
ancestry for the specified \code{D} and \code{K} values.}
}
}
\item{\code{KNNSynthetic}}{  a \code{data.frame} containing the inferred
ancestry for each synthetic data for different values of \code{K} and
\code{D}.
The \code{data.frame}
contains those columns:
\describe{
\item{\code{sample.id}}{ a \code{character} string representing the unique
identifier of the current synthetic data.}
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions) used to infer the ancestry. }
\item{\code{K}}{ a \code{numeric} representing the value of \code{K} (the
number of neighbors) used to infer the ancestry. }
\item{\code{infer.superPop}}{ a \code{character} string representing the
inferred ancestry for the specified \code{D} and \code{K} values.}
\item{\code{ref.superPop}}{ a \code{character} string representing the known
ancestry from the reference}
}
}
\item{\code{Ancestry}}{ a \code{data.frame} containing the inferred
ancestry for the current profile. The \code{data.frame} contains those
columns:
\describe{
\item{\code{sample.id}}{ a \code{character} string representing the unique
identifier of the current profile.}
\item{\code{D}}{ a \code{numeric} representing the value of \code{D} (the
number of dimensions) used to infer the ancestry.}
\item{\code{K}}{ a \code{numeric} representing the value of \code{K} (the
number of neighbors) used to infer the ancestry.}
\item{\code{SuperPop}}{ a \code{character} string representing the inferred
ancestry.}
}
}
\item{\code{pRAIDS}}{ a \code{list} of class \code{parametersRAIDS} a list of
the parameters.}
}
}
\description{
This function runs most steps leading to the ancestry inference
call on a specific profile. First, the function creates the
Profile GDS file for the specific profile using the information from a
RDS Sample description file and the Population Reference GDS file.
}
\details{
The runExomeAncestry() function generates 3 types of files
in the OUTPUT directory.
\describe{
\item{Ancestry Inference}{ The ancestry inference CSV file
(".Ancestry.csv" file)}
\item{Inference Informaton}{ The inference information RDS file
(".infoCall.rds" file)}
\item{Synthetic Information}{ The parameter information RDS files
from the synthetic inference ("KNN.synt.*.rds" files in a sub-directory)}
}

In addition, a sub-directory (named using the profile ID) is
also created.
}
\examples{

## Required library for GDS
## library(SNPRelate)

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata", package="RAIDS")


#################################################################
## The 1KG GDS file and the 1KG SNV Annotation GDS file
## need to be located in the same directory
## Note that the 1KG GDS file used for this example is a
## simplified version and CANNOT be used for any real analysis
#################################################################
path1KG <- file.path(dataDir, "tests")

fileReferenceGDS <- file.path(path1KG, "ex1_good_small_1KG.gds")
fileAnnotGDS <- file.path(path1KG, "ex1_good_small_1KG_Annot.gds")

#################################################################
## The Sample SNP pileup files (one per sample) need
## to be located in the same directory.
#################################################################
demoProfileEx1 <- file.path(dataDir, "example", "snpPileup", "ex1.txt.gz")

#################################################################
## The path where the Profile GDS Files (one per sample)
## will be created need to be specified.
#################################################################
pathProfileGDS <- file.path(tempdir(), "out.tmp")

####################################################################
## Fix seed to ensure reproducible results
####################################################################
set.seed(3043)

pRAIDS <- RAIDS:::paramRAIDS(profileFile=demoProfileEx1,
    genoSource="snp-pileup",
    pathProfileGDS=pathProfileGDS,
    fileReferenceGDS=fileReferenceGDS,
    fileReferenceAnnotGDS=fileAnnotGDS
    )

# gds1KG <- snpgdsOpen(fileReferenceGDS)
# dataRef <- select1KGPop(gds1KG, nbProfiles=2L)
# closefn.gds(gds1KG)


res <- ancestryInferencePCAKNN(pRAIDS=pRAIDS)

unlink(file.path(pathProfileGDS, paste0(res$pRAIDS$pedStudy$Name.ID[1], ".gds")) , force=TRUE)

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
