% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{appendGDSgenotype}
\alias{appendGDSgenotype}
\title{This function append the field genotype in the gds file}
\usage{
appendGDSgenotype(gds, listSample, PATHGENO, fileLSNP)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{PATHGENO}{TODO a PATH to a directory with the a file for each
samples with the genotype.}

\item{fileLSNP}{TODO list of SNP to keep in the file genotype}

\item{listSamples}{a \code{array} with the sample to keep}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
