% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{appendGDSgenotype}
\alias{appendGDSgenotype}
\title{This function append the field genotype in the gds file}
\usage{
appendGDSgenotype(gds, listSample, pathGeno, fileLSNP, verbose = FALSE)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{pathGeno}{TODO a PATH to a directory with the a file for each
samples with the genotype.}

\item{fileLSNP}{TODO list of SNP to keep in the file genotype}

\item{verbose}{a \code{logical} indicating if the function must print
messages when running. Default: \code{FALSE}.}

\item{listSamples}{a \code{array} with the sample to keep}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
