% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{appendStudy2GDS1KG}
\alias{appendStudy2GDS1KG}
\title{Create a study at GDS including the reference (first study add)}
\usage{
appendStudy2GDS1KG(
  PATHGENO = file.path("data", "sampleGeno"),
  fileNamePED,
  fileNameGDS,
  batch = 1,
  studyDF,
  listSamples = NULL
)
}
\arguments{
\item{PATHGENO}{TODO a PATH to the directory genotype file of 1KG
The directory sampleGeno must contain matFreqSNV.txt.bz2}

\item{fileNamePED}{TODO}

\item{fileNameGDS}{TODO}

\item{batch}{TODO}

\item{studyDF}{TODO}

\item{listSamples}{A \code{vector} of \code{string} corresponding to
the sample.ids. if NULL all samples}
}
\value{
None
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
