% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{basePCASample}
\alias{basePCASample}
\title{Compute principal component axes (PCA) on SNV data using the
reference samples}
\usage{
basePCASample(gds, listSample.Ref = NULL, listSNP = NULL, np = 1L)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{listSample.Ref}{a \code{vector} of \code{character} strings
corresponding to
the sample identifiers that will be used for the PCA.}

\item{listSNP}{a \code{vector} of \code{character} strings representing
the SNV identifiers retained for the PCA.}

\item{np}{a single positive \code{integer} representing the number of
threads. Default: \code{1L}.}
}
\value{
a \code{list} with 3 entries:
\itemize{
\item{SNP}{ a \code{vector} of \code{character} strings representing the
SNV identifiers used in the PCA.}
\item{pca.unrel}{ an object of class \code{snpgdsPCAClass} as generated
by the
\code{\link[SNPRelate:snpgdsPCA]{SNPRelate::snpgdsPCA}} function. }
\item{snp.load}{ an object of class \code{snpgdsPCASNPLoading} as generated
by the
\code{\link[SNPRelate:snpgdsPCASNPLoading]{SNPRelate::snpgdsPCASNPLoading}}
function. }
}
}
\description{
The function runs a Principal Component Analysis (PCA) on
the SNv genotype data. The function also loads SNVs into the PCA to
calculate the SNV eigenvectors. Those 2 steps are done with the
 \code{\link[SNPRelate]{snpgdsPCA}} and
 \code{\link[SNPRelate]{snpgdsPCASNPLoading}}
functions.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
