% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{calcAFMLRNA}
\alias{calcAFMLRNA}
\title{TOREVIEW Compute the log likelihood ratio base on the coverage (read depth)
of each allele in block (gene in the case of RNA-seq)}
\usage{
calcAFMLRNA(snpPosHetero)
}
\arguments{
\item{snpPosHetero}{For a specific gene (block) a \code{data.frame}
containing the SNV information.
The \code{data.frame} must contain those columns:
\itemize{
\item{cnt.ref} {a single \code{integer} representing the coverage for
the reference allele.}
\item{cnt.alt} {a single \code{integer} representing the coverage for
the alternative allele.}
\item{phase} {TOREVIEW a \code{integer} indicating the phase of the variant
if known, 3 if not known}
}}
}
\value{
TOREVIEW a \code{list}  for the block with the information of
relative to the heterozygotes.
The \code{list} contains:
\itemize{
\item{lR} {TOREVIEW a single \code{numeric} representing sum the log of read depth of the lowest depth
divide by the total depth of the position minus of likelhood of the allelic
fraction of 0.5.}
\item{aFraction} {TOREVIEW a single \code{numeric} representing the allele
fraction estimation.}
\item{sumAlleleLow} {TOREVIEW a \code{integer} representing the
sum of the allele read depth
of the lowest read alelle depth}
\item{sumAlleleHigh} {TOREVIEW a \code{integer} representing the
sum of the allele read depth
of the highsest read alelle depth}
}
}
\description{
TOREVIEW For the block sum the log of read depth of the lowest depth
divide by the total depth of the position minus of likelhood of the allelic
fraction of 0.5. If the phase is known, the variant varaint in the same
haplotype are group.
}
\examples{

dataDir <- system.file("extdata", package="RAIDS")

snpPos <- readRDS(file.path(dataDir, "demoAllelicFraction", "demSnpPos.rds"))

result <- RAIDS:::calcAFMLRNA(snpPos[which(
                snpPos$block.id == 2750 &
                snpPos$hetero), c("cnt.ref",
                    "cnt.alt", "phase")])
head(result)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
