% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeAlleleFraction}
\alias{computeAlleleFraction}
\title{TODO}
\usage{
computeAlleleFraction(snp.pos, chr, w = 10, cutOff = -3)
}
\arguments{
\item{snp.pos}{TODO}

\item{chr}{a single positive \code{integer} for the chromosome.}

\item{w}{a single positive \code{numeric} representing the size of the
window to compute the allelic fraction.
Default: \code{10}.}

\item{cutOff}{a single \code{numeric} representing TODO. Default: \code{-3}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
