% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeAllelicFractionDNA}
\alias{computeAllelicFractionDNA}
\title{TODO}
\usage{
computeAllelicFractionDNA(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffHomoScore = -3,
  wAR = 9,
  verbose = FALSE
)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{gdsSample}{TODO}

\item{sampleCurrent}{A \code{character} string corresponding to
the sample identifier as used in \code{\link{pruningSample}} function.}

\item{study.id}{A \code{character} string corresponding to the name of
the study as
used in \code{\link{pruningSample}} function.}

\item{chrInfo}{a vector chrInfo[i] = length(Hsapiens[[paste0("chr", i)]])
Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{minProb}{a single \code{numeric} between 0 and 1 representing TODO.
Default: \code{0.999}.}

\item{eProb}{a single \code{numeric} between 0 and 1 representing the
probability of sequencing error. Default: \code{0.001}.}

\item{cutOffLOH}{log of the score to be LOH . Default: \code{-5}.}

\item{cutOffHomoScore}{TODO. Default: \code{-3}.}

\item{wAR}{a single positive \code{integer} representing the size-1 of
the window used to compute an empty box. Default: \code{9}.}

\item{verbose}{a \code{logicial} indicating if the function should print
message when running. Default: \code{FALSE}.}
}
\value{
a \code{data.frame} with lap for the snv with depth > minCov. TODO
}
\description{
Create a data.frame with the allelic fraction TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
