% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{computeAllelicFractionRNA}
\alias{computeAllelicFractionRNA}
\title{Estimate the allelic fraction of the pruned SNVs for a specific
RNA-seq sample}
\usage{
computeAllelicFractionRNA(
  gdsReference,
  gdsSample,
  gdsRefAnnot,
  currentProfile,
  studyID,
  blockID,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffAR = 3,
  verbose
)
}
\arguments{
\item{gdsReference}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened 1KG GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened Profile GDS file.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the opened Reference SNV Annotation GDS file.}

\item{currentProfile}{a \code{character} string corresponding to
the sample identifier as used in \code{\link{pruningSample}} function.}

\item{studyID}{a \code{character} string corresponding to the name of
the study as
used in \code{\link{pruningSample}} function.}

\item{blockID}{a \code{character} string corresponding to the field gene
block in the GDS \code{gdsRefAnnot} to use split by gene.}

\item{chrInfo}{a \code{vector} of \code{integer} values representing
the length of the chromosomes.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{minProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
Default: \code{0.999}.}

\item{eProb}{a single \code{numeric} between 0 and 1 representing the
probability of sequencing error. Default: \code{0.001}.}

\item{cutOffLOH}{a single log of the score to be LOH TODO.
Default: \code{-5}.}

\item{cutOffAR}{a single \code{numeric} representing the cutoff, in
log score, to tag SNVs located in a gene has having an allelic fraction
different 0.5
Default: \code{3}.}

\item{verbose}{a \code{logicial} indicating if the function should print
message when running.}
}
\value{
a \code{data.frame} with lap for the pruned SNV dataset with
coverage > \code{minCov}. TODO
}
\description{
The function creates a \code{data.frame} containing the
allelic fraction for the pruned SNV dataset specific to a RNA-seq sample.
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
