% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{computeAllelicImbDNAChr}
\alias{computeAllelicImbDNAChr}
\title{TODO}
\usage{
computeAllelicImbDNAChr(snp.pos, chr, wAR = 10, cutOffEmptyBox = -3)
}
\arguments{
\item{snp.pos}{a \code{data.frame} containing the genotype information for
a SNV dataset. TODO}

\item{chr}{a single positive \code{integer} for the chromosome.}

\item{wAR}{a single positive \code{integer} representing the size-1 of
the window used to compute an empty box. Default: \code{10}.}

\item{cutOffEmptyBox}{TODO Default: \code{-3}.}
}
\value{
a \code{vector} of TODO representing the imbAR TODO.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
