% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeAncestryFromSyntheticFile}
\alias{computeAncestryFromSyntheticFile}
\title{TODO}
\usage{
computeAncestryFromSyntheticFile(
  gds,
  gdsSample,
  listFiles,
  sample.ana.id,
  spRef,
  study.id.syn,
  np = 1L,
  listCatPop = c("EAS", "EUR", "AFR", "AMR", "SAS"),
  fieldPopIn1KG = "superPop",
  fieldPopInfAnc = "SuperPop",
  kList = seq(2, 15, 1),
  pcaList = 2:15,
  algorithm = "exact",
  eigen.cnt = 32L,
  missing.rate = 0.025
)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
1KG GDS file.}

\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{listFiles}{TODO.}

\item{sample.ana.id}{TODO}

\item{spRef}{TODO}

\item{study.id.syn}{TODO}

\item{np}{a single \code{integer} representing the number of CPU to use.
Default: \code{1L}.}

\item{listCatPop}{TODO,
Default: \code{c("EAS", "EUR", "AFR", "AMR", "SAS")}.}

\item{fieldPopIn1KG}{TODO}

\item{fieldPopInfAnc}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}

\item{algorithm}{algorithm of the PCA "exact", "randomized"
(para snpgdsPCA)}

\item{eigen.cnt}{number of eigenvectors in PCA
(para snpgdsPCA)}

\item{missing.rate}{number of eigenvectors in PCA
(para snpgdsPCA)}
}
\value{
a \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
