% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeAncestryFromSyntheticFile}
\alias{computeAncestryFromSyntheticFile}
\title{TODO}
\usage{
computeAncestryFromSyntheticFile(
  gds,
  gdsSample,
  listFiles,
  sample.ana.id,
  spRef,
  study.id.syn,
  np = 1L,
  listCatPop = c("EAS", "EUR", "AFR", "AMR", "SAS"),
  fieldPopIn1KG = "superPop",
  fieldPopInfAnc = "SuperPop",
  kList = seq(2, 15, 1),
  pcaList = 2:15,
  algorithm = "exact",
  eigen.cnt = 32L,
  missing.rate = NaN
)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
1KG GDS file.}

\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{listFiles}{TODO.}

\item{sample.ana.id}{TODO}

\item{spRef}{TODO}

\item{study.id.syn}{TODO}

\item{np}{a single \code{integer} representing the number of CPU to use.
Default: \code{1L}.}

\item{listCatPop}{TODO,
Default: \code{c("EAS", "EUR", "AFR", "AMR", "SAS")}.}

\item{fieldPopIn1KG}{TODO}

\item{fieldPopInfAnc}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}

\item{missing.rate}{a \code{numeric} value representing the threshold
missing rate at with the SNVs are discarded; the SNVs are retained in the
\link[SNPRelate]{snpgdsPCA}
with "<= missing.rate" only; if \code{NaN}, no missing threshold.
Default: \code{NaN}.}
}
\value{
a \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
