% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis_internal.R
\encoding{UTF-8}
\name{computeKNNProfileSubSet}
\alias{computeKNNProfileSubSet}
\title{Run a k-nearest neighbors analysis on a subset of the
synthetic dataset}
\usage{
computeKNNProfileSubSet(listEigenvector, K, D, dfRef, pRAIDS)
}
\arguments{
\item{listEigenvector}{a \code{list} with 3 entries:
'sample.id', 'eigenvector.ref' and 'eigenvector'. The \code{list} represents
the PCA done on the 1KG reference profiles and the synthetic profiles
projected onto it.}

\item{K}{a \code{integer} representing value tested for the  K parameter.
The K parameter represents the
number of neighbors used in the K-nearest neighbors analysis.}

\item{D}{a \code{integer} representing
value tested for the  D parameter. The D parameter represents the
number of dimensions used in the PCA analysis.}

\item{dfRef}{\code{data.frame} TODO of \code{character}
strings representing the
known super population ancestry for the 1KG profiles. The 1KG profile
identifiers are used as names for the \code{vector}.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{list} containing 4 entries:
\describe{
\item{\code{sample.id}}{ a \code{vector} of \code{character} strings
representing the identifiers of the synthetic profiles analysed.}
\item{\code{sample1Kg}}{ a \code{vector} of \code{character} strings
representing the identifiers of the 1KG reference profiles used to
generate the synthetic profiles.}
\item{\code{sp}}{ a \code{vector} of \code{character} strings representing
the known super population ancestry of the 1KG reference profiles used
to generate the synthetic profiles.}
\item{\code{matKNN}}{ a \code{data.frame} containing the super population
inference for each synthetic profiles for different values of PCA
dimensions \code{D} and k-neighbors values \code{K}. The fourth column title
corresponds to the \code{fieldPopInfAnc} parameter.
The \code{data.frame} contains 4 columns:
\describe{
\item{\code{sample.id}}{ a \code{character} string representing
the identifier of the synthetic profile analysed.}
\item{\code{D}}{ a \code{numeric} strings representing
the value of the PCA dimension used to infer the super population.}
\item{\code{K}}{ a \code{numeric} strings representing
the value of the k-neighbors used to infer the super population.}
\item{\code{fieldPopInfAnc} value}{ a \code{character} string representing
the inferred ancestry.}
}
}
}
}
\description{
The function runs k-nearest neighbors analysis on a
subset of the synthetic data set. The function uses the 'knn' package.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
