% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNRefSample}
\alias{computeKNNRefSample}
\title{TODO}
\usage{
computeKNNRefSample(
  listEigenvector,
  listCatPop,
  spRef,
  fieldPopInfAnc = "SuperPop",
  kList = seq(2, 15, 1),
  pcaList = seq(2, 15, 1)
)
}
\arguments{
\item{listEigenvector}{TODO see return of computePCARefSample}

\item{listCatPop}{TODO}

\item{spRef}{TODO}

\item{fieldPopInfAnc}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
