% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNRefSynthetic}
\alias{computeKNNRefSynthetic}
\title{TODO}
\usage{
computeKNNRefSynthetic(
  gdsSample,
  listEigenvector,
  listCatPop,
  study.id.syn,
  spRef,
  fieldPopInfAnc = "SuperPop",
  kList = seq_len(15),
  pcaList = 2:15
)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{listEigenvector}{TODO see return of computePCAsynthetic}

\item{listCatPop}{TODO}

\item{study.id.syn}{a the study.id of the synthetic data}

\item{spRef}{TODO}

\item{fieldPopInfAnc}{TODO}

\item{kList}{TODO array of the k possible values}

\item{pcaList}{TODO array of the pca dimension possible values}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
