% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computeKNNSuperPoprSynthetic}
\alias{computeKNNSuperPoprSynthetic}
\title{TODO}
\usage{
computeKNNSuperPoprSynthetic(listEigenvector, sample.ref, study.annot, spRef)
}
\arguments{
\item{listEigenvector}{TODO see return of computePCAsynthetic}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}

\item{spRef}{TODO}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
