% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{computeKNNSuperPopSample}
\alias{computeKNNSuperPopSample}
\title{Deprecated}
\usage{
computeKNNSuperPopSample(
  gdsSample,
  listEigenvector,
  name.id,
  spRef,
  studyIDRef = "Ref.1KG",
  kList = seq_len(15),
  pcaList = 2:15
)
}
\arguments{
\item{listEigenvector}{TODO see return of computePCARefSample}

\item{spRef}{TODO}

\item{kList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _K_ parameter. The _K_ parameter represents the
number of neighbors used in the K-nearest neighbor analysis. If \code{NULL},
the value \code{seq_len(15)} is assigned.
Default: \code{seq_len(15)}.}

\item{pcaList}{TODO array of the pca dimension possible values}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors
and a table with KNN callfor different K and pca dimension.
}
\description{
Deprecated
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
