% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{computeLOHBlocksDNAChr}
\alias{computeLOHBlocksDNAChr}
\title{TODO}
\usage{
computeLOHBlocksDNAChr(gds, chrInfo, snp.pos, chr, genoN = 1e-04)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{chrInfo}{a vector chrInfo[i] = length(Hsapiens[[paste0("chr", i)]])
Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)}

\item{snp.pos}{a \code{data.frame} containing TODO.}

\item{chr}{a single positive \code{integer} for the chromosome.}

\item{genoN}{a single \code{numeric} between 0 and 1 representing TODO.
Default: \code{0.0001}.}
}
\value{
a \code{data.frame} containing:
\itemize{
\item{chr} {TODO}
\item{start} {TODO}
\item{end} {TODO}
\item{logLHR} {TODO}
\item{LH1} {TODO}
\item{LM1} {TODO}
\item{homoScore} {TODO}
\item{nbSNV} {TODO}
\item{nbPruned} {TODO}
\item{nbNorm} {TODO}
\item{LOH} {TODO}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
