% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCAMultiSynthetic}
\alias{computePCAMultiSynthetic}
\title{TODO}
\usage{
computePCAMultiSynthetic(gdsSample, listPCA, sampleRef, study.id.syn)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{listPCA}{TODO}

\item{sampleRef}{a \code{vector} of sample.id from 1KG with one entry for each synthetic to project.}

\item{study.id.syn}{a the study.id of the synthetic data}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO
gds <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
