% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCAMultiSynthetic}
\alias{computePCAMultiSynthetic}
\title{TODO}
\usage{
computePCAMultiSynthetic(
  gdsProfile,
  listPCA,
  sampleRef,
  studyIDSyn,
  verbose = FALSE
)
}
\arguments{
\item{gdsProfile}{an object of class \code{gds} opened related to
the sample}

\item{listPCA}{TODO}

\item{sampleRef}{a \code{vector} of sample.id from 1KG with one entry for
each synthetic to project.}

\item{studyIDSyn}{a \code{character} string corresponding to the study
identifier.
The study identifier must be present in the GDS Sample file.}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function. Default: \code{FALSE}.}
}
\value{
a \code{list} containing 3 entries:
\itemize{
\item{sample.id} { TODO }
\item{eigenvector.ref} { TODO }
\item{eigenvecto} { TODO }
}
}
\description{
TODO
}
\examples{

# TODO
gds <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
