% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis_internal.R
\encoding{UTF-8}
\name{computePCAProfile}
\alias{computePCAProfile}
\title{Calculate Principal Component Analysis (PCA) on SNV genotype data set}
\usage{
computePCAProfile(listPCA, profileId, pRAIDS)
}
\arguments{
\item{listPCA}{a TODO.}

\item{profileId}{a TODO.}

\item{pRAIDS}{a TODO.}
}
\value{
a \code{list} containing 2 entries:
\describe{
\item{pruned}{ a \code{vector} of SNV identifiers specifying selected SNVs
for the PCA analysis.}
\item{pca.unrel}{ a \code{snpgdsPCAClass} object containing the eigenvalues
as generated by \link[SNPRelate]{snpgdsPCA} function.}
}
}
\description{
The functions calculates the principal component analysis (PCA)
for a list of pruned SNVs present in a Profile GDS file. The
\link[SNPRelate]{snpgdsPCA} function is used to do the calculation.
}
\examples{

## TODO

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
