% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCARefSample}
\alias{computePCARefSample}
\title{Compute Principal Component Analysis (PCA) on SNV genotype for
data from a sample}
\usage{
computePCARefSample(
  gdsSample,
  name.id,
  study.id.ref = "Ref.1KG",
  np = 1L,
  algorithm = c("exact", "randomized"),
  eigen.cnt = 32L,
  missing.rate = NaN
)
}
\arguments{
\item{gdsSample}{an object of class \link[gdsfmt]{gds.class},
a GDS Sample file.}

\item{name.id}{a \code{character} string representing the sample
identifier.}

\item{study.id.ref}{id of the reference in study.annot}

\item{np}{a single positive \code{integer} representing the number of CPU
that will be used. Default: \code{1L}.}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}

\item{missing.rate}{a \code{numeric} value representing the threshold
missing rate at with the SNVs are discarded; the SNVs are retained in the
\link[SNPRelate]{snpgdsPCA}
with "<= missing.rate" only; if \code{NaN}, no missing threshold.
Default: \code{NaN}.}
}
\value{
a \code{list} containing 3 entries:
\itemize{
\item{sample.id} { TODO }
\item{eigenvector.ref} { TODO }
\item{eigenvector} { TODO }
}
}
\description{
TODO
}
\examples{

# TODO
gds <- "TOTO"

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
