% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCARefSample}
\alias{computePCARefSample}
\title{TODO}
\usage{
computePCARefSample(
  gdsSample,
  name.id,
  study.id.ref = "Ref.1KG",
  algorithm = "exact",
  eigen.cnt = 32L
)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{name.id}{the sample id
the gds}

\item{study.id.ref}{id of the reference in study.annot}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO
gds <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
