% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePCAsynthetic}
\alias{computePCAsynthetic}
\title{TODO}
\usage{
computePCAsynthetic(gdsSample, pruned, sample.id, sample.ref, study.annot)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{pruned}{TODO}

\item{sample.id}{TODO}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
