% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy_internal.R
\encoding{UTF-8}
\name{computePCAsynthetic}
\alias{computePCAsynthetic}
\title{Deprecated Function}
\usage{
computePCAsynthetic(
  gdsSample,
  pruned,
  sample.id,
  sample.ref,
  study.annot,
  algorithm = "exact",
  eigen.cnt = 32L
)
}
\arguments{
\item{gdsSample}{an object of class \code{gds} opened related to
the sample}

\item{pruned}{TODO}

\item{sample.id}{TODO}

\item{sample.ref}{TODO}

\item{study.annot}{a  \code{data.frame} with one entry from study.annot in
the gds}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}
}
\value{
A \code{list} TODO with the sample.id and eigenvectors.
}
\description{
TODO Deprecated
}
\examples{

# TODO
gds <- "TOTO"

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
