% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{computePoolSyntheticAncestryGr}
\alias{computePoolSyntheticAncestryGr}
\title{Run a PCA analysis and a K nearest neighbor analysis on a small set
of synthetic data}
\usage{
computePoolSyntheticAncestryGr(
  gds,
  gdsSample,
  sampleRM,
  spRef,
  study.id.syn,
  np = 1L,
  listCatPop = c("EAS", "EUR", "AFR", "AMR", "SAS"),
  fieldPopIn1KG = "superPop",
  fieldPopInfAnc = "SuperPop",
  kList = seq(2, 15, 1),
  pcaList = seq(2, 15, 1),
  algorithm = c("exact", "randomized"),
  eigen.cnt = 32L,
  missing.rate = 0.025
)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
1KG GDS file.}

\item{gdsSample}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, the
GDS Sample file.}

\item{sampleRM}{a \code{vector} of \code{character} strings representing
the identifiers of the 1KG reference samples that should not be used to
create the reference PCA.}

\item{spRef}{TODO}

\item{study.id.syn}{a \code{character} string corresponding to the study
identifier.
The study identifier must be present in the GDS Sample file.}

\item{np}{a single positive \code{integer} representing the number of
threads. Default: \code{1L}.}

\item{listCatPop}{a \code{vector} of \code{character} string
representing the list of possible ancestry assignations. Default:
\code{("EAS", "EUR", "AFR", "AMR", "SAS")}.}

\item{fieldPopIn1KG}{TODO. Default: \code{"superPop"}.}

\item{fieldPopInfAnc}{a \code{character} string representing the name of
the column that will contain the inferred ancestry for the specified
dataset. Default: \code{"SuperPop"}.}

\item{kList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _K_ parameter. The _K_ parameter represents the
number of neighbors used in the K-nearest neighbor analysis. If \code{NULL},
the value \code{seq(2,15,1)} is assigned.
Default: \code{seq(2,15,1)}.}

\item{pcaList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _D_ parameter. The _D_ parameter represents the
number of dimensions used in the PCA analysis.  If \code{NULL},
the value \code{seq(2,15,1)} is assigned.
Default: \code{seq(2,15,1)}.}

\item{algorithm}{a \code{character} string representing the algorithm used
to calculate the PCA. The 2 choices are "exact" (traditional exact
calculation) and "randomized" (fast PCA with randomized algorithm
introduced in Galinsky et al. 2016). Default: \code{"exact"}.}

\item{eigen.cnt}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}

\item{missing.rate}{a \code{numeric} value representing the threshold
missing rate at with the SNVs are discarded; the SNVs are retained in the
\link[SNPRelate]{snpgdsPCA} function
with "<= missing.rate" only; if \code{NaN}, no missing threshold.
Default: \code{0.025}.}
}
\value{
a \code{list} TODO with the sample.id and eigenvectors
and a table with KNN call for different K and pca dimension.
}
\description{
The function runs a PCA analysis using 1 synthetic profile
from each sub-continental population. The reference samples used to
create those synthetic profiles are first removed from the 1KG list
of samples that generates the reference PCA. Then, the retained synthetic
profiles are projected on the 1KG PCA space. Finally, a K nearest neighbor
analysis using a range of K and D values is done.
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
