% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_internal.R
\encoding{UTF-8}
\name{computeSyntheticConfMat}
\alias{computeSyntheticConfMat}
\title{TODO}
\usage{
computeSyntheticConfMat(matKNN, pedCall, refCall, predCall, listCall)
}
\arguments{
\item{matKNN}{TODO}

\item{pedCall}{TODO see return of prepPedSynthetic1KG}

\item{refCall}{TODO column name in pedCall with the call}

\item{predCall}{a \code{character} string representing the name of
the column that will contain the inferred ancestry for the specified
dataset.}

\item{listCall}{TODO array of the possible call}
}
\value{
\code{list} containing 2 entries:
\itemize{
\item{confMat} { TODO }
\item{matAccuracy} { a \code{data.frame} TODO}
}
}
\description{
TODO
}
\examples{

## TODO
gds <- "TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
\keyword{internal}
