% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis_internal.R
\encoding{UTF-8}
\name{createProfile2}
\alias{createProfile2}
\title{Create the Profile GDS file(s) for one or multiple specific profiles
using the information from a RDS Sample description file and the 1KG
GDS file}
\usage{
createProfile2(pRAIDS)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
The function returns a \code{parametersRAIDS} an object with all the RAIDS
parameters updated.
}
\description{
The function uses the information for the Reference GDS file
and the RDS Sample Description file to create the Profile GDS file. One
Profile GDS file is created per profile. One Profile GDS file will be
created for each entry present in the \code{listProfiles} parameter.
}
\examples{

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG.gds")

## The data.frame containing the information about the study
## The 3 mandatory columns: "study.id", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyDF <- data.frame(study.id = "MYDATA",
                        study.desc = "Description",
                        study.platform = "PLATFORM",
                        stringsAsFactors = FALSE)

## The data.frame containing the information about the samples
## The entries should be strings, not factors (stringsAsFactors=FALSE)
samplePED <- data.frame(Name.ID=c("ex1"),
                    Case.ID=c("Patient_h11"),
                    Diagnosis=rep("Cancer"),
                    Sample.Type=c("Primary Tumor"),
                    Source=c("Databank B"), stringsAsFactors=FALSE,
                    drop=FALSE)
rownames(samplePED) <- samplePED$Name.ID

pRAIDS <- RAIDS:::paramRAIDS(profileFile=file.path(dataDir, "ex1.txt.gz"),
    studyDF=studyDF,
    genoSource="snp-pileup",
    pedStudy=samplePED,
    pathProfileGDS=tempdir(),
    fileReferenceGDS=fileGDS
    )
## Create the Profile GDS File for samples in 'listSamples' vector
## (in this case, samples "ex1")
## The Profile GDS file is created in the pathProfileGDS directory
pRAIDS <- RAIDS:::createProfile2(pRAIDS)


## The Profile GDS file 'ex1.gds' has been created in the
## specified directory
list.files(pRAIDS$pathProfileGDS)

## Remove Profile GDS file (created for demo purpose)
unlink(file.path(pRAIDS$pathProfileGDS, "ex1.gds"), force=TRUE)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
