% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{dispatchParseReadCov}
\alias{dispatchParseReadCov}
\title{Read a VCF file with the genotypes use for the ancestry call}
\usage{
dispatchParseReadCov(pRAIDS)
}
\arguments{
\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
a \code{list}
\describe{
\item{matProfile}{
a \code{data.frame} containing at least:
\describe{
\item{Chromosome}{ a \code{numeric} representing the name of
the chromosome}
\item{Position}{ a \code{numeric} representing the position on the
chromosome}
\item{Ref}{ a \code{character} string representing the reference nucleotide}
\item{Alt}{ a \code{character} string representing the alternative
nucleotide}
\item{File1R}{ a \code{numeric} representing the count for
the reference nucleotide}
\item{File1A}{ a \code{numeric} representing the count for the
alternative nucleotide}
\item{count}{ a \code{numeric} representing the total count}
}}
\item{pRAIDS}{
a \code{parametersRAIDS} an object with all the RAIDS
parameters
}
}
}
\description{
The function reads VCF file and returns a data frame
containing the information about the read counts for the SNVs present in
the file.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
