% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{estimateAllelicFraction}
\alias{estimateAllelicFraction}
\title{TODO}
\usage{
estimateAllelicFraction(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  chrInfo,
  studyType = "DNA",
  minCov = 10,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffHomoScore = -3,
  wAR = 9
)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{gdsSample}{TODO}

\item{sampleCurrent}{A \code{string} corresponding to
the sample.id
use in LDpruning}

\item{study.id}{A \code{string} corresponding to the study
use in LDpruning}

\item{chrInfo}{a vector chrInfo[i] = length(Hsapiens[[paste0("chr", i)]])
Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)}

\item{studyType}{a \code{string} with value as DNA, ...}

\item{minCov}{an \code{integer} default 10}

\item{minProb}{an \code{numeric} betweeen 0 and 1}

\item{eProb}{an \code{numeric} betweeen 0 and 1 probability of sequencing error}

\item{cutOffLOH}{log of the score to be LOH default -5}

\item{cutOffHomoScore}{TODO}

\item{wAR}{size-1 of the window to compute an empty box}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
