% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{fileReferenceAnnotGDS}
\alias{fileReferenceAnnotGDS}
\title{Generic function for getting the fileReferenceAnnotGDS slot in a class}
\usage{
fileReferenceAnnotGDS(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the fileReferenceAnnotGDS slot in the S4 object.
}
\description{
A generic function for getting the fileReferenceAnnotGDS slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(fileReferenceAnnotGDS="character"))

# Create an instance
obj <- new("MyClass", fileReferenceAnnotGDS="123")

# Call the generic (assuming a method is implemented)
# fileReferenceAnnotGDS(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
