% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{gds2tfamSample}
\alias{gds2tfamSample}
\title{create a file tfam file for plink from the gds file}
\usage{
gds2tfamSample(gds, listSample, sampleANNO, pedOUT)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{sampleANNO}{a \code{data.frame} with at least column sex and the name
must be sample.id}

\item{pedOUT}{TODO a PATH and file name to the output file}

\item{listSamples}{a \code{array} with the sample to keep}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
