% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{gds2tfamSample}
\alias{gds2tfamSample}
\title{create a file tfam file for plink from the gdsProfile file}
\usage{
gds2tfamSample(gdsProfile, listSample, sampleANNO, pedOUT)
}
\arguments{
\item{gdsProfile}{an object of class \link[gdsfmt]{gds.class} (a GDS
file), the open Profile GDS file.}

\item{listSample}{a \code{array} with the sample to keep}

\item{sampleANNO}{a \code{data.frame} with at least column sex and the name
must be sample.id}

\item{pedOUT}{TODO a PATH and file name to the output file}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{export}
