% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{gds2tped}
\alias{gds2tped}
\title{create a file tped file for plink from the gds file}
\usage{
gds2tped(gds, listSample, listSNP, pedOUT)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{listSample}{a \code{array} with the sample to keep}

\item{listSNP}{a \code{array} with the snp.id to keep}

\item{pedOUT}{TODO a PATH and file name to the output file}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{export}
