% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{generateGDS1KG}
\alias{generateGDS1KG}
\title{Generate the GDS file that will contain the information from 1KG}
\usage{
generateGDS1KG(
  PATHGENO = file.path("data", "sampleGeno"),
  fileNamePED,
  fileListSNP,
  fileSNPSel,
  fileNameGDS,
  listSamples = NULL
)
}
\arguments{
\item{PATHGENO}{a \code{character} string representing the path where
the 1K genotyping files for each sample are located. The name of the
genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file.
Default: \code{"./data/sampleGeno"}.

The directory sampleGeno must contain matFreqSNV.txt.bz2}

\item{fileNamePED}{TODO}

\item{fileListSNP}{TODO}

\item{fileSNPSel}{TODO}

\item{fileNameGDS}{TODO}

\item{listSamples}{a \code{vector} of \code{character} string corresponding
to samples (must be the sample.ids) that will be retained and added to the
GDS file. When \code{NULL}, all the samples are retained.
Default: \code{NULL}.}
}
\value{
None.
}
\description{
This function generates the GDS file that will contain the
information from 1KG. The function also add the samples information, the
SNP information and the genotyping information into the GDS file.
}
\details{
More information about GDS file format can be found here:
[Bioconductor gdsfmt](https://bioconductor.org/packages/gdsfmt/)
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
