% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{generateGDS1KG}
\alias{generateGDS1KG}
\title{Generate the base GDS file from 1KG}
\usage{
generateGDS1KG(
  PATHGENO = file.path("data", "sampleGeno"),
  fileNamePED,
  fileListSNP,
  fileSNPSel,
  fileNameGDS,
  listSamples = NULL
)
}
\arguments{
\item{PATHGENO}{TODO a PATH to the directory genotype file of 1KG
The directory sampleGeno must contain matFreqSNV.txt.bz2}

\item{fileNamePED}{TODO}

\item{fileListSNP}{TODO}

\item{fileSNPSel}{TODO}

\item{fileNameGDS}{TODO}

\item{listSamples}{a \code{vector} of \code{string} corresponding to the sample.ids
if NULL all the samples}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
