% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{generateGDS1KGgenotypeFromSNPPileup}
\alias{generateGDS1KGgenotypeFromSNPPileup}
\title{TODO This function append the genotype and the file related to the pileup}
\usage{
generateGDS1KGgenotypeFromSNPPileup(
  gds,
  PATHGENO,
  listSamples,
  listPos,
  offset,
  minCov = 10,
  minProb = 0.999,
  seqError = 0.001
)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{PATHGENO}{TODO a PATH to a directory with the a file for each samples with the genotype.}

\item{listSamples}{a \code{array} with the sample to keep}

\item{listPos}{a \code{array}}

\item{offset}{a \code{integer} to adjust if the genome start at 0 or 1}

\item{minCov}{a \code{array} with the sample to keep}

\item{minProb}{a \code{array} with the sample to keep}

\item{seqError}{a \code{array} with the sample to keep}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
\keyword{internal}
