% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper_internal.R
\encoding{UTF-8}
\name{generateGDS1KGgenotypeFromSNPPileup}
\alias{generateGDS1KGgenotypeFromSNPPileup}
\title{TODO This function append the genotype and the file related to the
pileup}
\usage{
generateGDS1KGgenotypeFromSNPPileup(
  pathGeno,
  listSamples,
  listPos,
  offset,
  minCov = 10,
  minProb = 0.999,
  seqError = 0.001,
  pedStudy,
  batch,
  studyDF,
  PATHGDSSAMPLE = NULL,
  genoSource = c("snp-pileup", "generic"),
  verbose
)
}
\arguments{
\item{pathGeno}{a \code{character} string representing the path to a
directory with the genotype files for the profiles, as generated by
snp-pileup. The path must exist.}

\item{listSamples}{a \code{vector} of \code{character} string corresponding
to the sample identifiers that will have a GDS Sample file created. The
sample identifiers must be present in the "Name.ID" column of the
\code{data.frame} passed to the \code{pedDF} parameter.}

\item{listPos}{a \code{array} TODO}

\item{offset}{a \code{integer} to adjust if the genome start at 0 or 1}

\item{minCov}{a single positive \code{integer} representing the minimum
coverage needed to keep the SNPs in the analysis. Default: \code{10}.}

\item{minProb}{a \code{array} with the sample to keep}

\item{seqError}{a single positive \code{numeric} between 0 and 1
representing the sequencing error rate. Default: \code{0.001}.}

\item{batch}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.}

\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings.}

\item{PATHGDSSAMPLE}{TODO a PATH to a directory where a gds specific
to the samples with coverage info is keep}

\item{genoSource}{a \code{stirng} with two possible values:
snp-pileup and generic. It specify if the genotype files
are generate by snp-pileup(Facets) or generic format csv
with the column at least the columns:
Chromosome,Position,Ref,Alt,Count,File1R,File1A
where Count is the deep at the position,
FileR is the deep of the reference allele, and
File1A is the deep of the specific alternative allele}

\item{verbose}{a \code{logical} indicating if the function must print
messages when running.}

\item{pedDF}{a \code{data.frame} with the information about the sample(s).
Those are mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in
\code{character} strings. The \code{data.frame}
must contain the information for all the samples passed in the
\code{listSamples} parameter.}
}
\value{
The  function returns \code{0L} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO
gds <- "Demo GDS TODO"


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
