% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDSgenotype}
\alias{generateGDSgenotype}
\title{Add information related to profile genotype into a Reference GDS file}
\usage{
generateGDSgenotype(gds, pathGeno, fileSNPsRDS, listSamples, verbose)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened Reference GDS file.}

\item{pathGeno}{a \code{character} string representing the path where
the 1K genotyping files for each sample are located. The name of the
genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file.}

\item{fileSNPsRDS}{a \code{character} string representing the path and file
name of the RDS file that contains the indexes of the retained SNPs. The
file must exist. The file must be a RDS file.}

\item{listSamples}{a \code{vector} of \code{character} string corresponding
to profiles (must be the profile identifiers) that will be
retained and added to the Reference GDS file.}

\item{verbose}{a \code{logical} indicating if the function must print
messages when running.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function adds the genotype fields with the associated
information into the Reference GDS file for the selected profiles.
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
