% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDSgenotype}
\alias{generateGDSgenotype}
\title{This function create the field genotype in the gds file}
\usage{
generateGDSgenotype(gds, PATHGENO, fileLSNP, listSamples)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{PATHGENO}{TODO a PATH to a directory with the a file for each samples
with the genotype.}

\item{fileLSNP}{TODO list of SNP to keep in the file genotype}

\item{listSamples}{a \code{array} with the sample to keep}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
