% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{generateGDSSample}
\alias{generateGDSSample}
\title{This function initialise the sample information in the gds file.
The information is extract from data.frame pedDF}
\usage{
generateGDSSample(gds, pedDF, listSamples = NULL)
}
\arguments{
\item{gds}{a \code{gds}.}

\item{pedDF}{a \code{data.frame} with the sample info. Must have the column
sample.id, Name.ID, sex, pop.group, superPop and batch. The unique id of pedDF
is Name.ID and the row.name is Name.ID too.}

\item{listSamples}{a \code{array} with the sample from pedDF to keep}
}
\value{
the a \code{array} with the sample from pedDF keept
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
\keyword{internal}
