% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDSSNPinfo}
\alias{generateGDSSNPinfo}
\title{Add information related to SNVs into a Reference GDS file}
\usage{
generateGDSSNPinfo(gdsReference, fileFREQ, verbose)
}
\arguments{
\item{gdsReference}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened Reference GDS file.}

\item{fileFREQ}{a \code{character} string representing the path and file
name of the RDS file with the filtered SNP information.}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function.}
}
\value{
The integer \code{0L} when successful.
}
\description{
the function adds the SNV information into a Reference
GDS file.
}
\examples{

## Required package
library(gdsfmt)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## The RDS file containing the filtered SNP information
fileFilerterSNVs <- file.path(dataDir, "mapSNVSelected_Demo.rds")

## Different code depending of the withr package availability
if (requireNamespace("withr", quietly = TRUE)) {

    ## Temporary Reference GDS file
    file1KG <- withr::local_file("1KG_TEMP_002.gds")
    filenewGDS <- createfn.gds(file1KG)

    ## Add SNV information to Reference GDS
    RAIDS:::generateGDSSNPinfo(gdsReference=filenewGDS,
        fileFREQ=fileFilerterSNVs, verbose=TRUE)

    ## Close GDS file (important)
    closefn.gds(filenewGDS)

    ## Remove temporary 1KG_TEMP_002.gds file
    withr::deferred_run()

} else {

    ## Temporary Reference GDS file
    file1KG <- file.path("1KG_TEMP_002.gds")
    filenewGDS <- createfn.gds(file1KG)

    ## Add SNV information to Reference GDS
    RAIDS:::generateGDSSNPinfo(gdsReference=filenewGDS,
        fileFREQ=fileFilerterSNVs, verbose=TRUE)

    ## Close GDS file (important)
    closefn.gds(filenewGDS)

    ## Remove temporary 1KG_TEMP_002.gds file
    unlink(file1KG)

}

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
