% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDSSNPinfo}
\alias{generateGDSSNPinfo}
\title{This function create the fields related to the snp}
\usage{
generateGDSSNPinfo(gds, fileFREQ)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{fileFREQ}{string with the path and the fileNames to a files with
frequence information TODO describe the file}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
