% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{generateGDSSNPinfo}
\alias{generateGDSSNPinfo}
\title{This function creates the fields related to the snp TODO}
\usage{
generateGDSSNPinfo(gds, fileFREQ, verbose = TRUE)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{fileFREQ}{a \code{character} string with the path and the file
naame to a RDS file containing the frequency information
TODO describe the file}

\item{verbose}{a \code{logical} indicating if messages should be printed
to show how the different steps in the function. Default: \code{TRUE}.}
}
\value{
The integer \code{0L} when successful.
}
\description{
TODO
}
\examples{

# TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
