% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG_internal.R
\encoding{UTF-8}
\name{generateGeneBlock}
\alias{generateGeneBlock}
\title{Generate two indexes based on gene annotation for gdsAnnot1KG block}
\usage{
generateGeneBlock(gdsReference, winSize = 10000, EnsDb)
}
\arguments{
\item{gdsReference}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file (reference).}

\item{winSize}{a single positive \code{integer} representing the
size of the window to use to group the SNVs when the SNVs are in a
non-coding region. Default: \code{10000}.}

\item{EnsDb}{An object of class \code{EnsDb} with the Ensembl genome
annotation. By default, the \code{EnsDb.Hsapiens.v86} class has been used.}
}
\value{
a \code{data.frame} with those columns:
\itemize{
\item{chr} {}
\item{pos} {}
\item{snp.allele} {}
\item{Exon} {}
\item{GName} {}
\item{Gene} {}
\item{GeneS} {}
}
"chr", "pos", "snp.allele", "Exon", "GName", "Gene", "GeneS"
Example for GName and the two indexes "Gene", "GeneS"
GName Gene GeneS
470                                 ENSG00000230021   17  3820
471                                 ENSG00000230021   17  3820
472                 ENSG00000230021:ENSG00000228794   17  3825
473                 ENSG00000230021:ENSG00000228794   17  3825
481 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
482 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
483 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
492                 ENSG00000230021:ENSG00000228794   17  3825
493                 ENSG00000230021:ENSG00000228794   17  3825
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
\keyword{internal}
