% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\name{generateGeneBlock}
\alias{generateGeneBlock}
\title{Generate two indexes base on gene annotation for gdsAnnot1KG block}
\usage{
generateGeneBlock(gds, winSize = 10000, EnsDb = EnsDb.Hsapiens.v86)
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the 1KG GDS file.}

\item{winSize}{TODO}

\item{EnsDb}{An object with the ensembl genome annotation
Default: \code{EnsDb.Hsapiens.v86}.}
}
\value{
a \code{data.frame} with those columns:
"chr", "pos", "snp.allele", "Exon", "GName", "Gene", "GeneS"
Example for GName and the two indexes "Gene", "GeneS"
GName Gene GeneS
470                                 ENSG00000230021   17  3820
471                                 ENSG00000230021   17  3820
472                 ENSG00000230021:ENSG00000228794   17  3825
473                 ENSG00000230021:ENSG00000228794   17  3825
481 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
482 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
483 ENSG00000230021:ENSG00000228794:ENSG00000225880   17  3826
492                 ENSG00000230021:ENSG00000228794   17  3825
493                 ENSG00000230021:ENSG00000228794   17  3825
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
\keyword{internal}
