% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{generateMapSnvSel}
\alias{generateMapSnvSel}
\title{Generate the filter SNP information file in RDS format}
\usage{
generateMapSnvSel(cutOff = 0.01, fileSNV, fileLSNP, fileFREQ)
}
\arguments{
\item{cutOff}{a single \code{numeric} value, the cut-off
for the frequency in at least one super population. Default: \code{0.01}.}

\item{fileSNV}{a \code{character} string representing the path and
file name of the bulk SNP information file from 1KG. The file must exist.}

\item{fileLSNP}{TODO}

\item{fileFREQ}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
The function applies a cut-off filter to the SNP information
file to retain only the SNP that have a frequency superior or equal to the
specified cut-off in at least one super population. The information about
the retained SNPs is saved in a RDS format file. A RDS file containing the
indexes of the retained SNP is also created.
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
