% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{generatePhase1KG2GDS}
\alias{generatePhase1KG2GDS}
\title{TODO}
\usage{
generatePhase1KG2GDS(gds, gdsPhase, PATHGENO, fileLSNP, verbose = FALSE)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{gdsPhase}{TODO}

\item{PATHGENO}{a \code{character} string representing the path where
the 1K genotyping files for each sample are located. The name of the
genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file.
Default: \code{"./data/sampleGeno"}.}

\item{fileLSNP}{TODO}

\item{verbose}{a \code{logicial} indicating if the function should
print messages when running. Default: \code{FALSE}.}
}
\value{
TODO a \code{vector} of \code{string}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
