% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{getRef1KGPop}
\alias{getRef1KGPop}
\title{Extract the specified column from the 1KG GDS 'sample.ref' node
for the reference samples}
\usage{
getRef1KGPop(gds, popName = "superPop")
}
\arguments{
\item{gds}{an object of class
\link[gdsfmt]{gds.class} (a GDS file), the opened 1KG GDS file.}

\item{popName}{a \code{character} string representing the name of the column
that will be fetched in the \code{data.frame} present in the 1KG GDS
"sample.ref" node. The column must be present in the \code{data.frame}.
 Default: \code{"superPop"}.}
}
\value{
\code{data.frame} TODO study.annot with study.annot == study.id and
with the column popName.
}
\description{
The function extract the specified column for the 'sample.ref'
node present in the 1KG GDS file. The column must be present in the
\code{data.frame} saved in the 'sample.ref' node. Only the information for
the reference samples is returned.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## Open existing 1K GDS file with "sample.ref" node
GDS_file <- file.path(data.dir, "1KG_Demo_with_sampleREF.gds")
gdsFile <- snpgdsOpen(GDS_file)

## Extract super population information for the 1KG samples
getRef1KGPop(gds=gdsFile, popName="superPop")

## Close 1K GDS file
closefn.gds(gdsFile)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
