% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{getTableSNV}
\alias{getTableSNV}
\title{TODO}
\usage{
getTableSNV(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  minCov = 10,
  minProb = 0.999,
  eProb = 0.001
)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a 1KG
GDS file.}

\item{gdsSample}{an object of class \code{\link[gdsfmt]{gdsn.class}}
(a GDS node), or \code{\link[gdsfmt]{gds.class}} (a GDS file) containing
the information about one sample (a GDS Sample file).}

\item{sampleCurrent}{a \code{character} string corresponding to
the sample identifier used in \code{\link{pruningSample}} function.}

\item{study.id}{a \code{character} string corresponding to the study
identifier used in \code{\link{pruningSample}} function.}

\item{minCov}{a single positive \code{integer} representing the
minimum coverage needed to retain TODO. Default: \code{10}.}

\item{minProb}{a single \code{numeric} between \code{0} and \code{1}
representing TODO.
Default: \code{0.999}.}

\item{eProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error. Default: \code{0.001}.}
}
\value{
a \code{data.frame} containing:
\itemize{
\item{cnt.tot} {a single \code{integer} representing the total coverage for
the SNV.}
\item{cnt.ref} {a single \code{integer} representing the coverage for
the reference allele.}
\item{cnt.alt} {a single \code{integer} representing the coverage for
the alternative allele.}
\item{snp.pos} {a single \code{integer} representing the SNV position.}
\item{snp.chr} {a single \code{integer} representing the SNV chromosome.}
\item{normal.geno} {a single \code{numeric} \code{3} indicating that
the normal genotype is unknown. TODO}
\item{snp.index} {The \code{boolean} \code{FALSE} indicating TODO}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
