% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{groupChr1KGSNV}
\alias{groupChr1KGSNV}
\title{Merge the pruning files by chromosome in one file}
\usage{
groupChr1KGSNV(PATHGENOCHR, PATHOUT)
}
\arguments{
\item{PATHGENOCHR}{TODO}

\item{PATHOUT}{TODO}
}
\value{
The integer \code{0} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
