% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{groupChrPruning}
\alias{groupChrPruning}
\title{Merge the pruning files by chromosome in one file}
\usage{
groupChrPruning(PATHPRUNED, filePref, fileOUT)
}
\arguments{
\item{PATHPRUNED}{}

\item{filePref}{}

\item{fileOUT}{}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
\keyword{internal}
