% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{groupChrPruning}
\alias{groupChrPruning}
\title{Merge the pruning files by chromosome in one file}
\usage{
groupChrPruning(pathPrunedGDS, filePref, fileOUT)
}
\arguments{
\item{pathPrunedGDS}{TODO}

\item{filePref}{TODO}

\item{fileOUT}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
