% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{identifyRelative}
\alias{identifyRelative}
\title{Identify genetically related and unrelated patients in 1KG files}
\usage{
identifyRelative(gds, maf = 0.05, thresh = 2^(-11/2), fileIBD, filePart)
}
\arguments{
\item{gds}{a \code{character} string representing the path and file
name of the GDS file that contains the 1KG information. The GDS file must
contain the SNP information, the genotyping information and
the pedigree information from 1000 Genomes.
The extension of the file must be '.gds'.}

\item{maf}{a single \code{numeric} representing the threshold for the minor
allele frequency. Only the SNPs with ">= maf" will be used.
Default: \code{0.05}.}

\item{thresh}{a single \code{numeric} representing the threshold value used
to decide if a pair of individuals is ancestrally divergent.
Default: \code{2^(-11/2)}.}

\item{fileIBD}{TODO}

\item{filePart}{TODO}
}
\value{
\code{NULL} invisibly.
}
\description{
The function identify patients that are genetically related in
the 1KG files. It generates a list of unrelated as well as a list of
related patients.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
