% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{identifyRelative}
\alias{identifyRelative}
\title{Identify genetically related and unrelated patients in 1KG files}
\usage{
identifyRelative(gds, maf = 0.05, thresh = 2^(-11/2), fileIBD, filePart)
}
\arguments{
\item{gds}{TODO}

\item{maf}{TODO}

\item{thresh}{TODO}

\item{fileIBD}{TODO}

\item{filePart}{TODO}
}
\value{
\code{NULL} invisibly.
}
\description{
The function identify patients that are genetically related in
the 1KG files. It generates a list of unrelated as well as a list of
related patients.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
