% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{identifyRelative}
\alias{identifyRelative}
\title{Extract a list of id unrelated and a list related}
\usage{
identifyRelative(gds, maf = 0.05, thresh = 2^(-11/2), fileIBD, filePart)
}
\arguments{
\item{gds}{TODO}

\item{maf}{TODO}

\item{thresh}{TODO}

\item{fileIBD}{TODO}

\item{filePart}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
