% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{minCov<-,RAIDSparam-method}
\alias{minCov<-,RAIDSparam-method}
\title{Setter function for replacement of minCov slot in
a RAIDSparam class}
\usage{
\S4method{minCov}{RAIDSparam}(x) <- value

\S4method{minCov}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single positive \code{integer} representing the minimum
required coverage.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.

the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of minCov slot in a
\code{RAIDSparam} class.

A function for replacement of minCov slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the minCov slot in the object
minCov(paramDemo) <- 3L


## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the minCov slot in the object
minCov(paramDemo) <- 3L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
