% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{np<-,RAIDSparam-method}
\alias{np<-,RAIDSparam-method}
\title{Setter function for replacement of np slot in
a RAIDSparam class}
\usage{
\S4method{np}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a single positive \code{numeric} between \code{0} and
\code{1} that represents the frequency of phase switching in the
synthetic profiles.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of np slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new value to the np slot in the object
np(paramDemo) <-3L

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
