% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{paramAncestry,RAIDSparam-method}
\alias{paramAncestry,RAIDSparam-method}
\title{A getter for the paramAncestry slot in a RAIDSparam class}
\usage{
\S4method{paramAncestry}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a \code{list} of parameters related to ancestry.
}
\description{
A function for getting the paramAncestry slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the paramAncestry slot for the object
paramAncestry(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
