% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\name{pedStudy,RAIDSparam-method}
\alias{pedStudy,RAIDSparam-method}
\title{Getter method for pedStudy slot in the "RAIDSparam" class}
\usage{
\S4method{pedStudy}{RAIDSparam}(x)
}
\description{
Getter method for pedStudy slot in the "RAIDSparam" class
}
