% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy<-,RAIDSparam-method}
\alias{pedStudy<-,RAIDSparam-method}
\title{Setter function for replacement of pedStudy slot in a RAIDSparam class}
\usage{
\S4method{pedStudy}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{data.frame} containing the information
TODO with those mandatory columns: "Name.ID", "Case.ID", "Sample.Type",
"Diagnosis", and "Source". All columns must be character strings
(no factors). All row names should correspond to the values in the
"Name.ID" column.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of pedStudy slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Demo PED study
pedStudy <- data.frame(Name.ID=c("TEST11", "Test12"),
    Case.ID=c("Case1", "Case2"), 
    Sample.Type=c("Cancer", "Cancer"), 
    Diagnosis=c("sarcoma", "sarcoma"),
    Source=c("CSHL", "CSHL"), stringsAsFactors=FALSE, 
    row.names=c("TEST11", "Test12"))

## Assign the new PED study to the pedStudy slot in the object
pedStudy(paramDemo) <- pedStudy

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
