% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy<-}
\alias{pedStudy<-}
\title{Generic function for replacement of pedStudy slot in a class}
\usage{
pedStudy(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{a \code{data.frame} containing the information
TODO with those mandatory columns: "Name.ID", "Case.ID", "Sample.Type",
"Diagnosis", and "Source". All columns must be character strings
(no factors). All row names should correspond to the values in the
"Name.ID" column.}
}
\value{
the modified object when the new value is valid.
}
\description{
A generic function for replacement of pedStudy slot in a
class. The function is implemented for the \code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
