% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pedStudy<-}
\alias{pedStudy<-}
\title{Generic function for replacement of pedStudy slot in a class}
\usage{
pedStudy(x) <- value
}
\description{
A generic function for replacement of pedStudy slot in a
class. The function is implemented for the \code{RAIDSparam} class.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
