% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\name{prepPed1KG}
\alias{prepPed1KG}
\title{Prepare the ped file for 1KG}
\usage{
prepPed1KG(pedFile, PATHGENO = file.path("data", "sampleGeno"), batch.v = 0)
}
\arguments{
\item{pedFile}{TODO a \code{string} path and file name of ped file from 1KG}

\item{PATHGENO}{TODO}

\item{batch.v}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
