% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{prepPed1KG}
\alias{prepPed1KG}
\title{Prepare the pedigree file using pedigree information from 1KG}
\usage{
prepPed1KG(pedFile, pathGeno = file.path("data", "sampleGeno"), batch.v = 0L)
}
\arguments{
\item{pedFile}{a \code{character} string representing the path and
file name of the pedigree file (PED file) that contains the information
related to the profiles present in the 1KG GDS file. The PED file must
exist.}

\item{pathGeno}{a \code{character} string representing the path where
the 1KG genotyping files for each profile are located. Only the profiles
with associated genotyping files are retained in the creation of the final
\code{data.frame}. The name of the genotyping files must correspond to
the individual identification (Individual.ID) in the pedigree file
(PED file).
Default: \code{"./data/sampleGeno"}.}

\item{batch.v}{a\code{integer} that uniquely identifies the source of the
pedigree information. The 1KG is usually \code{0L}. Default: \code{0L}.}
}
\value{
a \code{data.frame} containing the needed pedigree information
from 1KG. The \code{data.frame} contains those columns:
\itemize{
\item{sample.id}{a \code{character} string representing the profile unique
ID.}
\item{Name.ID}{a \code{character} string representing the profile name.}
\item{sex}{a \code{character} string representing the sex of the profile.}
\item{pop.group}{a \code{character} string representing the
sub-continental ancestry of the profile.}
\item{superPop }{a \code{character} string representing the continental
ancestry of the profile.}
\item{superPop }{a \code{integer} representing the batch of the profile.}
}
}
\description{
Using the pedigree file from 1KG, this function extracts
needed information and formats it into a \code{data.frame} so in can
be used in following steps of the ancestry inference process. The
function also requires that the genotyping files associated to each
sample be available in a specified directory.
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## Demo pedigree file
pedDemoFile <- file.path(dataDir, "PedigreeDemo.ped")

## Create a data.frame containing the information of the retained
## samples (samples with existing genotyping files)
prepPed1KG(pedFile=pedDemoFile, pathGeno=dataDir, batch.v=0L)


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
