% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryAnalysis.R
\encoding{UTF-8}
\name{prepPedSyntheticRef}
\alias{prepPedSyntheticRef}
\title{Extract the sample information from the 1KG GDS file for a list
of profiles associated to a specific study in the Profile GDS file}
\usage{
prepPedSyntheticRef(pRAIDS)
}
\arguments{
\item{pRAIDS}{an object of class TODO.}
}
\value{
\code{data.frame} containing the columns extracted from the
GDS Sample 'study.annot' node with a extra column named as the 'popName'
parameter that has been extracted from the 1KG GDS 'sample.annot' node.
Only the rows corresponding to the specified study ('studyID' parameter)
are returned.
}
\description{
The function extracts the information for the profiles
associated to a specific study in the GDS Sample file. The information is
extracted from the 'study.annot' node as a 'data.frame'.

Then, the function used the 1KG GDS file to extract specific information
about each sample and add it, as an extra column, to the 'data.frame'.

As example, this function can extract the synthetic profiles
for a GDS Sample and the super-population of the 1KG samples used to
generate each synthetic profile would be added
as an extra column to the final 'data.frame'.
}
\details{
As example, this function can extract the synthetic profiles
for a Profile GDS and the super-population of the 1KG samples used to
generate each synthetic profile would be added
as an extra column to the final 'data.frame'. In that situation, the
'popName' parameter would correspond to the super-population column and the
'studyID' parameter would be the name given to the synthetic dataset.
}
\examples{

## Required library
library(gdsfmt)

## TODO


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
