% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{prepSynthetic}
\alias{prepSynthetic}
\title{Prepare the annotations for the synthetic data}
\usage{
prepSynthetic(
  gdsSampleFile,
  listSampleRef,
  data.id.profile,
  studyDF,
  nbSim = 1L,
  prefId = "",
  pRecomb = 0.01,
  minProb = 0.999,
  seqError = 0.001
)
}
\arguments{
\item{gdsSampleFile}{a \code{character} string representing the file name
of the GDS Sample file containing the information about the sample
to be analyzed.}

\item{listSampleRef}{a \code{character} string representing the
identifiers of the selected 1KG samples that will be used as reference.}

\item{data.id.profile}{a \code{character} string TODO}

\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor).}

\item{nbSim}{a single positive \code{integer} representing the number of
simulations per combination of sample and profile. Default: \code{1L}.}

\item{prefId}{a \code{character} string TODO. Default: \code{""}.}

\item{pRecomb}{a single \code{numeric} between 0 and 1 TODO.
Default: \code{0.01}.}

\item{minProb}{a single \code{numeric} TODO. Default: \code{0.999}.}

\item{seqError}{a single positive \code{numeric} between 0 and 1
representing the sequencing error rate. Default: \code{0.001}.}
}
\value{
\code{0L} when successful.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
