% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{prepSynthetic}
\alias{prepSynthetic}
\title{TODO}
\usage{
prepSynthetic(
  gdsSampleFile,
  listSampleRef,
  data.id.profile,
  studyDF,
  nbSim = 1,
  pRecomb = 0.01,
  minProb = 0.999,
  seqError = 0.001
)
}
\arguments{
\item{gdsSampleFile}{the path of an object of class \code{gds} related to
the sample}

\item{listSampleRef}{a \code{string} TODO}

\item{data.id.profile}{a \code{string} TODO}

\item{studyDF}{TODO}

\item{nbSim}{a \code{integer} TODO}

\item{pRecomb}{a \code{numeric} between 0 and 1 TODO}

\item{gds}{an object of class \code{gds} opened}
}
\value{
TODO a \code{vector} of \code{string}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
