% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{processBlockChr}
\alias{processBlockChr}
\title{Extract the block information from the block file associated to a
specific population and chromosome}
\usage{
processBlockChr(snpKeep, pathBlock, superPop, chr)
}
\arguments{
\item{snpKeep}{a \code{vector} of \code{integer} representing the
positions of the retained SNVs for the current chromosome.}

\item{pathBlock}{a \code{character} string representing the path to the
block files.}

\item{superPop}{a \code{character} string representing the current
continental population (ex: EUR, AFR, etc.).}

\item{chr}{a \code{integer} representing the current chromosome.}
}
\value{
a \code{array} pf TODO with the sample from pedDF keept
}
\description{
The function reads the information form the specified block
file associated to a specified population and a specified chromosome. The
the block information for a list of SNV positions is formatted and returned.
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
