% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{processBlockChr}
\alias{processBlockChr}
\title{TODO}
\usage{
processBlockChr(snpKeep, pathBlock, superPop, chr)
}
\arguments{
\item{snpKeep}{TODO}

\item{pathBlock}{a \code{character} string representing the path to the
block files.}

\item{superPop}{a \code{character} string representing the current
continental population (ex: EUR, AFR, etc.).}

\item{chr}{a \code{integer} representing the current chromosome.}
}
\value{
the a \code{array} with the sample from pedDF keept
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
