% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_internal.R
\encoding{UTF-8}
\name{processPileupChrBin}
\alias{processPileupChrBin}
\title{Extract SNV information from pileup file for a selected chromosome}
\usage{
processPileupChrBin(chr, resPileup, varDf, verbose)
}
\arguments{
\item{chr}{a \code{character} string representing the name of the
chromosome to keep}

\item{resPileup}{a \code{data.frame} as generated by the \code{pileup}
function from \code{Rsamtools} package}

\item{varDf}{a \code{data.frame} representing the positions to keep}

\item{verbose}{a \code{logical} indicating if messages should be printed}
}
\value{
a \code{data.frame} containing at least:
\describe{
\item{seqnames}{ a \code{character} representing the name of the chromosome}
\item{pos}{ a \code{numeric} representing the position on the chromosome}
\item{REF}{ a \code{character} string representing the reference nucleotide}
\item{ALT}{ a \code{character} string representing the alternative
nucleotide}
\item{A}{ a \code{numeric} representing the count for the A nucleotide}
\item{C}{ a \code{numeric} representing the count for the C nucleotide}
\item{G}{ a \code{numeric} representing the count for the G nucleotide}
\item{T}{ a \code{numeric} representing the count for the T nucleotide}
\item{count}{ a \code{numeric} representing the total count}
}
}
\description{
The function reads pileup file and
returns a \code{data.frame}
containing the information about the read counts for the SNVs present in
the selected chromosome.
}
\examples{

## Directory where demo SNP-pileup file
dataDir <- system.file("extdata/example/snpPileup", package="RAIDS")


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
