% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{profileFile,RAIDSparam-method}
\alias{profileFile,RAIDSparam-method}
\title{A getter for the profileFile slot in a RAIDSparam class}
\usage{
\S4method{profileFile}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
\code{NULL} or \code{character} string representing the path to
the file with genotype and the allele information of the profile.
}
\description{
A function for getting the profileFile slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the profileFile slot for the object
profileFile(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
