% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\name{projectSample2PCA}
\alias{projectSample2PCA}
\title{TODO}
\usage{
projectSample2PCA(gds, listPCA, sample.current, np = 1)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{listPCA}{a \code{list}  with with two objects pca.unrel -> \code{snpgdsPCAClass}
and a snp.load -> \code{snpgdsPCASNPLoading}}

\item{sample.current}{the sample.id to project in the PCA}

\item{np}{a \code{integer} reprenting the number of thread}
}
\value{
TODO Objects p\code{snpgdsPCAClass}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
