% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{projectSample2PCA}
\alias{projectSample2PCA}
\title{Project profile onto existing principal component axes (PCA)}
\usage{
projectSample2PCA(gds, listPCA, currentProfile, np = 1L, verbose = FALSE)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, an
opened Profile GDS file.}

\item{listPCA}{a \code{list} containing two entries:
\itemize{
\item{pca.unrel} {\code{snpgdsPCAClass} object}
\item{snp.load} {\code{snpgdsPCASNPLoading} object}
}}

\item{currentProfile}{a \code{character} string representing the
identifiant of the profile to be projected in the PCA.}

\item{np}{a single positive \code{integer} representing the number of
threads. Default: \code{1L}.}

\item{verbose}{a \code{logical} passed to the PCA function.
Default: \code{FALSE}.}
}
\value{
a \code{snpgdsPCAClass} object, a \code{list} that contains:
\itemize{
   \item{sample.id} {the sample ids used in the analysis}
   \item{snp.id} {the SNP ids used in the analysis}
   \item{eigenvalues} {eigenvalues}
   \item{eigenvect} {eigenvactors, “# of samples” x “eigen.cnt”}
   \item{TraceXTX} {the trace of the genetic covariance matrix}
   \item{Bayesian} {whether use bayerisan normalization}
}
}
\description{
This function calculates the profile eigenvectors using
the specified SNP loadings.
}
\details{
More information about the method used to calculate the patient eigenvectors
can be found at the Bioconductor SNPRelate website:
https://bioconductor.org/packages/SNPRelate/
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", "RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
