% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1KG.R
\encoding{UTF-8}
\name{pruning1KG.Chr}
\alias{pruning1KG.Chr}
\title{Find the pruned snv in 1KG by chr}
\usage{
pruning1KG.Chr(
  gds,
  method = "corr",
  listSamples = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE,
  chr = NULL,
  minAF = NULL,
  outPref = "pruned_1KG",
  keepObj = FALSE
)
}
\arguments{
\item{gds}{an object of class
\code{\link[SNPRelate:SNPGDSFileClass]{SNPRelate::SNPGDSFileClass}}, a SNP
GDS file.}

\item{method}{a \code{character string} TODO . Default: \code{"corr"}.}

\item{listSamples}{TODO}

\item{slide.max.bp.v}{TODO}

\item{ld.threshold.v}{TODO}

\item{np}{TODO . Default: \code{NULL}.}

\item{verbose.v}{a \code{logical} specifying if the function must provide
more information about the process. Default: \code{FALSE}.}

\item{chr}{TODO}

\item{minAF}{TODO}

\item{outPref}{TODO}

\item{keepObj}{a \code{logical} specifying if the function must save the
the processed information into a RDS object. Default: \code{FALSE}.}
}
\value{
\code{NULL} invisibly.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
