% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{pruningMethod,RAIDSparam-method}
\alias{pruningMethod,RAIDSparam-method}
\title{A getter for the pruningMethod slot in a RAIDSparam class}
\usage{
\S4method{pruningMethod}{RAIDSparam}(x)

\S4method{pruningMethod}{RAIDSparam}(x)
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}
}
\value{
a single \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function.

a single \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function.
}
\description{
A function for getting the pruningMethod slot in a
\code{RAIDSparam} class.

A function for getting the pruningMethod slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pruningMethod slot for the object
pruningMethod(paramDemo)


## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Extract the pruningMethod slot for the object
pruningMethod(paramDemo)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
