% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{pruningSample}
\alias{pruningSample}
\title{Compute the list of pruned SNVs for a specific sample using the
information from the 1KG GDS file and a linkage disequilibrium analysis}
\usage{
pruningSample(
  gds,
  method = c("corr", "r", "dprime", "composite"),
  sampleCurrent,
  study.id,
  listSNP = NULL,
  slide.max.bp.v = 500000L,
  ld.threshold.v = sqrt(0.1),
  np = 1L,
  verbose.v = FALSE,
  chr = NULL,
  minAF.SuperPop = NULL,
  keepGDSpruned = TRUE,
  PATHSAMPLEGDS = NULL,
  keepFile = FALSE,
  PATHPRUNED = ".",
  outPref = "pruned"
)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gds.class} (a GDS file), the
1 KG GDS file.}

\item{method}{a \code{character} string that represents the method that will
be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime" and "composite". Default: \code{"corr"}.}

\item{sampleCurrent}{a \code{character} string
corresponding to the sample identifier used in LD pruning done by the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. A GDS Sample file
corresponding to the sample identifier must exist and be located in the
\code{PATHSAMPLEGDS} directory.}

\item{study.id}{a \code{character} string corresponding to the study
identifier used in the \code{\link[SNPRelate]{snpgdsLDpruning}} function.
The study identifier must be present in the GDS Sample file.}

\item{listSNP}{a \code{vector} of SNVs identifiers specifying selected to
be passed the the pruning function;
if \code{NULL}, all SNVs are used in the
\code{\link[SNPRelate]{snpgdsLDpruning}} function. Default: \code{NULL}.}

\item{slide.max.bp.v}{a single positive \code{integer} that represents
the maximum basepairs (bp) in the sliding window. This parameter is used
for the LD pruning done in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.
Default: \code{500000L}.}

\item{ld.threshold.v}{a single \code{numeric} value that represents the LD
threshold used in the \code{\link[SNPRelate]{snpgdsLDpruning}} function.
Default: \code{sqrt(0.1)}.}

\item{np}{a single positive \code{integer} specifying the number of
threads to be used. Default: \code{1L}.}

\item{verbose.v}{a \code{logicial} indicating if information is shown
during the process in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function.  Default: \code{FALSE}.}

\item{chr}{a \code{character} string representing the chromosome where the
selected SNVs should belong. Only one chromosome can be handled. If
\code{NULL}, the chromosome is not used as a filtering criterion.
Default: \code{NULL}.}

\item{minAF.SuperPop}{a single positive \code{numeric} representing the
minimum allelic frequency used to select the SNVs. If \code{NULL}, the
allelic frequency is not used as a filtering criterion. Default: \code{NULL}.}

\item{keepGDSpruned}{a \code{logicial} indicating if the information about
the pruned SNVs should be added to the GDS Sample file.
Default: \code{TRUE}.}

\item{PATHSAMPLEGDS}{a \code{character} string representing the directory
where the GDS Sample file will be created. The directory must exist.}

\item{keepFile}{a \code{logical} indicating if RDS files containing the
information about the pruned SNVs must be
created. Default: \code{FALSE}.}

\item{PATHPRUNED}{a \code{character} string representing an existing
directory. The directory must exist. Default: \code{"."}.}

\item{outPref}{a \code{character} string that represents the prefix of the
RDS files that will be generated. The RDS files are only generated when
the parameter \code{keepFile}=\code{TRUE}. Default: \code{"pruned"}.}
}
\value{
The function returns \code{0L} when successful.
}
\description{
This function computes the list of pruned SNVs for a
specific sample. When
a group of SNVs are in linkage disequilibrium, only one SNV from that group
is retained. The linkage disequilibrium is calculated with the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The initial list of
SNVs that are passed to the \code{\link[SNPRelate]{snpgdsLDpruning}}()
function can be specified by the user.
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
