% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{pruningSample}
\alias{pruningSample}
\title{TODO}
\usage{
pruningSample(
  gds,
  method = "corr",
  sampleCurrent,
  study.id,
  listSNP = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE,
  chr = NULL,
  minAF.SuperPop = NULL,
  keepGDSpruned = TRUE,
  PATHSAMPLEGDS = NULL,
  keepFile = FALSE,
  PATHPRUNED = ".",
  outPref = "pruned"
)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{method}{a \code{character} string representing the method used in
SNPRelate::snpgdsLDpruning() function.
Default: \code{"corr"}.}

\item{sampleCurrent}{A \code{character} string corresponding to
the sample.id
use in LDpruning}

\item{study.id}{A \code{string} corresponding to the study
use in LDpruning}

\item{listSNP}{the list of snp.id keep. TODO. Default: \code{NULL}.}

\item{slide.max.bp.v}{a single \code{numeric} TODO. Default: \code{5e5}.}

\item{ld.threshold.v}{a single \code{numeric} TODO.
Default: \code{sqrt(0.1)}.}

\item{np}{TODO. Default: \code{1}.}

\item{verbose.v}{TODO. Default: \code{FALSE}.}

\item{chr}{TODO. Default: \code{NULL}.}

\item{minAF.SuperPop}{TODO. Default: \code{NULL}.}

\item{keepGDSpruned}{a \code{boolean} TODO. Default: \code{TRUE}.}

\item{PATHSAMPLEGDS}{TODO}

\item{keepFile}{a \code{boolean} indicating if the file must be TODO.
Default: \code{FALSE}.}

\item{PATHPRUNED}{a \code{character} string TODO. Default: \code{"."}.}

\item{outPref}{TODO. Default: \code{"pruned"}.}
}
\value{
\code{0L} when successful.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
