% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\docType{class}
\encoding{UTF-8}
\name{RAIDSparam-class}
\alias{RAIDSparam-class}
\title{An S4 class to represent the RAIDS parameters}
\description{
An S4 class to represent the RAIDS parameters
}
\details{
The \code{chrInfo} slot contains the length of the chromosomes. The
length of the chromosomes can be obtain through the
\code{\link[Seqinfo]{seqlengths}} library.

As example, for hg38 genome:

\if{html}{\out{<div class="sourceCode">}}\preformatted{
if (requireNamespace("Seqinfo", quietly=TRUE) &&
     requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly=TRUE)) \{
     chrInfo <- Seqinfo::seqlengths(BSgenome.Hsapiens.UCSC.hg38::Hsapiens)[1:25]
\}

}\if{html}{\out{</div>}}
}
\section{Slots}{

\describe{
\item{\code{studyDF}}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factors). Default:
\code{data.frame(study.id="NotDef", study.desc="NotDef", 
study.platform="NotDef", stringsAsFactors=FALSE)}.}

\item{\code{studyDFSyn}}{a \code{data.frame} containing the information about the
synthetic data to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factors). Default:
\code{data.frame(study.id="NotDef.Synthetic", 
study.desc="NotDef synthetic data", study.platform="Synthetic", 
stringsAsFactors=FALSE)}.}

\item{\code{pedStudy}}{a \code{data.frame} containing the information TODO
with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", and "Source". All columns must be in
\code{character} strings (no factors). All row names should correspond to the
values in the "Name.ID" column.
Default: \code{data.frame(Name.ID=c("ProfileId"), Case.ID=c("ProfileId"), 
Sample.Type=c("type"), Diagnosis="NotDef", Source=c("NotDef"), 
stringsAsFactors=FALSE, row.names = c("ProfileId"))}.}

\item{\code{studyType}}{a single \code{character} string representing the type of
study. The possible choices are: "LD" and "GeneAware". The type of study
affects the way the estimation of the allelic fraction is done.
Default: \code{"LD"}.}

\item{\code{genoSource}}{a single \code{character} string corresponding to the type
of file with the genotype and the allele information of the profile that
will be provided in the 'profileFile' slot. The valid options are: "VCF",
"generic", "snp-pileup", and "bam". The "generic" format CSV file
must have at least these columns: 'Chromosome', 'Position', 'Ref', 'Alt',
'Count', 'File1R', and 'File1A'. The 'Count' is the depth at the
specified position; 'FileR' is the depth of the reference allele, and
'File1A' is the depth of the specific alternative allele.
Finally, in the case of a "VCF" file, the file must have at least those
genotype fields: GT, AD, and DP. Default: \code{NULL}.}

\item{\code{blockTypeId}}{a single \code{character} string corresponding to
the block type used to extract the block identifiers. The block type must
be present in the GDS Reference Annotation file.
Default: \code{"GeneS.Ensembl.Hsapiens.v86"}.}

\item{\code{reference}}{a \code{character} string with two possible values:
'1KGv1.0', '1k_hgdpV0.1'. It specifies the type of inference.
Default: \code{"1KGv1.0"}.}

\item{\code{genome}}{a \code{character} string with one possible value:
'HG38'. It specifies the genome used. Default: \code{"HG38"}.}

\item{\code{chrInfo}}{a \code{vector} of positive \code{integer} values
representing the length of the chromosomes. See 'details' section.
Default: \code{seqlengths(Hsapiens)[seq_len(25)]}.}

\item{\code{paramAncestry}}{a \code{list} of parameters related to ancestry.
The \code{list} should contain those three entries: \code{ScanBamParam},
\code{PileupParam}, and \code{yieldSize}.
Default: \code{list(ScanBamParam=NULL, PileupParam=NULL, 
yieldSize=10000000)}.}

\item{\code{profileFile}}{a \code{character} string representing the path to the
file with genotype and the allele information of the profile. The format
of the accepted file is determined by the value in the 'genoSource' slot.
A profile would have a file with extension "\emph{vcf.gz" when 'genoSource' slot
is "VCF". If 'genoSource' is "generic" or "snp-pileup", then "}.txt.gz".
If 'genoSource' is "bam", then "\emph{.bam" (the file needs to be indexed with a
existing corresponding "}.bai" file). Default: \code{NULL}.}

\item{\code{profileFileGeno}}{TODO}

\item{\code{pathProfileGDS}}{a single \code{character} string representing the
path to the directory where the GDS Profile files will be created. If
specified, the directory must exist.
Default: \code{NULL}.}

\item{\code{fileReferenceGDS}}{a single \code{character} string representing
the file name of the Reference GDS file. If specified, the file
must exist. Default: \code{NULL}.}

\item{\code{fileReferenceAnnotGDS}}{a \code{character} string representing the
file name of the Population Reference GDS Annotation file. If specified,
the file must exist. Default: \code{NULL}.}

\item{\code{inferenceType}}{a single \code{character} string representing the
genotype ancestry inference method. The two possible values:
'PCAknn' and 'haploAdmixture'. Default: \code{"PCAknn"}.}

\item{\code{sampleRef}}{a \code{vector} of \code{character} strings that represent
a subset of sample identifiers from the GDS Reference File. Those samples
will be used for the ancestry inference. If \code{NULL},
all the samples are retained. This is usefull if you want to removed some
specific samples (ex: samples with admixed ancestry).
Default: \code{NULL}.}

\item{\code{batch}}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.
Default: \code{1L}.}

\item{\code{prefix}}{a single \code{character} string representing TODO.
Default: \code{"1"}.}

\item{\code{nbSim}}{a positive \code{integer} representing TODO.
Default: \code{1L}.}

\item{\code{offset}}{a single \code{integer} that is added to the SNP position to
switch from 0-based to 1-based coordinate when needed (or reverse).
Default: \code{-1L}.}

\item{\code{minCov}}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{\code{minProb}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
Default: \code{0.999}.}

\item{\code{seqError}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error.
Default: \code{0.001}.}

\item{\code{seqErrorSyn}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error for synthetic.
Default: \code{0.001}.}

\item{\code{pRecomb}}{a single positive \code{numeric} between \code{0} and
\code{1} that represents the frequency of phase switching in the
synthetic profiles. Default: \code{0.01}.}

\item{\code{np}}{a single positive \code{integer} specifying the number of
threads to be used. Default: \code{1L}.}

\item{\code{listPos}}{a \code{data.frame} containing 2 columns named:
"snp.chromosome" and "snp.position". The first column,
called "snp.chromosome", contains the name of the chromosome where the
SNV is located. The second column, called "snp.position", contains the
position of the SNV on the chromosome. Default: \code{NULL}.}

\item{\code{syntheticRefDF}}{a \code{data.frame} containing a subset of
reference profiles for each sub-population present in the Reference GDS
file. The \code{data.frame} must have those columns:
\describe{
\item{sample.id}{ a \code{character} string representing the sample
identifier. }
\item{pop.group}{ a \code{character} string representing the
subcontinental population assigned to the sample. }
\item{superPop}{ a \code{character} string representing the
super-population assigned to the sample. }
}
Default: \code{NULL}.}

\item{\code{pruningMethod}}{a \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime", and "composite". Default: \code{"corr"}.}

\item{\code{slideWindowMaxBP}}{a single positive \code{integer} that represents
the maximum basepairs (bp) in the sliding window. This parameter is used
for the LD pruning done by the \code{\link[SNPRelate]{snpgdsLDpruning}}
function. Default: \code{500000L}.}

\item{\code{thresholdLD}}{a single positive \code{numeric} value that represents
the LD threshold used in the \code{\link[SNPRelate]{snpgdsLDpruning}}
function. Default: \code{sqrt(0.1)}.}

\item{\code{specificSNV}}{a \code{data.frame} containing 2 columns. The first
column, called "snp.chromosome" contains the name of the chromosome where
the SNV is located. The second column, called "snp.position", contains the
position of the SNV on the chromosome. Optionally, the column "snvKeep"
contains the SNV index in the reference or -1 if not in the reference. It
is used during the pruning step. Default: \code{NULL}.}

\item{\code{genoType}}{a TODO. Default: \code{"geno.ref"}.}

\item{\code{phaseType}}{a TODO. Default: \code{"phase.ref"}.}

\item{\code{phase}}{a \code{logical} indicating TODO. Default: \code{FALSE}.}

\item{\code{PCAmissingRate}}{a positive \code{numeric} representing the maximum
missing rate retained accepted to use SNPs in the PCA analysis done
with the the \link[SNPRelate]{snpgdsPCA} function. If \code{NaN}, no
missing threshold. Default: \code{0.025}.}

\item{\code{PCAalgorithm}}{a single \code{character} string representing the
algorithm to use with the \link[SNPRelate]{snpgdsPCA} function. The
algorithm must be implemented and available to the
\link[SNPRelate]{snpgdsPCA} function. The current options are:
'exact' or 'randomized'.
Default: \code{"exact"}.}

\item{\code{eigenCount}}{a single \code{integer} indicating the number of
eigenvectors that will be in the output of the \link[SNPRelate]{snpgdsPCA}
function; if 'eigenCount' <= 0, then all eigenvectors are returned.
Default: \code{32L}.}

\item{\code{eigenCountSyn}}{a TODO. Default: \code{15L}.}

\item{\code{kList}}{a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{K} parameter. The \emph{K} parameter represents the
number of neighbors used in the K-nearest neighbor analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{\code{pcaList}}{a \code{vector} of positive \code{integer} representing
the values tested for the  \emph{D} parameter. The \emph{D} parameter represents the
number of dimensions used in the PCA analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{\code{fieldPopInRef}}{a \code{character} string representing the name of the
column that contains the known ancestry for the reference profiles in
the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter). Default: \code{"superPop"}.}

\item{\code{fieldPopInfAnc}}{a \code{character} string representing the name of
the column in the data frame that contains the inferred
super-population ancestry for the samples. The column should be
present in the data frame. Default: \code{"superPop"}.}

\item{\code{fieldSubPop}}{a \code{character} string representing the name of
the column in the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter) that contains the sub-population
information for the samples. The column should be present in the Population
Reference GDS file. Default: \code{"pop.group"}.}

\item{\code{verbose}}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}.}
}}

\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{classes}
