% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\docType{class}
\encoding{UTF-8}
\name{RAIDSparams-class}
\alias{RAIDSparams-class}
\title{An S4 class to represent the RAIDS parameters}
\description{
An S4 class to represent the RAIDS parameters
}
\section{Slots}{

\describe{
\item{\code{studyDF}}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). Default: \code{NULL}.}

\item{\code{studyDFSyn}}{a \code{data.frame} containing the information about the
synthetic data to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). Default: \code{NULL}.}

\item{\code{studyType}}{a \code{character} string representing the type of study.
The possible choices are: "LD" and "GeneAware". The type of study affects
the way the estimation of the allelic fraction is done.
Default: \code{"LD"}.}

\item{\code{genoSource}}{TODO}

\item{\code{blockTypeId}}{TODO}

\item{\code{reference}}{a \code{character} string with two possible values:
'1KGv1.0', '1k_hgdpV0.1'. It specifies the type of inference.
Default: \code{"1KGc1.0"}.}

\item{\code{chrInfo}}{TODO}

\item{\code{paramAncestry}}{TODO}

\item{\code{profileFile}}{TODO}

\item{\code{profileFileGeno}}{TODO}

\item{\code{pathProfileGDS}}{TODO}

\item{\code{fileReferenceGDS}}{TODO}

\item{\code{fileReferenceAnnotGDS}}{TODO}

\item{\code{inferenceType}}{TODO}

\item{\code{sampleRef}}{TODO}

\item{\code{batch}}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.
Default: \code{1L}.}

\item{\code{prefix}}{TODO}

\item{\code{nbSim}}{TODO}

\item{\code{offset}}{a single \code{integer} that is added to the SNP position to
switch from 0-based to 1-based coordinate when needed (or reverse).
Default: \code{-1L}.}

\item{\code{minCov}}{TODO}

\item{\code{minProb}}{TODO}

\item{\code{seqError}}{TODO}

\item{\code{seqErrorSyn}}{TODO}

\item{\code{pRecomb}}{TODO}

\item{\code{np}}{TODO}

\item{\code{listPos}}{TODO}

\item{\code{syntheticRefDF}}{TODO}

\item{\code{pruningMethod}}{a \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime", and "composite". Default: \code{"corr"}.}

\item{\code{slideWindowMaxBP}}{TODO}

\item{\code{thresholdLD}}{TODO}

\item{\code{specificSNV}}{TODO}

\item{\code{genoType}}{TODO}

\item{\code{phaseType}}{TODO}

\item{\code{phase}}{TODO}

\item{\code{PCAmissingRate}}{TODO}

\item{\code{PCAalgorithm}}{TODO}

\item{\code{eigenCount}}{TODO}

\item{\code{eigenCountSyn}}{TODO}

\item{\code{kList}}{TODO}

\item{\code{pcaList}}{TODO}

\item{\code{fieldPopInRef}}{TODO}

\item{\code{fieldPopInfAnc}}{TODO}

\item{\code{fieldSubPop}}{TODO}

\item{\code{verbose}}{TODO}
}}

\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
