% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\docType{class}
\encoding{UTF-8}
\name{RAIDSparams-class}
\alias{RAIDSparams-class}
\title{An S4 class to represent the RAIDS parameters}
\description{
An S4 class to represent the RAIDS parameters
}
\section{Slots}{

\describe{
\item{\code{studyDF}}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). Default:
\code{data.frame(study.id="NotDef", study.desc="NotDef", 
study.platform="NotDef", stringsAsFactors=FALSE)}.}

\item{\code{studyDFSyn}}{a \code{data.frame} containing the information about the
synthetic data to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). Default: \code{NULL}.}

\item{\code{pedStudy}}{a \code{data.frame} containing the information TODO
with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", and "Source". All columns must be in
\code{character} strings (no factor). All row names should correspond to the
values in the "Name.ID" column.
Default: \code{data.frame(Name.ID=c("ProfileId"), Case.ID=c("ProfileId"), 
Sample.Type=c("type"), Diagnosis="NotDef", Source=c("NotDef"), 
stringsAsFactors=FALSE, row.names = c("ProfileId"))}.}

\item{\code{studyType}}{a \code{character} string representing the type of study.
The possible choices are: "LD" and "GeneAware". The type of study affects
the way the estimation of the allelic fraction is done.
Default: \code{"LD"}.}

\item{\code{genoSource}}{TODO}

\item{\code{blockTypeId}}{TODO}

\item{\code{reference}}{a \code{character} string with two possible values:
'1KGv1.0', '1k_hgdpV0.1'. It specifies the type of inference.
Default: \code{"1KGc1.0"}.}

\item{\code{genome}}{a \code{character} string with one possible value:
'HG38'. It specifies the genome used. Default: \code{"HG38"}.}

\item{\code{chrInfo}}{TODO}

\item{\code{paramAncestry}}{TODO}

\item{\code{profileFile}}{TODO}

\item{\code{profileFileGeno}}{TODO}

\item{\code{pathProfileGDS}}{TODO}

\item{\code{fileReferenceGDS}}{TODO}

\item{\code{fileReferenceAnnotGDS}}{TODO}

\item{\code{inferenceType}}{TODO}

\item{\code{sampleRef}}{TODO}

\item{\code{batch}}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.
Default: \code{1L}.}

\item{\code{prefix}}{a single \code{character} string representing TODO.
Default: \code{"1"}.}

\item{\code{nbSim}}{a positive \code{integer} representing TODO.
Default: \code{1L}.}

\item{\code{offset}}{a single \code{integer} that is added to the SNP position to
switch from 0-based to 1-based coordinate when needed (or reverse).
Default: \code{-1L}.}

\item{\code{minCov}}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{\code{minProb}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
Default: \code{0.999}.}

\item{\code{seqError}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error.
Default: \code{0.001}.}

\item{\code{seqErrorSyn}}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error for synthetic.
Default: \code{0.001}.}

\item{\code{pRecomb}}{TODO}

\item{\code{np}}{TODO}

\item{\code{listPos}}{TODO}

\item{\code{syntheticRefDF}}{TODO}

\item{\code{pruningMethod}}{a \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime", and "composite". Default: \code{"corr"}.}

\item{\code{slideWindowMaxBP}}{TODO}

\item{\code{thresholdLD}}{TODO}

\item{\code{specificSNV}}{TODO}

\item{\code{genoType}}{TODO}

\item{\code{phaseType}}{TODO}

\item{\code{phase}}{a \code{logical} indicating TODO. Default: \code{FALSE}.}

\item{\code{PCAmissingRate}}{TODO}

\item{\code{PCAalgorithm}}{TODO}

\item{\code{eigenCount}}{TODO}

\item{\code{eigenCountSyn}}{TODO}

\item{\code{kList}}{a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{K} parameter. The \emph{K} parameter represents the
number of neighbors used in the K-nearest neighbor analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{\code{pcaList}}{a \code{vector} of positive \code{integer} representing
the values tested for the  \emph{D} parameter. The \emph{D} parameter represents the
number of dimensions used in the PCA analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{\code{fieldPopInRef}}{a \code{character} string representing the name of the
column that contains the known ancestry for the reference profiles in
the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter). Default: \code{"superPop"}.}

\item{\code{fieldPopInfAnc}}{a \code{character} string representing the name of
the column in the data frame that contains the inferred
super-population ancestry for the samples. The column should be
present in the data frame. Default: \code{"superPop"}.}

\item{\code{fieldSubPop}}{a \code{character} string representing the name of
the column in the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter) that contains the sub-population
information for the samples. The column should be present in the Population
Reference GDS file. Default: \code{"pop.group"}.}

\item{\code{verbose}}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}.}
}}

\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
