% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{RAIDSparam}
\alias{RAIDSparam}
\title{Create a RAIDSparam object}
\usage{
RAIDSparam(
  studyDF = NULL,
  studyDFSyn = NULL,
  pedStudy = NULL,
  studyType = c("LD", "GeneAware"),
  genoSource = NULL,
  blockTypeId = "GeneS.Ensembl.Hsapiens.v86",
  reference = "1KGc1.0",
  genome = "HG38",
  chrInfo = NULL,
  paramAncestry = NULL,
  profileFile = NULL,
  profileFileGeno = NULL,
  pathProfileGDS = NULL,
  fileReferenceGDS = NULL,
  fileReferenceAnnotGDS = NULL,
  inferenceType = "PCAknn",
  sampleRef = NULL,
  batch = 1L,
  prefix = "1",
  nbSim = 1L,
  offset = -1L,
  minCov = 10L,
  minProb = 0.999,
  seqError = 0.001,
  seqErrorSyn = 0.001,
  pRecomb = 0.01,
  np = 1L,
  listPos = NULL,
  syntheticRefDF = NULL,
  pruningMethod = c("corr", "r", "dprime", "composite"),
  slideWindowMaxBP = 500000L,
  thresholdLD = sqrt(0.1),
  specificSNV = NULL,
  genoType = "geno.ref",
  phaseType = "phase.ref",
  phase = FALSE,
  PCAmissingRate = 0.025,
  PCAalgorithm = "exact",
  eigenCount = 32L,
  eigenCountSyn = 15L,
  kList = seq(2L, 15L, 1L),
  pcaList = seq(2L, 15L, 1L),
  fieldPopInRef = "superPop",
  fieldPopInfAnc = "superPop",
  fieldSubPop = "pop.group",
  verbose = FALSE
)
}
\arguments{
\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). If \code{NULL}, the
following will be assigned:
\code{data.frame(study.id="NotDef", study.desc="NotDef", 
study.platform="NotDef", stringsAsFactors=FALSE)}.
Default: \code{NULL}.}

\item{studyDFSyn}{a \code{data.frame} containing the information about the
synthetic data to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "study.id", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor). If \code{NULL}, the
following will be assigned: \code{data.frame(study.id="NotDef.Synthetic", 
study.desc="NotDef synthetic data", study.platform="Synthetic", 
stringsAsFactors=FALSE)}.
Default: \code{NULL}.}

\item{pedStudy}{a \code{data.frame} containing the information TODO
with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", and "Source". All columns must be in
\code{character} strings (no factor). All row names should correspond to the
values in the "Name.ID" column. If \code{NULL}, the
following will be assigned:  \code{data.frame(Name.ID=c("ProfileId"), 
Case.ID=c("ProfileId"), Sample.Type=c("type"), Diagnosis="NotDef", 
Source=c("NotDef"), stringsAsFactors=FALSE, row.names = c("ProfileId"))}.
Default: \code{NULL}.}

\item{studyType}{a \code{character} string representing the type of study.
The possible choices are: "LD" and "GeneAware". The type of study affects
the way the estimation of the allelic fraction is done.
Default: \code{"LD"}.}

\item{genoSource}{TODO. The valid options are: "VCF", "generic",
"snp-pileup", and "bam". Default: \code{NULL}.}

\item{blockTypeId}{a single \code{character} string corresponding to
the block type used to extract the block identifiers. The block type must
be present in the GDS Reference Annotation file.
Default: \code{"GeneS.Ensembl.Hsapiens.v86"}.}

\item{reference}{a \code{character} string with two possible values:
'1KGv1.0', '1k_hgdpV0.1'. It specifies the type of inference.
Default: \code{"1KGc1.0"}.}

\item{genome}{a \code{character} string with one possible value:
'HG38'. It specifies the genome used. Default: \code{"HG38"}.}

\item{chrInfo}{TODO}

\item{paramAncestry}{a \code{list} of parameters related to ancestry.
The \code{list} should contain those three entries: \code{ScanBamParam},
\code{PileupParam}, and \code{yieldSize}. If \code{NULL}, the default value
\code{list(ScanBamParam=NULL, PileupParam=NULL, yieldSize=10000000)} will
be assigned to the parameter. Default: \code{NULL}.}

\item{profileFile}{a \code{character} string representing the path to the
file with genotype and the allele information of the profile. The format
of the accepted file is determined by the value in the 'genoSource' slot.
A profile would have a file with extension "\emph{vcf.gz" when 'genoSource' slot
is "VCF". If 'genoSource' is "generic" or "snp-pileup", then "}.txt.gz".
If 'genoSource' is "bam", then "\emph{.bam" (the file need to be indexed with a
existing corresponding "}.bai" file). Default: \code{NULL}.}

\item{profileFileGeno}{TODO}

\item{pathProfileGDS}{TODO}

\item{fileReferenceGDS}{TODO}

\item{fileReferenceAnnotGDS}{TODO}

\item{inferenceType}{TODO}

\item{sampleRef}{TODO}

\item{batch}{a single positive \code{integer} representing the current
identifier for the batch. Beware, this field is not stored anymore.
Default: \code{1L}.}

\item{prefix}{a single \code{character} string representing TODO.
Default: \code{"1"}.}

\item{nbSim}{a positive \code{integer} representing TODO.
Default: \code{1L}.}

\item{offset}{a single \code{integer} that is added to the SNP position to
switch from 0-based to 1-based coordinate when needed (or reverse).
Default: \code{-1L}.}

\item{minCov}{a single positive \code{integer} representing the minimum
required coverage. Default: \code{10L}.}

\item{minProb}{a single \code{numeric} between \code{0} and \code{1}
representing the probability that the calculated genotype call is correct.
Default: \code{0.999}.}

\item{seqError}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error.
Default: \code{0.001}.}

\item{seqErrorSyn}{a single \code{numeric} between \code{0} and \code{1}
representing the probability of sequencing error for synthetic.
Default: \code{0.001}.}

\item{pRecomb}{TODO}

\item{np}{TODO}

\item{listPos}{TODO}

\item{syntheticRefDF}{TODO}

\item{pruningMethod}{a \code{character} string representing the method that
will be used to calculate the linkage disequilibrium in the
\code{\link[SNPRelate]{snpgdsLDpruning}}() function. The 4 possible values
are: "corr", "r", "dprime", and "composite". Default: \code{"corr"}.}

\item{slideWindowMaxBP}{TODO}

\item{thresholdLD}{TODO}

\item{specificSNV}{TODO}

\item{genoType}{TODO}

\item{phaseType}{TODO}

\item{phase}{a \code{logical} indicating TODO. Default: \code{FALSE}.}

\item{PCAmissingRate}{TODO}

\item{PCAalgorithm}{TODO}

\item{eigenCount}{TODO}

\item{eigenCountSyn}{TODO}

\item{kList}{a \code{vector} of positive \code{integer} representing the
values tested for the  \emph{K} parameter. The \emph{K} parameter represents the
number of neighbors used in the K-nearest neighbor analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{pcaList}{a \code{vector} of positive \code{integer} representing
the values tested for the  \emph{D} parameter. The \emph{D} parameter represents the
number of dimensions used in the PCA analysis.
Default: \code{seq(2L, 15L, 1L)}.}

\item{fieldPopInRef}{a \code{character} string representing the name of the
column that contains the known ancestry for the reference profiles in
the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter). Default: \code{"superPop"}.}

\item{fieldPopInfAnc}{a \code{character} string representing the name of
the column in the data frame that contains the inferred
super-population ancestry for the samples. The column should be
present in the data frame. Default: \code{"superPop"}.}

\item{fieldSubPop}{a \code{character} string representing the name of
the column in the Population Reference GDS file (corresponding to the
\code{fileReferenceGDS} parameter) that contains the sub-population
information for the samples. The column should be present in the Population
Reference GDS file. Default: \code{"pop.group"}.}

\item{verbose}{a \code{logical} indicating if message information should be
printed. Default: \code{FALSE}}
}
\value{
an object of class \code{RAIDSparam} that contains all the
required parameters needed by the RAIDS workflow.
}
\description{
The function uses all parameters to create a
\code{RAIDSparam} class object that contains all the parameters used
in the RAIDS workflow.
}
\examples{


 ## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
