% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{runADMIXTURE}
\alias{runADMIXTURE}
\title{Call ADMIXTURE}
\usage{
runADMIXTURE(pathOut, fileBed, seedAdmix = 3435L, pRAIDS)
}
\arguments{
\item{pathOut}{a \code{character} representing path
where the file for ADMIXTURE (BED,BIM, FAM, and P.in)
are located and where the output of ADMIXTURE
will be generated.}

\item{fileBed}{a \code{character} represented
the BED file name (without path) for a}

\item{seedAdmix}{a \code{integer} representing the seed
pass to ADMIXTURE.
Default: \code{3435L}.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{0L} when successful.
}
\description{
TODO
}
\details{
TODO
}
\examples{


## Load the known ancestry for the demo 1KG reference profiles
data(demoKnownSuperPop1KG)

## The Reference GDS file
path1KG <- system.file("extdata/tests", package="RAIDS")

## Path to the demo Profile GDS file is located in this package
dataDir <- system.file("extdata/demoAncestryCall", package="RAIDS")

# The name of the synthetic study
studyID <- "MYDATA"

studyDF <- data.frame(study.id=studyID,
                             study.desc=studyID,
                             study.platform="NotDef",
                             stringsAsFactors=FALSE)
pathProfileGDS <- file.path(dataDir) # , "ex1.gds"
fileReferenceGDS <- system.file("extdata/tests/ex1_good_small_1KG.gds", package="RAIDS")
pedStudy <- data.frame(Name.ID=c("ex1"),
                             Case.ID=c("ex1"),
                             Sample.Type=c("type"),
                             Diagnosis="NotDef",
                             Source=c("NotDef"),
                             stringsAsFactors=FALSE)
     row.names(pedStudy) <- pedStudy$Name.ID

pRAIDS <- paramRAIDS(studyDF=studyDF,
                      pedStudy=pedStudy,
                      pathProfileGDS=pathProfileGDS,
                      fileReferenceGDS=fileReferenceGDS,
                      fieldPopInfAnc="SuperPop")

### TODO call
### 


}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
