% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{runExomeAncestry}
\alias{runExomeAncestry}
\title{TODO}
\usage{
runExomeAncestry(
  pedStudy,
  studyDF,
  pathProfileGDS,
  pathGeno,
  PATH_OUT,
  File.gds_1KG,
  File.gds_1KG.Annot,
  chrInfo,
  dataRefSyn
)
}
\arguments{
\item{pedStudy}{a \code{data.frame} with those mandatory columns: "Name.ID",
"Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in
\code{character} strings (no factor). The \code{data.frame}
must contain the information for all the samples passed in the
\code{listSamples} parameter. Only \code{fileNamePED} or \code{pedStudy}
can be defined.}

\item{studyDF}{a \code{data.frame} containing the information about the
study associated to the analysed sample(s). The \code{data.frame} must have
those 3 columns: "studyID", "study.desc", "study.platform". All columns
must be in \code{character} strings (no factor).}

\item{pathProfileGDS}{a \code{character} string representing the path to
the directory where the GDS Profile files will be created.
Default: \code{NULL}.}

\item{pathGeno}{a \code{character} string representing the path to the
directory containing the VCF output of SNP-pileup for each sample. The
SNP-pileup files must be compressed (gz files) and have the name identifiers
of the samples. A sample with "Name.ID" identifier would have an
associated SNP-pileup file called "Name.ID.txt.gz".}

\item{PATH_OUT}{a \code{character} string representing the path to
the directory where the out files are created.
Default: \code{NULL}.}

\item{File.gds_1KG}{a \code{character} string representing the file name of
the 1KG GDS file. The file must exist.}

\item{File.gds_1KG.Annot}{a \code{character} string representing the file name of
the 1KG GDS annotation file. The file must exist.}

\item{chrInfo}{a \code{vector} of \code{integer} values representing
the length of the chromosomes. See 'details' section.}

\item{dataRefSyn}{a \code{data.frame} containing those columns:
\itemize{
\item{sample.id} { a \code{character} string representing the sample
identifier. }
\item{pop.group} { a \code{character} string representing the
subcontinental population assigned to the sample. }
\item{superPop} { a \code{character} string representing the
super-population assigned to the sample. }
} can be the return of select1KGPop}
}
\value{
The integer \code{0L} when successful.

The *computeAncestryFromSyntheticFile()* function generates 3 types of files
in the *OUTPUT* directory.

* The ancestry inference CSV file (".Ancestry.csv" file)
* The inference information RDS file (".infoCall.rds" file)
* The parameter information RDS files from the synthetic inference ("KNN.synt.*.rds" files in a sub-directory)

In addition, a sub-directory (named using the *profile ID*) is
also created.
}
\description{
TODO
}
\examples{

## Path to the demo 1KG GDS file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

#################################################################
## The path and file name for the PED RDS file
## will the information about the analyzed samples
#################################################################

filePED <- file.path(data.dir, "example", "pedEx.rds")
ped <- readRDS(filePED)
head(ped)

#################################################################
## The 1KG GDS file and the 1KG SNV Annotation GDS file
## need to be located in the same directory
## Note that the 1KG GDS file used for this example is a
## simplified version and CANNOT be used for any real analysis
#################################################################

PATH_1KG <- file.path(data.dir, "example", "gdsRef")

fileName.GDS.1KG <- "ex1kg.gds"
file.GDS <- file.path(PATH_1KG, fileName.GDS.1KG)

fileName.GDS.Annot <- "exAnnot1kg.gds"
file.Annot <- file.path(PATH_1KG, fileName.GDS.Annot)


#################################################################
## The Sample SNP pileup files (one per sample) need
## to be located in the same directory.
#################################################################
pathGeno <- file.path(data.dir, "example", "snpPileup")

#################################################################
## The path where the Sample GDS files (one per sample)
## will be created need to be specified.
#################################################################
pathProfileGDS <- file.path(data.dir, "example", "out.tmp")

PATHOUT <- file.path(data.dir, "example", "res.out")

#################################################################
## A data frame containing general information about the study
## is also required. The data frame must have
## those 3 columns: "studyID", "study.desc", "study.platform"
#################################################################
studyDF <- data.frame(study.id = "MYDATA",
                      study.desc = "Description",
                      study.platform = "PLATFORM",
                      stringsAsFactors = FALSE)

gds_1KG <- snpgdsOpen(file.GDS)

####################################################################
## Fix RNG seed to ensure reproducible results
####################################################################

set.seed(3043)


dataRef <- select1KGPop(gds_1KG, nbSamples = 2L)

closefn.gds(gds_1KG)

## Chromosome length information
## chr23 is chrX, chr24 is chrY and chrM is 25
chrInfo <- c(248956422L, 242193529L, 198295559L, 190214555L,
    181538259L, 170805979L, 159345973L, 145138636L, 138394717L, 133797422L,
    135086622L, 133275309L, 114364328L, 107043718L, 101991189L, 90338345L,
    83257441L,  80373285L,  58617616L,  64444167L,  46709983L, 50818468L,
    156040895L, 57227415L,  16569L)

## A formal way to get the chormosome length information
## library(BSgenome.Hsapiens.UCSC.hg38)
## chrInfo <- integer(25L)
## for(i in seq_len(22L)){ chrInfo[i] <-
##                          length(Hsapiens[[paste0("chr", i)]])}
## chrInfo[23] <- length(Hsapiens[["chrX"]])
## chrInfo[24] <- length(Hsapiens[["chrY"]])
## chrInfo[25] <- length(Hsapiens[["chrM"]])
\dontrun{
runExomeAncestry(pedStudy = ped, studyDF = studyDF,
                    pathProfileGDS = pathProfileGDS,
                    pathGeno = pathGeno,
                    PATH_OUT = PATHOUT,
                    File.gds_1KG = file.GDS,
                    File.gds_1KG.Annot = file.Annot,
                    chrInfo = chrInfo,
                    dataRefSyn = dataRef)

unlink(pathProfileGDS, recursive = TRUE, force=TRUE)
unlink(PATHOUT, recursive = TRUE, force=TRUE)
}
}
\references{
Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ,
Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution
of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72.
doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.
}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
