% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{runIBDKING}
\alias{runIBDKING}
\title{Function just wrap snpgdsIBDKING}
\usage{
runIBDKING(gds, sampleId = NULL, snp.id = NULL, maf = 0.05)
}
\arguments{
\item{gds}{a \code{gds} object.}

\item{sampleId}{a \code{array} with the sample to keep
if NULL all}

\item{snp.id}{a \code{array} with the snp.id to keep
if NULL all}

\item{maf}{a \code{numeric} mininum allelic frequency keep.}
}
\value{
TODO ibd.robust
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
