% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\encoding{UTF-8}
\name{runLDPruning}
\alias{runLDPruning}
\title{TODO just a wrapper of snpgdsLDpruning}
\usage{
runLDPruning(
  gds,
  method = "corr",
  listSamples = NULL,
  listKeep = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE
)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{method}{the parameter method in SNPRelate::snpgdsLDpruning}

\item{listSamples}{A \code{vector} of \code{string} corresponding to
the sample.ids
use in LDpruning}

\item{listKeep}{the list of snp.id keep}

\item{slide.max.bp.v}{TODO}

\item{ld.threshold.v}{TODO}

\item{np}{a single positive \code{integer} specifying the number of
threads to be used. Default: \code{1}.}

\item{verbose.v}{a \code{logical} indicating if information is shown during
the process. Default: \code{FALSE}.}
}
\value{
a \code{list} of SNP IDs stratified by chromosomes as generated
by \code{\link[SNPRelate:snpgdsLDpruning]{SNPRelate:snpgdsLDpruning()}}.
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO
gds <- "Demo GDS TODO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
