% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsWrapper.R
\name{runLDPruning}
\alias{runLDPruning}
\title{TODO just a wrapper of snpgdsLDpruning}
\usage{
runLDPruning(
  gds,
  method = "corr",
  listSamples = NULL,
  listKeep = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE
)
}
\arguments{
\item{gds}{an object of class \code{gds} opened}

\item{method}{the parameter method in SNPRelate::snpgdsLDpruning}

\item{listSamples}{A \code{vector} of \code{string} corresponding to
the sample.ids
use in LDpruning}

\item{listKeep}{the list of snp.id keep}

\item{slide.max.bp.v}{}

\item{ld.threshold.v}{}

\item{np}{}

\item{verbose.v}{}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
