% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{sampleRef<-,RAIDSparam-method}
\alias{sampleRef<-,RAIDSparam-method}
\title{Setter function for replacement of sampleRef slot in
a RAIDSparam class}
\usage{
\S4method{sampleRef}{RAIDSparam}(x) <- value
}
\arguments{
\item{x}{a \code{RAIDSparam} object.}

\item{value}{\code{NULL} or a \code{vector} of \code{character} strings
that represent a subset of sample identifiers from the GDS Reference File.
Those samples will be used for the ancestry inference. If \code{NULL}, all
the samples are retained.}
}
\value{
the modified \code{RAIDSparam} object when the new value is valid.
}
\description{
A function for replacement of sampleRef slot in a
\code{RAIDSparam} class.
}
\examples{

## Create a RAIDSparam object
paramDemo <- RAIDSparam()

## Assign a new values to the sampleRef slot in the object
sampleRef(paramDemo) <- c("Sample1", "Sample2")

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
