% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{select1KGPop}
\alias{select1KGPop}
\title{TODO}
\usage{
select1KGPop(gds, nbSamples)
}
\arguments{
\item{gds}{an object of class \code{gds} opened with 1k genome in it}

\item{nbSample}{a \code{numeric} between 0 and 1 TODO}
}
\value{
TODO a \code{vector} of \code{string}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alex Krasnitz
}
