% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{select1KGPop}
\alias{select1KGPop}
\title{TODO}
\usage{
select1KGPop(gds, nbSamples)
}
\arguments{
\item{gds}{an object of class \link[gdsfmt]{gdsn.class} (a GDS node), or
\link[gdsfmt]{gds.class} (a GDS file) containing the information about
1000 Genome (1kG).}

\item{nbSamples}{a single positive \code{integer}.}
}
\value{

}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
