% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processStudy.R
\encoding{UTF-8}
\name{selParaPCAUpQuartile}
\alias{selParaPCAUpQuartile}
\title{TODO}
\usage{
selParaPCAUpQuartile(
  matKNN.All,
  pedCall,
  refCall,
  predCall,
  listCall,
  kList = seq(3, 15, 1),
  pcaList = seq(2, 15, 1)
)
}
\arguments{
\item{matKNN.All}{TODO see it is rbind matKNN of the
computeKNNSuperPoprSynthetic return from group synthetic data}

\item{pedCall}{TODO see return of prepPedSynthetic1KG}

\item{refCall}{TODO column name in pedCall with the call}

\item{predCall}{TODO column name in matKNN with the call}

\item{listCall}{TODO array of the possible call}

\item{kList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _K_ parameter. The _K_ parameter represents the
number of neighbors used in the K-nearest neighbor analysis. If \code{NULL},
the value \code{seq(3,15,1)} is assigned.
Default: \code{seq(3,15,1)}.}

\item{pcaList}{a \code{vector} of \code{integer} representing  the list of
values tested for the  _D_ parameter. The _D_ parameter represents the
number of dimensions used in the PCA analysis.  If \code{NULL},
the value \code{seq(2,15,1)} is assigned.
Default: \code{seq(2,15,1)}.}
}
\value{
a \code{list} containing 5 entries:
\itemize{
\item{dfPCA} {TODO}
\item{dfPop} {TODO}
\item{D} {TODO}
\item{K} {TODO}
\item{listD} {TODO}
}
}
\description{
TODO
}
\examples{

# TODO
listEigenvector <- "TOTO"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
