% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{snvListVCF}
\alias{snvListVCF}
\title{Generate a vcf for the SNP with at least freqCutoff for all frequency
for at least one super population}
\usage{
snvListVCF(gds, fileOUT, offset = 0, freqCutoff = NULL)
}
\arguments{
\item{gds}{a \code{character} string representing the path and file
name of the GDS file that contains the 1KG information. The GDS file must
contain the SNP information, the genotyping information and
the pedigree information from 1000 Genomes.
The extension of the file must be '.gds'.}

\item{fileOUT}{TODO}

\item{offset}{TODO. Default: \code{0}.}

\item{freqCutoff}{TODO. Default: \code{NULL}.}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="aicsPaper")

## TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
