% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{snvListVCF}
\alias{snvListVCF}
\title{Generate a vcf for the SNP with at least freqCutoff for all frequency
for at least one super population}
\usage{
snvListVCF(gds, fileOUT, offset = 0, freqCutoff = NULL)
}
\arguments{
\item{fileOUT}{TODO}

\item{PATHPRUNED}{TODO}

\item{filePref}{TODO}
}
\value{
TODO a \code{vector} of \code{numeric}
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Desch&ecirc;nes and Alexander Krasnitz
}
\keyword{internal}
