% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{splitSelectByPop}
\alias{splitSelectByPop}
\title{TODO}
\usage{
splitSelectByPop(dataRef)
}
\arguments{
\item{dataRef}{a \code{data.frame} containing those columns:
\itemize{
\item{sample.id} { TODO }
\item{pop.group} { TODO }
\item{superPop} { TODO }
}}
}
\value{
a \code{data.frame} containing TODO
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alex Krasnitz
}
