% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\name{studyDF,RAIDSparam-method}
\alias{studyDF,RAIDSparam-method}
\title{Getter method for studyDF slot in the "RAIDSparam" class}
\usage{
\S4method{studyDF}{RAIDSparam}(x)
}
\description{
Getter method for studyDF slot in the "RAIDSparam" class
}
