% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\name{studyDF<-,RAIDSparam-method}
\alias{studyDF<-,RAIDSparam-method}
\title{Implement the method for the "StudRAIDSparament" class}
\usage{
\S4method{studyDF}{RAIDSparam}(x) <- value
}
\description{
Implement the method for the "StudRAIDSparament" class
}
