% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\name{studyDFSyn<-,RAIDSparam-method}
\alias{studyDFSyn<-,RAIDSparam-method}
\title{Implement the method for the "RAIDSparam" class}
\usage{
\S4method{studyDFSyn}{RAIDSparam}(x) <- value
}
\description{
Implement the method for the "RAIDSparam" class
}
