% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyDFSyn<-}
\alias{studyDFSyn<-}
\title{Generic function for replacement of studyDFSyn slot in a class}
\usage{
studyDFSyn(x) <- value
}
\arguments{
\item{x}{an S4 object.}

\item{value}{the new value to assign or update.}
}
\value{
the modified S4 object when the new value is valid.
}
\description{
A generic function for replacement of studyDFSyn slot in a
S4 object.
}
\examples{
 
# Define a dummy class to show usage
setClass("MyClass", slots = list(studyDFSyn="character"))

# Create an instance
obj <- new("MyClass", studyDFSyn="123")

# Call the generic (assuming a method is implemented)
# studyDFSyn(obj) <- "333"

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
