% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIDSparam_class.R
\encoding{UTF-8}
\name{studyType}
\alias{studyType}
\title{Generic function for getting the studyType slot in a class}
\usage{
studyType(x)
}
\arguments{
\item{x}{a S4 object.}
}
\value{
a value from the studyType slot in the S4 object.
}
\description{
A generic function for getting the studyType slot in a
S4 object.
}
\examples{

# Define a dummy class to show usage
setClass("MyClass", slots = list(studyType="character"))

# Create an instance
obj <- new("MyClass", studyType="123")

# Call the generic (assuming a method is implemented)
# studyType(obj)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
