% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\encoding{UTF-8}
\name{syntheticGeno}
\alias{syntheticGeno}
\title{Generate synthetic profiles for each cancer sample and 1KG
reference sample combination and add them to the GDS Sample file}
\usage{
syntheticGeno(
  gds,
  gdsRefAnnot,
  gdsSampleFile,
  data.id.profile,
  listSampleRef,
  nbSim = 1L,
  prefId = "",
  pRecomb = 0.01,
  minProb = 0.999,
  seqError = 0.001
)
}
\arguments{
\item{gds}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the 1KG GDS file.}

\item{gdsRefAnnot}{an object of class \code{\link[gdsfmt]{gds.class}}
(a GDS file), the1 1KG SNV Annotation GDS file.}

\item{gdsSampleFile}{a \code{character} string representing the file name of
the GDS Sample file containing the information about the sample.
The file must exist.}

\item{data.id.profile}{a \code{character} string representing the unique
identifier of the cancer sample.}

\item{listSampleRef}{a \code{vector} of \code{character} string
representing the sample identifiers of the 1KG selected reference samples.}

\item{nbSim}{a single positive \code{integer} representing the number of
simulations that will be generated per sample + 1KG reference combination.
Default: \code{1L}.}

\item{prefId}{a \code{character} string that represent the prefix that will
be added to the name of the synthetic profiles generated by the function.
Default: \code{""}.}

\item{pRecomb}{a single positive \code{numeric} between 0 and 1 that
represents the frequency of phase switching in the synthetic profiles,
Default: \code{0.01}.}

\item{minProb}{a single positive \code{numeric} between 0 and 1 that
represents the probability that the genotype is correct. TODO.
Default: \code{0.999}.}

\item{seqError}{a single positive \code{numeric} between 0 and 1
representing the sequencing error rate. Default: \code{0.001}.}
}
\value{
\code{OL} when the function is successful.
}
\description{
The functions uses one cancer sample in combination with one
1KG reference sample to generate an synthetic profile that is saved in the
GDS Sample file. When more than one 1KG reference samples are specified,
the function recursively generates synthetic profiles for
each cancer sample + 1KG reference sample combination. The number of
synthetic profiles generated by combination if specified by the number of
simulation requested.
}
\examples{

# TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
