% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{tableBlockAF}
\alias{tableBlockAF}
\title{TOREVIEW Compile the information about the variants
for each bloc.}
\usage{
tableBlockAF(snpPos)
}
\arguments{
\item{snpPos}{For a specific chromosome a \code{data.frame} with lap for
the SNV dataset with
coverage > \code{minCov}.}
}
\value{
TOREVIEW resBlock a \code{data.frame} containing only heterozygote
SNV information. The
\code{data.frame} contain those columns:
\itemize{
\item{block} {a single \code{integer} representing the unique identifier
of the block.}
\item{aRF} {a single \code{numeric} representing the final allelic fraction not compute
here -1 for all entries.}
\item{aFraction} {a single \code{integer} representing the possible allelic
fraction in absence of LOH.}
\item{lR} {a single \code{integer} representing the coverage for
the alternative allele.}
\item{nPhase} {a single \code{integer} representin the number of SNV
phases.}
\item{sumAlleleLow} {a single \code{integer} TOREVIEW sum of the allele with
the less coverage.}
\item{sumAlleleHigh} {a single \code{integer} TOREVIEW sum of the allele
with more coverage.}
\item{lH} {a single \code{numeric} for the homozygotes log10 of the product
frequencies of the allele not found in the profile (not a probability).}
\item{lM} {a single \code{numeric} log10 product frequency allele
in population.}
\item{lRhomo} {a single \code{numeric} score lH - lM.}
\item{nbHomo} {a single \code{integer} representing the number of
homozygote SNVs per block.}
\item{nbKeep} {a single \code{integer} representing the number of
SNVs retained per block.}
\item{nbHetero} {a single \code{integer} representing the number of
heterozygote SNVs per block.}
}
}
\description{
TOREVIEW For each block, the function evaluates a score
about lost of heterizygocity and allelic fration. It generates some
information about the variants in the block like number of homozigte or
heterozygote.
In the case of RNA-seq the blocks are genes.
}
\examples{

dataDir <- system.file("extdata", package="RAIDS")

snpPos <- readRDS(file.path(dataDir, "demoAllelicFraction", "demSnpPos.rds"))

result <- RAIDS:::tableBlockAF(snpPos[which(snpPos$snp.chr == 1),])
head(result)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
