% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{testAlleleFractionChange}
\alias{testAlleleFractionChange}
\title{TODO}
\usage{
testAlleleFractionChange(matCov, pCutOff = -3, vMean)
}
\arguments{
\item{matCov}{TODO}

\item{pCutOff}{TODO Default: \code{-3}.}

\item{vMean}{TODO}
}
\value{
a \code{list} containing 3 entries:
\itemize{
\item{pWin} {TODO.}
\item{pCut} {TODO.}
\item{pCut1} {TODO.}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
