% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction_internal.R
\encoding{UTF-8}
\name{testAlleleFractionChange}
\alias{testAlleleFractionChange}
\title{TODO}
\usage{
testAlleleFractionChange(matCov, pCutOff = -3, vMean)
}
\arguments{
\item{matCov}{a \code{data.frame} containing only heterozygote SNVs. The
\code{data.frame} must contain those columns:
\itemize{
\item{cnt.ref} {a single \code{integer} representing the coverage for
the reference allele.}
\item{cnt.alt} {a single \code{integer} representing the coverage for
the alternative allele.}
}}

\item{pCutOff}{a \code{numeric} representing the cut-off for considering
a region imbalanced when comparing likelihood to gave allelic fraction
change and likelihood not to have allelic fraction change.
Default: \code{-3}.}

\item{vMean}{a \code{numeric} representing the}
}
\value{
a \code{list} containing 4 entries:
\itemize{
\item{pWin}{TODO}
\item{p}{TODO}
\item{pCut}{a \code{integer} indicating if all SNVs tested
positive (1=TRUE, 0-FALSE).
TODO }
\item{pCut1}{Calculate a global statistic using all SNVs TODO}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
\keyword{internal}
