% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicFraction.R
\encoding{UTF-8}
\name{testEmptyBox}
\alias{testEmptyBox}
\title{TODO}
\usage{
testEmptyBox(matCov, pCutOff = -3)
}
\arguments{
\item{matCov}{TODO}

\item{pCutOff}{TODO}
}
\value{
a \code{list} TODO containing:
\itemize{
\item{pWin}{TODO}
\item{p}{TODO}
\item{pCut}{TODO}
\item{pCut1}{TODO}
}
}
\description{
TODO
}
\examples{

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
