% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersRAIDS_internal.R
\encoding{UTF-8}
\name{validateParamRAIDS}
\alias{validateParamRAIDS}
\title{Validate that the parametersRAIDS object is valid}
\usage{
validateParamRAIDS(parameters)
}
\arguments{
\item{parameters}{a \code{parametersRAIDS} an object with all the
parameters used by the RAIDS workflow.}
}
\value{
\code{TRUE} when all the parameters in the object are valid;
otherwise \code{FALSE}
}
\description{
This function validates all the parameters present in the
\code{parametersRAIDS} object.
}
\examples{
##TODO


}
\author{
Pascal Belleau and Astrid Deschênes
}
\keyword{internal}
