% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersRAIDS_internal.R
\encoding{UTF-8}
\name{validateParamRAIDS_subpart02}
\alias{validateParamRAIDS_subpart02}
\title{Validate the second subsection fo the input parameters}
\usage{
validateParamRAIDS_subpart02(parameters)
}
\arguments{
\item{parameters}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
\code{TRUE} when all the parameters tested in the object are valid;
otherwise \code{FALSE}
}
\description{
This function validates a subsection of the parameters
present in the \code{parametersRAIDS} object.
}
\examples{
##TODO


}
\author{
Pascal Belleau and Astrid Deschênes
}
\keyword{internal}
