% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{writeBedBimFilesProfileFilter}
\alias{writeBedBimFilesProfileFilter}
\title{Generate a plink bed file}
\usage{
writeBedBimFilesProfileFilter(pathOut, fileP, listProfile, indexS = 1, pRAIDS)
}
\arguments{
\item{pathOut}{TODO}

\item{fileP}{a \code{character} string representing the path and
the file name of the new bim file.}

\item{listProfile}{a \code{vector} of \code{character} representing the list
of profiles from gdsProfile to keep.}

\item{indexS}{a \code{integer} TODO. Default: \code{1}.}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters.}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
