% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\encoding{UTF-8}
\name{writeBedRef}
\alias{writeBedRef}
\title{Generate a plink bed file}
\usage{
writeBedRef(fileOut, listRM = NULL, subset = FALSE, pRAIDS)
}
\arguments{
\item{fileOut}{a \code{character} string representing the path and
the file name of the new bim file.}

\item{listRM}{a \code{vector} of \code{character} strings containing the
identifiers for the reference samples that need to be removed for the
PCA analysis.}

\item{subset}{a \code{logical} extract the genotypes from the reference in
listRM (TRUE) or extract from the reference not in listRM (FALSE).}

\item{pRAIDS}{a \code{parametersRAIDS} an object with all the RAIDS
parameters}
}
\value{
The integer \code{0L} when successful.
}
\description{
This function extract the SNPs that pass a frequency cut-off
in at least one super population
from a GDS SNP information file and save the retained SNP information into
a VCF file.
}
\examples{

## Required library
library(gdsfmt)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## Demo 1KG Reference GDS file
fileGDS <- openfn.gds(file.path(dataDir,
                    "PopulationReferenceDemo.gds"))

## Output VCF file that will be created (temporary)
vcfFile <- file.path(tempdir(), "Demo_TMP_01.vcf")

## Create a VCF file with the SNV dataset present in the GDS file
## No cutoff on frequency, so all SNVs are saved
snvListVCF(gdsReference=fileGDS, fileOut=vcfFile, offset=0L,
                    freqCutoff=NULL)

## Close GDS file (IMPORTANT)
closefn.gds(fileGDS)

## Remove temporary VCF file
unlink(vcfFile, force=TRUE)

}
\author{
Pascal Belleau, Astrid Deschênes and Alexander Krasnitz
}
